### 2020-12-14 #2908 @whedon
[PRE REVIEW]: samplics: a Python Package for selecting, weighting and analyzing data from complex sampling designs.
**Submitting author:** @MamadouSDiallo (<a href="http://orcid.org/0000-0002-0376-3631">Mamadou Diallo</a>)
**Repository:** <a href="https://github.com/survey-methods/samplics/" target ="_blank">https://github.com/survey-methods/samplics/</a>
**Version:** 0.2.6
