### 2020-12-14 #2908 @MamadouSDiallo
Dear Openjournals/Joss-Reviews

As JOSS editor, I suggest lorenanicole.

For potential reviewers, i suggest katschueler (katja.schueler@kgu.de),
ejhigson (ejhigson@gmail.com), nhejazi (nhejazi@berkeley.edu), and ohmaddi (
oh73@cornell.edu)

Best regards
