### 2020-12-14 #2908 @kthyng
Hi @MamadouSDiallo. Thank you for your submission. I have a question about it: the software readme ends with "This is a beta version. At this stage, this project is not recommended to be used for production or any project that the user depend on." Can you explain this note further? In particular, I am looking for information that will help me and the editorial board understand if this fulfills our [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) and I'm not sure that the beta version of a software package will do so.
