### 2020-12-14 #2908 @MamadouSDiallo
Dear Kristen

I have not updated the readme page for months. Since the time I wrote that,
I have added many tests and the software is certainly ready to be used. I
still plan to keep the numbering scheme in a beta format 0.X for a while to
get the chance to incorporate more users' feedback in the API design.

The software is not a one-off program. It is meant to be a tool for
analyzing survey sampling and as such will continue incorporating validated
methods for analyzing complex survey samples. I hope that the evolving
nature of the package is not a reason for rejection by JOSS. However,
please let me know if that disqualifies the software.

Looking forward to your feedback

Best regards

On Mon, Dec 14, 2020 at 4:37 PM Kristen Thyng <notifications@github.com>
wrote:

> Hi @MamadouSDiallo <https://github.com/MamadouSDiallo>. Thank you for
> your submission. I have a question about it: the software readme ends with
> "This is a beta version. At this stage, this project is not recommended to
> be used for production or any project that the user depend on." Can you
> explain this note further? In particular, I am looking for information that
> will help me and the editorial board understand if this fulfills our submission
> guidelines
> <https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements>
> and I'm not sure that the beta version of a software package will do so.
>
