### 2020-12-15 #2908 @kthyng
Hi @MamadouSDiallo, thanks for the info. One guideline for if a package is ready for submission to JOSS is if it is relatively complete in terms of functionality.  It’s normal to add features, but the base functionality should be solid. I would think the API would be relatively stable as well. You can check out the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). The readme is part of the documentation and thus should be up to date before going to review.

Given this information, do you need more time to update your software before we decide about going to review?
