### 2020-12-15 #2908 @MamadouSDiallo
Dear Kirsten

Thank you for the clarification. Yes, the software is relatively complete
and I do not expect the API to change in a meaningful way. The readme is
behind compared to the actual documentation of the tool (
https://samplics.readthedocs.io/en/latest/).

I will improve the readme and possibly tweak a few things based on JOSS
review criteria. How much time can you give me to implement these
improvements before you start the review?

Best regards

On Tue, Dec 15, 2020 at 9:54 AM Kristen Thyng <notifications@github.com>
wrote:

> Hi @MamadouSDiallo <https://github.com/MamadouSDiallo>, thanks for the
> info. One guideline for if a package is ready for submission to JOSS is if
> it is relatively complete in terms of functionality. It’s normal to add
> features, but the base functionality should be solid. I would think the API
> would be relatively stable as well. You can check out the review criteria
> <https://joss.readthedocs.io/en/latest/review_criteria.html>. The readme
> is part of the documentation and thus should be up to date before going to
> review.
>
> Given this information, do you need more time to update your software
> before we decide about going to review?
>
