### 2020-12-15 #2908 @MamadouSDiallo
I will work on these tasks and come back to you within two weeks. Bets
regards.

On Tue, Dec 15, 2020 at 11:03 AM Kristen Thyng <notifications@github.com>
wrote:

> @MamadouSDiallo <https://github.com/MamadouSDiallo> If you can complete
> these tasks in a week or two, then let's just leave this open for now. If
> not, you can withdraw the paper and resubmit.
>
> Note in the comments above from our bot whedon, too, you are missing a
> statement of need in your paper, and the paper itself is not compiling
> correctly. Please address these items too.
>
