+++ README.rst
@@ -99,10 +99,11 @@ License
 -------
 `MIT <https://github.com/survey-methods/samplics/blob/master/license.txt>`_
 
-Project status
+Contact 
 --------------
-This is a beta version. At this stage, this project is not recommended to be
-used for production or any project that the user depend on.
+| Mamadou S. Diallo 
+| msdiallo@QuantifyAfrica.org
+| https://www.linkedin.com/in/mamadousdiallo/ 
 
 
 
