+++ README.rst
@@ -101,9 +101,7 @@ License
 
 Contact 
 --------------
-| Mamadou S. Diallo 
-| msdiallo@QuantifyAfrica.org
-| https://www.linkedin.com/in/mamadousdiallo/ 
+created by `Mamadou S. Diallo <https://twitter.com/MamadouSDiallo>` - feel free to contact me!
 
 
 
