+++ README.rst
@@ -15,7 +15,7 @@
 In large scale surveys, often complex random mechanisms are used to select
 samples. Estimates derived from such samples must reflect the random
 mechanism. *Samplics* is a python package that implements a set of
-sampling techniques for complex survey designs. These techniques are organized into the following four subpackages.
+sampling techniques for complex survey designs. These survey sampling techniques are organized into the following four subpackages.
 
 **Sampling** provides a set of random selection techniques used to draw a sample from a population. It also provides procedures for calculating sample sizes. The sampling subpackage contains: 
 
