+++ README.rst
@@ -45,7 +45,12 @@ Usage
 Let's assume that we have a population and we would like to select a sample from it. The goal is to calculate the sample size for an expected proportion of 0.80 with a precision of 0.10.
 
 .. code:: python
-
+    :width: 200px
+    :align: center
+    :height: 100px
+    :alt: alternate text
+    :figclass: align-center
+    
     import samplics
     from samplics.sampling import SampleSize
 
