+++ README.md
@@ -165,11 +165,11 @@ To predict small area parameters, we can use code similar to:
 
 _pip install samplics_
 
-Python 3.6 or newer is required and the main dependencies are [numpy](https://numpy.org), [pandas](https://pandas.pydata.org), [scpy](https://www.scipy.org), and [statsmodel] (https://www.statsmodels.org/stable/index.h.tml).
+Python 3.6 or newer is required and the main dependencies are [numpy](https://numpy.org), [pandas](https://pandas.pydata.org), [scpy](https://www.scipy.org), and [statsmodel](https://www.statsmodels.org/stable/index.h.tml).
 
 ## License
 
-[MIT] (https://github.com/survey-methods/samplics/blob/master/license.txt)
+[MIT](https://github.com/survey-methods/samplics/blob/master/license.txt)
 
 ## Contact
 
