+++ README.md
@@ -2,10 +2,8 @@
 
 <img src="https://travis-ci.com/survey-methods/samplics.svg?token=WwRayqkQBt1W4ihyTzvw&branch=master">
 <!-- :target: https://travis-ci.com/survey-methods/samplics -->
-
 <img src="https://codecov.io/gh/survey-methods/samplics/branch/master/graph/badge.svg?token=7C0LBB5N8Y">
 <!-- :target: https://codecov.io/gh/survey-methods/samplics -->
-
 <img src="https://readthedocs.org/projects/samplics/badge/?version=latest">
 <!-- :target: https://samplics.readthedocs.io/en/latest/?badge=latest
 :alt: Documentation Status -->
