+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "Select, weight and analyze complex sample data"
 authors = ["Mamadou S Diallo <msdiallo@quantifyafrica.org>"]
 license = "MIT"
 
-readme = "README.rst"
+readme = "README.md"
 
 homepage = "https://samplics.org/"
 repository = "https://github.com/survey-methods/samplics"
