+++ noxfile.py
@@ -1,7 +1,7 @@
 import nox
 
 
-@nox.session(python=["3.6", "3.7"])
+@nox.session(python=["3.6", "3.7", "3.8"])
 def tests(session):
     session.run("poetry", "install")
     session.run("pytest", "-v", "tests")
