+++ tests/utils/test_basic_functions.py
@@ -78,10 +78,19 @@ def test_transform_exp_inverse(y):
     y_transformed = transform(y + constant, llambda, constant, inverse=True)
     assert np.isclose(y_transformed, np.exp(np.log(1 + (y + constant) * llambda) / llambda)).all()
 
-
+@pytest.mark.xfail
 @pytest.mark.parametrize("y", [y1, y2])
 def test_plot_skewness(y):
     plot_skewness(y, block=False)
+
+@pytest.mark.xfail
+@pytest.mark.parametrize("y", [y1, y2])
+def test_plot_kurtosis(y):
+    plot_kurtosis(y, block=False)
+
+@pytest.mark.parametrize("y", [y1, y2])
+def test_plot_skewness(y):
+    plot_skewness(y, coef_min=-2, coef_max=3, block=False)
 
 
 @pytest.mark.parametrize("y", [y1, y2])
