+++ pyproject.toml
@@ -30,12 +30,12 @@ packages = [
 #exclude = ["tests", "docs", "datasets"] - exclude by default since not inside src
 
 [tool.poetry.dependencies]
-python = "^3.6.1"
-numpy = "^1.15.1"
-pandas = "^1.0.1"
-scipy = "^1.4.1"
-statsmodels = "^0.10.1"
-matplotlib = "^3.1.3"
+python = "^3.6.1" #Pandas 1.0 requires at least Python 3.6.1
+numpy = "^1.15"
+pandas = "^1.0"
+scipy = "^1.1"
+statsmodels = "^0.10"
+matplotlib = "^3.1"
 
 [tool.poetry.dev-dependencies]
 mypy = "^0.770"
@@ -48,7 +48,7 @@ sphinx = "^3.0"
 sphinx-bootstrap-theme = "^0.7.1"
 nbsphinx = "^0.5.0"
 jupyterlab = "^1.2.4"
-nb_black = "^1.0.7"
+# nb_black = "^1.0.7"
 certifi = "^2019.11.28"
 pytest-cov = "^2.8.1"
 nox = "^2019.11.9"
@@ -58,7 +58,7 @@ codecov = "^2.1.3"
 
 [tool.black]
 line-length = 99
-target-version = ['py36', 'py37', 'py38']
+target-version = ['py36', 'py37', 'py38', 'py39']
 exclude = '''
 /(
     \.eggs
@@ -74,7 +74,19 @@ exclude = '''
 )/
 '''
 
+[tool.isort]
+profile = "black"
+force_single_line = true
+atomic = true
+include_trailing_comma = true
+lines_after_imports = 2
+lines_between_types = 1
+use_parentheses = true
+src_paths = ["src", "tests"]
+filter_files = true
+known_first_party = "samplics"
+
 [build-system]
-requires = ["poetry>=0.12"]
+requires = ["poetry>=1.0.0"]
 build-backend = "poetry.masonry.api"
 
