+++ README copy.rst
@@ -1,180 +1,0 @@
-==========
-*SAMPLICS*
-==========
-.. image:: https://travis-ci.com/survey-methods/samplics.svg?token=WwRayqkQBt1W4ihyTzvw&branch=master
-  :target: https://travis-ci.com/survey-methods/samplics
-
-.. image:: https://codecov.io/gh/survey-methods/samplics/branch/master/graph/badge.svg?token=7C0LBB5N8Y
-  :target: https://codecov.io/gh/survey-methods/samplics     
-
-.. image:: https://readthedocs.org/projects/samplics/badge/?version=latest
-  :target: https://samplics.readthedocs.io/en/latest/?badge=latest
-  :alt: Documentation Status
-
-|
-In large scale surveys, often complex random mechanisms are used to select
-samples. Estimates derived from such samples must reflect the random
-mechanism. *Samplics* is a python package that implements a set of
-sampling techniques for complex survey designs. These survey sampling techniques are organized into the following four subpackages.
-
-**Sampling** provides a set of random selection techniques used to draw a sample from a population. It also provides procedures for calculating sample sizes. The sampling subpackage contains: 
-
-* Sample size calculation and allocation: Wald and Fleiss methods for proportions. 
-* Equal probability of selection: simple random sampling (SRS) and systematic selection (SYS)
-* Probability proportional to size (PPS): Systematic, Brewer's method, Hanurav-Vijayan method, Murphy's method, and Rao-Sampford's method.
-
-**Weighting** provides the procedures for adjusting sample weights. More specifically, the weighting subpackage allows the following:
-
-* Weight adjustment due to nonresponse
-* Weight poststratification, calibration and normalization
-* Weight replication i.e. Bootstrap, BRR, and Jackknife
-
-**Estimation** provides methods for estimating the parameters of interest with uncertainty measures that are consistent with the sampling design. The estimation subpackage implements the following types of estimation methods:
-
-* Taylor-based, also called linearization methods
-* Replication-based estimation i.e. Boostrap, BRR, and Jackknife
-* Regression-based e.g. generalized regression (GREG)
-
-**Small Area Estimation (SAE).** When the sample size is not large enough to produce reliable / stable domain level estimates, SAE techniques can be used to model the output variable of interest to produce domain level estimates. This subpackage provides Area-level and Unit-level SAE methods. 
-
-For more details, visit https://samplics.readthedocs.io/en/latest/
-
-Usage
-------
-
-Let's assume that we have a population and we would like to select a sample from it. The goal is to calculate the sample size for an expected proportion of 0.80 with a precision of 0.10.
-
-.. code-block:: python
-
-    import samplics
-    from samplics.sampling import SampleSize
-
-    sample_size = SampleSize(parameter = "proportion")
-    sample_size.calculate(target=0.80, precision=0.10)
-
-    assert size_nat_wald.samp_size["__none__"] == 62
-
-Furthermore, the population is located in four natural regions i.e. North, South, East, and West. We could be interested in calculating sample sizes based on region specific requirements e.g. expected proportions, desired precisions and associated design effects.  
-
-.. code:: python
-
-    import samplics
-    from samplics.sampling import SampleSize
-
-    sample_size = SampleSize(parameter="proportion", method="wald", stratification=True)
-
[…]
