+++ joss/paper.md
@@ -22,6 +22,14 @@ Survey sampling techniques are used in various fields to obtain information abou
 
 Until the initiation of `samplics`, Python did not have a library for analyzing complex survey samples similar to the [R survey package](https://CRAN.R-project.org/package=survey) [@lumley04] and several commercial software such as SAS, SPSS, Stata, and more. `samplics` is a Python package developed to provide a comprehensive set of python code to select random samples, adjust sample weights, produce design-based survey estimates, and predict small area parameters.
 
+# Statement of Need
+
+`samplics` aims at providing a comprehensive statistical package for analyzing survey sample data. The primary target audiences are survey statisticians and other data analysts working with sample data obtained from complex survey designs. This package can be used for producing, analyzing, and using official statistics. It also can be useful for teaching statistical concepts given the wide use of Python in Education and the simplicity of the `samplics` APIs.
+
+When designing a survey, `samplics` can calculate sample sizes by stratum base on expected proportions and level of precision for the indicator of interest as well as measures of the complexity of the design such as survey design effects. After sample sizes are determined, `samplics` can calculate probabilities of selection according to the sample selection strategy. To ensure representativeness of the random sample, `samplics` will compute design weights and adjust them for non-response, post-stratification, and calibration. `samplics` provides Taylor-based and replication-based techniques for computing population parameter estimates and associated measures of uncertainty. Finally, `samplics` has a small area estimation subpackage that can predict small area parameters. Note that `samplics` can be used independently for any of the steps described in the paragraph.
+
+In the sections below, more details are provided on the survey sampling techniques implemented in `samplics`.
+
 # Survey Sampling Techniques
 
 In large scale surveys, often complex random mechanisms are used to select samples. Estimations obtained from such samples must reflect the random mechanism to ensure correct approximations of the population parameters by sample estimates [@cochran77], [@kish65], and [@lohr10]. For Python users, `samplics` implements an extensive set of these sampling techniques for complex survey designs.
@@ -38,32 +46,32 @@ Sample weighting is the main mechanism used in surveys to formalize the represen
 
[…]
