### 2020-12-26 #2908 @MamadouSDiallo
Dear Kristen

I have updated the GitHub repo and I have added a statement of need to the
paper. I have tested that the paper compiles using the Whedon paper preview
tool.

Looking forward to your feedback.

Best regards

On Wed, Dec 16, 2020 at 1:09 PM Kristen Thyng <notifications@github.com>
wrote:

> Ok thank you. I've added the paused label to help indicate this.
>
