### 2020-12-28 da8c46b
started categorical module and added tests for expansion module
