### 2020-12-28 fed9157
started categorical module and added tests for expansion module
