+++ pyproject.toml
@@ -41,7 +41,7 @@ matplotlib = "^3.1"
 mypy = "^0.770"
 pytest = "^5.0.1"
 flake8 = "^3.7.9"
-black = "^19.10b0"
+black = "^20.8b1"
 ipython = "^7.11.0"
 ipykernel = "^5.1.3"
 sphinx = "^3.0"
@@ -58,7 +58,7 @@ codecov = "^2.1.3"
 
 [tool.black]
 line-length = 99
-target-version = ['py36', 'py37', 'py38', 'py39']
+target-version = ['py36', 'py37', 'py38']
 exclude = '''
 /(
     \.eggs
