### 2020-12-28 4e92abe
started cross-tabulation module
