### 2020-12-28 7282b2c
started cross-tabulation module
