### 2020-12-28 cb19077
added fpc to estimation module
