+++ src/samplics/estimation/expansion.py
@@ -13,12 +13,13 @@ and Wolter, K.M. (2007) [#w2007]_.
 """
 
 import math
-from typing import TypeVar, Any, Dict, List, Optional, Tuple
+from typing import TypeVar, Any, Dict, List, Optional, Union, Tuple
 
 import numpy as np
+from numpy.lib.arraysetops import isin
 import pandas as pd
-from samplics.utils import formats
-from samplics.utils.types import Array, StringNumber
+from samplics.utils.formats import fpc_as_dict, numpy_array
+from samplics.utils.types import Array, Number, StringNumber
 from scipy.stats import t as student
 
 TypeTaylorEst = TypeVar("TypeTaylorEst", bound="TaylorEstimator")
@@ -48,6 +49,7 @@ class _SurveyEstimator:
         self.deff: Dict[StringNumber, Any] = {}
         self.lower_ci: Dict[StringNumber, Any] = {}
         self.upper_ci: Dict[StringNumber, Any] = {}
+        self.fpc: Dict[StringNumber, Any] = {}
         self.strata: List[StringNumber] = ["__none__"]
         self.domains: List[StringNumber] = ["__none__"]
         self.method: str = "taylor"
@@ -94,22 +96,22 @@ class _SurveyEstimator:
         Optional[np.ndarray],
         Optional[np.ndarray],
     ]:
-        y = formats.numpy_array(y)
-        samp_weight = formats.numpy_array(samp_weight)
+        y = numpy_array(y)
+        samp_weight = numpy_array(samp_weight)
         if x is not None:
-            x = formats.numpy_array(x)
+            x = numpy_array(x)
             x = x[~excluded_units]
         if stratum is not None:
-            stratum = formats.numpy_array(stratum)
+            stratum = numpy_array(stratum)
             stratum = stratum[~excluded_units]
         if domain is not None:
-            domain = formats.numpy_array(domain)
+            domain = numpy_array(domain)
             domain = domain[~excluded_units]
         if psu is not None:
-            psu = formats.numpy_array(psu)
+            psu = numpy_array(psu)
             psu = psu[~excluded_units]
         if ssu is not None:
-            ssu = formats.numpy_array(ssu)
+            ssu = numpy_array(ssu)
             ssu = ssu[~excluded_units]
 
         return (
@@ -129,8 +131,8 @@ class _SurveyEstimator:
         psu: np.ndarray = None,
     ) -> None:
 
-        stratum = formats.numpy_array(stratum)
-        psu = formats.numpy_array(psu)
+        stratum = numpy_array(stratum)
+        psu = numpy_array(psu)
 
         if stratum.size <= 1:
             self.number_psus = np.unique(psu).size if psu.size > 1 else samp_weight.size
@@ -139,7 +141,7 @@ class _SurveyEstimator:
             self.number_psus = np.unique([stratum, psu], axis=1).shape[1]
             self.number_strata = np.unique(stratum).size
         else:
-            samp_weight = formats.numpy_array(samp_weight)
+            samp_weight = numpy_array(samp_weight)
             self.number_psus = samp_weight.size
             self.number_strata = np.unique(stratum).size
 
@@ -343,14 +345,17 @@ class TaylorEstimator(_SurveyEstimator):
         stratum: np.ndarray,
         psu: np.ndarray,
         ssu: Optional[np.ndarray] = None,
+        fpc: Dict[StringNumber, Number] = {"__none__": 1},
     ) -> np.float64:
         """Computes the variance across stratum """
 
         if stratum is None:
             number_psus = np.unique(psu).size
-            var_est = self._variance_stratum_between(
+            var_est = fpc["__none__"] * self._variance_stratum_between(
                 y_score, samp_weight, number_psus, psu
-            ) + self._variance_stratum_within(y_score, number_psus, psu, ssu)
[…]
