+++ src/samplics/utils/formats.py
@@ -138,3 +138,15 @@ def remove_nans(excluded_units: Array, *args) -> Tuple:
         vars.append(var[~excluded_units])
 
     return vars
+
+
+def fpc_as_dict(stratum: Array, fpc: Union[Array, Number]):
+
+    if stratum is None and isinstance(fpc, (int, float)):
+        return {"__none__": fpc}
+    elif stratum is not None and isinstance(fpc, (int, float)):
+        return dict(zip(stratum, np.repeat(fpc, stratum.shape[0])))
+    elif stratum is not None and isinstance(fpc, np.ndarray):
+        return dict(zip(stratum, fpc))
+    else:
+        raise TypeError("stratum and fpc are not compatible!")
