+++ tests/estimation/test_expansion_fpc.py
@@ -1,0 +1,536 @@
+import numpy as np
+import pandas as pd
+
+from samplics.estimation import TaylorEstimator
+
+one_stage_fpc = pd.read_csv("./tests/estimation/one_stage_fpc.csv")
+
+fpc_array_psu = 1 - one_stage_fpc["nh"] / one_stage_fpc["Nh"]
+fpc_dict_psu = dict(zip(one_stage_fpc["stratid"], fpc_array_psu))
+fpc_dict_no_psu = {1: 1 - 5 / 15, 2: 1 - 6 / 9, 3: 1 - 10 / 20}
+
+stratum = one_stage_fpc["stratid"]
+psu = one_stage_fpc["psuid"]
+weight = one_stage_fpc["weight"]
+domain = one_stage_fpc["domain"]
+x = one_stage_fpc["x"]
+y = one_stage_fpc["y"]
+fpc_psu = 1 - 5 / 44
+fpc_no_psu = 1 - 21 / 44
+
+"""Taylor Approximation WITHOUT Stratification for TOTAL"""
+svy_total_without_str = TaylorEstimator("total")
+
+
+def test_total_estimator_without_str():
+    svy_total_without_str.estimate(y, weight, psu=psu, fpc=fpc_psu)
+
+    assert np.isclose(svy_total_without_str.point_est["__none__"], 853.12)
+    assert np.isclose(svy_total_without_str.stderror["__none__"], 153.2627)
+    assert np.isclose(svy_total_without_str.lower_ci["__none__"], 427.5946)
+    assert np.isclose(svy_total_without_str.upper_ci["__none__"], 1278.645)
+
+
+def test_total_estimator_without_str_nor_psu():
+    svy_total_without_str.estimate(y, weight, fpc=fpc_no_psu)
+
+    assert np.isclose(svy_total_without_str.point_est["__none__"], 853.12)
+    assert np.isclose(svy_total_without_str.stderror["__none__"], 67.14785)
+    assert np.isclose(svy_total_without_str.lower_ci["__none__"], 713.052)
+    assert np.isclose(svy_total_without_str.upper_ci["__none__"], 993.188)
+
+
+svy_total_without_str_domain = TaylorEstimator("total")
+
+
+def test_total_estimator_without_str_domain():
+    svy_total_without_str_domain.estimate(y, weight, psu=psu, fpc=fpc_psu, domain=domain)
+
+    assert np.isclose(svy_total_without_str_domain.point_est[1], 528.52)
+    assert np.isclose(svy_total_without_str_domain.point_est[2], 324.6)
+    assert np.isclose(svy_total_without_str_domain.stderror[1], 195.5778)
+    assert np.isclose(svy_total_without_str_domain.stderror[2], 116.0081)
+    assert np.isclose(svy_total_without_str_domain.lower_ci[1], -14.49112)
+    assert np.isclose(svy_total_without_str_domain.lower_ci[2], 2.510021)
+    assert np.isclose(svy_total_without_str_domain.upper_ci[1], 1071.531)
+    assert np.isclose(svy_total_without_str_domain.upper_ci[2], 646.69)
+
+
+def test_total_estimator_without_str_nor_psu_domain():
+    svy_total_without_str_domain.estimate(y, weight, domain=domain, fpc=fpc_no_psu)
+
+    assert np.isclose(svy_total_without_str_domain.point_est[1], 528.52)
+    assert np.isclose(svy_total_without_str_domain.point_est[2], 324.6)
+    assert np.isclose(svy_total_without_str_domain.stderror[1], 92.9434)
+    assert np.isclose(svy_total_without_str_domain.stderror[2], 69.55674)
[…]
