+++ tests/estimation/test_expansion_fpc.py
@@ -15,6 +15,7 @@ weight = one_stage_fpc["weight"]
 domain = one_stage_fpc["domain"]
 x = one_stage_fpc["x"]
 y = one_stage_fpc["y"]
+z = one_stage_fpc["z"]
 fpc_psu = 1 - 5 / 44
 fpc_no_psu = 1 - 21 / 44
 
@@ -336,201 +337,155 @@ def test_ratio_estimator_with_str_nor_psu_domain():
     assert np.isclose(svy_ratio_with_str_domain.upper_ci[2], 1.898337)
 
 
-# """Taylor Approximation WITHOUT Stratification for PROPORTION"""
-# svy_prop_without_str = TaylorEstimator("proportion")
+"""Taylor Approximation WITHOUT Stratification for PROPORTION"""
+svy_prop_without_str = TaylorEstimator("proportion")
 
 
-# def test_prop_estimator_without_str():
-#     svy_prop_without_str.estimate(y, weight, psu=psu, remove_nan=True)
+def test_prop_estimator_without_str():
+    svy_prop_without_str.estimate(z, weight, psu=psu, fpc=fpc_psu)
 
-#     assert np.isclose(svy_prop_without_str.point_est["__none__"][0.0], 0.186_377_5)
-#     assert np.isclose(svy_prop_without_str.point_est["__none__"][1.0], 0.813_622_5)
-#     assert np.isclose(svy_prop_without_str.variance["__none__"][0.0], 0.020_285_6 ** 2)
-#     assert np.isclose(svy_prop_without_str.variance["__none__"][1.0], 0.020_285_6 ** 2)
-#     assert np.isclose(svy_prop_without_str.stderror["__none__"][0.0], 0.020_285_6)
-#     assert np.isclose(svy_prop_without_str.stderror["__none__"][1.0], 0.020_285_6)
-#     assert np.isclose(svy_prop_without_str.lower_ci["__none__"][0.0], 0.149_023)
-#     assert np.isclose(svy_prop_without_str.lower_ci["__none__"][1.0], 0.769_441_4)
-#     assert np.isclose(svy_prop_without_str.upper_ci["__none__"][0.0], 0.230_558_6)
-#     assert np.isclose(svy_prop_without_str.upper_ci["__none__"][1.0], 0.850_977)
-#     assert np.isclose(svy_prop_without_str.coef_var["__none__"][0.0], 0.020_285_6 / 0.186_377_5)
-#     assert np.isclose(svy_prop_without_str.coef_var["__none__"][1.0], 0.020_285_6 / 0.813_622_5)
+    assert np.isclose(svy_prop_without_str.point_est["__none__"][0.0], 0.3253012)
+    assert np.isclose(svy_prop_without_str.point_est["__none__"][1.0], 0.6746988)
+    assert np.isclose(svy_prop_without_str.stderror["__none__"][0.0], 0.1499394)
+    assert np.isclose(svy_prop_without_str.stderror["__none__"][1.0], 0.1499394)
+    assert np.isclose(svy_prop_without_str.lower_ci["__none__"][0.0], 0.0674673)
+    assert np.isclose(svy_prop_without_str.lower_ci["__none__"][1.0], 0.2373558)
+    assert np.isclose(svy_prop_without_str.upper_ci["__none__"][0.0], 0.7626442)
+    assert np.isclose(svy_prop_without_str.upper_ci["__none__"][1.0], 0.9325327)
 
 
-# def test_prop_estimator_without_str_nor_psu():
-#     svy_prop_without_str.estimate(y, weight, remove_nan=True)
+def test_prop_estimator_without_str_nor_psu():
[…]
