### 2020-12-29 4a65b70
addind cross-tabulation
