+++ src/samplics/utils/types.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 
 
-Array = Union[np.ndarray, pd.Series]
+Array = Union[np.ndarray, pd.Series, list]
 Number = Union[float, int]
 StringNumber = Union[str, float, int]
 DictStrNum = Dict[StringNumber, Number]
