+++ src/samplics/categorical/tabulation.py
@@ -14,7 +14,7 @@ import pandas as pd
 from scipy.stats import t as student
 from scipy.stats.stats import f_oneway
 
-from samplics.utils.formats import numpy_array, remove_nans
+from samplics.utils.formats import numpy_array, remove_nans, sample_size_dict
 from samplics.utils.types import Array, Number, StringNumber
 
 from samplics.estimation import TaylorEstimator
@@ -28,7 +28,7 @@ class OneWay:
         ciprop_method: str = "logit",
     ) -> None:
 
-        if parameter.lower() not in ("count", "prop", "proportion"):
+        if parameter.lower() in ("count", "prop", "proportion"):
             self.parameter = parameter.lower()
         else:
             raise ValueError("parameter must be 'count' or 'proportion'")
@@ -39,13 +39,47 @@ class OneWay:
         self.alpha = alpha
         self.ciprop_method = ciprop_method
 
-    # def _tabulate()
+    def _estimate(
+        self,
+        y: Array,
+        samp_weight: Optional[Union[Array, Number]] = None,
+        stratum: Optional[Array] = None,
+        psu: Optional[Array] = None,
+        ssu: Optional[Array] = None,
+        domain: Optional[Array] = None,
+        fpc: Union[Dict, float] = 1,
+        deff: bool = False,
+        coef_variation: bool = False,
+        remove_nan: bool = False,
+    ) -> TaylorEstimator:
+
+        if self.parameter == "count":
+            tbl_est = TaylorEstimator(parameter="total", alpha=self.alpha)
+        elif self.parameter == "proportion":
+            tbl_est = TaylorEstimator(parameter=self.parameter, alpha=self.alpha)
+        else:
+            raise ValueError("parameter must be 'count' or 'proportion'")
+
+        tbl_est.estimate(
+            y,
+            samp_weight=samp_weight,
+            stratum=stratum,
+            psu=psu,
+            ssu=ssu,
+            domain=domain,
+            fpc=fpc,
+            deff=deff,
+            coef_variation=coef_variation,
+            remove_nan=remove_nan,
+        )
+
+        return tbl_est
 
     def tabulate(
         self,
         vars: Array,
         varnames: Optional[List[str]] = None,
-        samp_weight: Optional[Union[Array]] = None,
+        samp_weight: Optional[Union[Array, Number]] = None,
         stratum: Optional[Array] = None,
         psu: Optional[Array] = None,
         ssu: Optional[Array] = None,
@@ -59,23 +93,46 @@ class OneWay:
         docstring
         """
 
+        if vars is None:
+            raise AssertionError("vars need to be an array-like object")
+
         vars_np = numpy_array(vars)
-        nb_vars = 1 if vars_np.shape[1] is None else vars_np.shape[1]
+        nb_vars = 1 if len(vars_np.shape) == 1 else vars_np.shape[1]
 
-        if self.parameter == "proportion":
-            pass
-        elif self.parameter == "count":
-            for _ in range(0, nb_vars):
-                count_est = TaylorEstimator(
-                    parameter=self.parameter, alpha=self.alpha, ciprop_method=self.ciprop_method
-                )
-                count_est.estimate(
-                    vars_np[0], samp_weight, stratum, psu, ssu, domain, fpc, deff, coef_variation
-                )
-                breakpoint()
-
+        if samp_weight is None:
+            samp_weight = np.ones(vars_np.shape[0])
+        elif isinstance(samp_weight, (int, float)):
+            samp_weight = np.repeat(samp_weight, vars_np.shape[0])
         else:
-            raise ValueError("parameter must be 'count' or 'proportion'")
+            samp_weight = numpy_array(samp_weight)
+
+        if nb_vars == 1:
+            tbl_est = self._estimate(
+                y=vars_np,
[…]
