+++ tests/tabulation/test_oneway.py
@@ -1,5 +1,0 @@
-import numpy as np
-import pandas as pd
-
-from samplics.estimation import TaylorEstimator
-from samplics.tabulation import OneWay
