+++ src/samplics/utils/basic_functions.py
@@ -10,27 +10,42 @@ Functions:
 respectively. 
 
 """
-from typing import Optional
+from typing import List, Optional
 
 import numpy as np
+import pandas as pd
 import matplotlib.pyplot as plt
+from numpy.lib.arraysetops import isin
 
 
 from samplics.utils import formats
 from samplics.utils.types import Array, Number
 
 
+def set_variables_names(vars: Array, varnames: List[str], prefix: str) -> List[str]:
+
+    if varnames is None:
+        if isinstance(vars, pd.DataFrame):
+            return list(vars.columns)
+        elif isinstance(vars, pd.Series) and vars.name is not None:
+            return vars.name
+        else:
+            return [prefix + "_" + k for k in range(len(vars.shape))]
+    else:
+        return varnames
+
+
 def sumby(
     group: np.ndarray, y: np.ndarray
 ) -> np.ndarray:  # Could use pd.grouby().sum(), may scale better
-    """Sums the numpy array by group. 
+    """Sums the numpy array by group.
 
     Args:
-        group (np.ndarray): provides the group associated to each observation. 
-        y (np.ndarray): variable of interest for the summation. 
+        group (np.ndarray): provides the group associated to each observation.
+        y (np.ndarray): variable of interest for the summation.
 
     Returns:
-        np.ndarray: the sums by group. 
+        np.ndarray: the sums by group.
     """
 
     groups = np.unique(group)
@@ -47,11 +62,11 @@ def averageby(
     """[summary]
 
     Args:
-        group (np.ndarray): provides the group associated to each observation. 
-        y (np.ndarray): variable of interest for the averaging. 
+        group (np.ndarray): provides the group associated to each observation.
+        y (np.ndarray): variable of interest for the averaging.
 
     Returns:
-        np.ndarray: the averages by group. 
+        np.ndarray: the averages by group.
     """
 
     groups = np.unique(group)
@@ -63,17 +78,20 @@ def averageby(
 
 
 def transform(
-    y: np.ndarray, llambda: Optional[Number] = None, constant: float = 0.0, inverse: bool = True,
+    y: np.ndarray,
+    llambda: Optional[Number] = None,
+    constant: float = 0.0,
+    inverse: bool = True,
 ) -> np.ndarray:
-    """Transforms the variable of interest using the Boxcox method or its inverse. 
+    """Transforms the variable of interest using the Boxcox method or its inverse.
 
     Args:
-        y (np.ndarray): variable of interest. 
+        y (np.ndarray): variable of interest.
         llambda (Optional[Number], optional): Boxcox parameter lambda. A value of zero indicates
         a logarithm transformation. Defaults to None.
-        constant (float, optional): An additive term to ensure that the logarithm function is 
+        constant (float, optional): An additive term to ensure that the logarithm function is
         applied to positive values. Defaults to 0.0.
-        inverse (bool, optional): indicates if the Boxcox transformation is applied or its 
+        inverse (bool, optional): indicates if the Boxcox transformation is applied or its
         inverse. Defaults to True.
 
     Returns:
@@ -87,11 +105,11 @@ def transform(
         else:
             if ((y + constant) <= 0).any():
                 raise ValueError("log function not defined for negative numbers")
-            else: 
+            else:
                 return np.log(y + constant)
     elif llambda != 0.0:
         if inverse:
-            if ((1 + y * llambda) <= 0).any(): 
+            if ((1 + y * llambda) <= 0).any():
                 raise ValueError("log function not defined for negative numbers")
[…]
