+++ src/samplics/utils/basic_functions.py
@@ -30,7 +30,15 @@ def set_variables_names(vars: Array, varnames: List[str], prefix: str) -> List[s
         elif isinstance(vars, pd.Series) and vars.name is not None:
             return [vars.name]
         else:
-            return [prefix + "_" + str(k) for k in range(1, len(vars.shape) + 1)]
+            if isinstance(vars, (pd.Series, np.ndarray)):
+                if len(vars.shape) == 2:
+                    return [prefix + "_" + str(k) for k in range(1, vars.shape[1] + 1)]
+                else:
+                    return [prefix + "_" + str(k) for k in range(1, len(vars.shape) + 1)]
+            elif isinstance(vars, (tuple, list)):
+                return [var.name for var in vars]
+            else:
+                raise TypeError("vars should be an array-like")
     else:
         return varnames
 
