+++ tests/categorical/test_oneway.py
@@ -13,6 +13,17 @@ birth_cat = birthcat["birthcat"]
 pop = birthcat["pop"]
 
 
+# tbl_prop = OneWay("proportion")
+# tbl_prop.tabulate(birth_cat, remove_nan=True)
+
+
+# def test_oneway_prop_one_var_count():
+#     assert tbl_prop.table["birthcat"][1] == 0.26
+#     assert tbl_prop.table["birthcat"][2] == 0.4875
+#     assert tbl_prop.table["birthcat"][3] == 0.2524
+# exit()
+
+
 tbl_count = OneWay("count")
 # print(tbl_count)
 
@@ -38,21 +49,21 @@ def test_oneway_count_one_var_count():
 
 
 def test_oneway_count_one_var_stderr():
-    assert np.isclose(tbl_count.stderror["birthcat"][1], 13.3337)
-    assert np.isclose(tbl_count.stderror["birthcat"][2], 15.1940)
-    assert np.isclose(tbl_count.stderror["birthcat"][3], 13.2050)
+    assert np.isclose(tbl_count.stderror["birthcat"][1], 13.3337, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["birthcat"][2], 15.1940, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["birthcat"][3], 13.2050, atol=1e-4)
 
 
 def test_oneway_count_one_var_lower_ci():
-    assert np.isclose(tbl_count.lower_ci["birthcat"][1], 213.8321)
-    assert np.isclose(tbl_count.lower_ci["birthcat"][2], 420.1812)
-    assert np.isclose(tbl_count.lower_ci["birthcat"][3], 207.0847)
+    assert np.isclose(tbl_count.lower_ci["birthcat"][1], 213.8321, atol=1e-4)
+    assert np.isclose(tbl_count.lower_ci["birthcat"][2], 420.1812, atol=1e-4)
+    assert np.isclose(tbl_count.lower_ci["birthcat"][3], 207.0847, atol=1e-4)
 
 
 def test_oneway_count_one_var_upper_ci():
-    assert np.isclose(tbl_count.upper_ci["birthcat"][1], 266.1679)
-    assert np.isclose(tbl_count.upper_ci["birthcat"][2], 479.8188)
-    assert np.isclose(tbl_count.upper_ci["birthcat"][3], 258.9153)
+    assert np.isclose(tbl_count.upper_ci["birthcat"][1], 266.1679, atol=1e-4)
+    assert np.isclose(tbl_count.upper_ci["birthcat"][2], 479.8188, atol=1e-4)
+    assert np.isclose(tbl_count.upper_ci["birthcat"][3], 258.9153, atol=1e-4)
 
 
 def test_oneway_count_one_var_deff_false():
@@ -66,6 +77,38 @@ def test_oneway_count_one_var_deff_false():
 #     assert tbl_count.deff["birthcat"][1] == 1
 #     assert tbl_count.deff["birthcat"][2] == 1
 #     assert tbl_count.deff["birthcat"][3] == 1
+
+tbl_prop = OneWay("proportion")
+tbl_prop.tabulate(birth_cat, remove_nan=True)
+print(tbl_prop.__dict__)
+
+
+def test_oneway_prop_one_var_prop():
+    assert np.isclose(tbl_prop.table["birthcat"][1], 0.2600, atol=1e-4)
+    assert np.isclose(tbl_prop.table["birthcat"][2], 0.4875, atol=1e-4)
+    assert np.isclose(tbl_prop.table["birthcat"][3], 0.2524, atol=1e-4)
+
+
[…]
