+++ src/samplics/categorical/tabulation.py
@@ -45,43 +45,53 @@ class OneWay:
 
     def _estimate(
         self,
-        y: Array,
+        var_of_ones: Array,
+        var: Array,
         samp_weight: Array = None,
         stratum: Optional[Array] = None,
         psu: Optional[Array] = None,
         ssu: Optional[Array] = None,
-        domain: Optional[Array] = None,
         fpc: Union[Dict, float] = 1,
         deff: bool = False,
         coef_variation: bool = False,
         remove_nan: bool = False,
     ) -> TaylorEstimator:
 
-        if self.parameter == "count":
-            tbl_est = TaylorEstimator(parameter="total", alpha=self.alpha)
-        elif self.parameter == "proportion":
-            tbl_est = TaylorEstimator(parameter=self.parameter, alpha=self.alpha)
-        else:
-            raise ValueError("parameter must be 'count' or 'proportion'")
-
         if remove_nan:
-            excluded_units = np.isnan(domain) | np.isnan(samp_weight)
-            y, samp_weight, stratum, domain, psu, ssu = remove_nans(
-                excluded_units, y, samp_weight, stratum, domain, psu, ssu
+            excluded_units = np.isnan(var) | np.isnan(samp_weight)
+            var_of_ones, samp_weight, stratum, var, psu, ssu = remove_nans(
+                excluded_units, var_of_ones, samp_weight, stratum, var, psu, ssu
             )
 
-        tbl_est.estimate(
-            y,
-            samp_weight=samp_weight,
-            stratum=stratum,
-            psu=psu,
-            ssu=ssu,
-            domain=domain,
-            fpc=fpc,
-            deff=deff,
-            coef_variation=coef_variation,
-            remove_nan=remove_nan,
-        )
+        if self.parameter == "count":
+            tbl_est = TaylorEstimator(parameter="total", alpha=self.alpha)
+            tbl_est.estimate(
+                y=var_of_ones,
+                samp_weight=samp_weight,
+                stratum=stratum,
+                psu=psu,
+                ssu=ssu,
+                domain=var,
+                fpc=fpc,
+                deff=deff,
+                coef_variation=coef_variation,
+                remove_nan=remove_nan,
+            )
+        elif self.parameter == "proportion":
+            tbl_est = TaylorEstimator(parameter=self.parameter, alpha=self.alpha)
+            tbl_est.estimate(
+                y=var,
+                samp_weight=samp_weight,
+                stratum=stratum,
+                psu=psu,
+                ssu=ssu,
+                fpc=fpc,
+                deff=deff,
+                coef_variation=coef_variation,
+                remove_nan=remove_nan,
+            )
+        else:
+            raise ValueError("parameter must be 'count' or 'proportion'")
 
         return tbl_est
 
@@ -93,7 +103,7 @@ class OneWay:
         stratum: Optional[Array] = None,
         psu: Optional[Array] = None,
         ssu: Optional[Array] = None,
-        # by: Optional[Array] = None,
+        # Todo: by: Optional[Array] = None,
         fpc: Union[Dict, float] = 1,
         deff: bool = False,
         coef_variation: bool = False,
@@ -135,42 +145,58 @@ class OneWay:
 
         if nb_vars == 1:
             tbl_est = self._estimate(
-                y=np.ones(vars_np.shape[0]),
+                var_of_ones=np.ones(vars_np.shape[0]),
+                var=vars_np,
                 samp_weight=samp_weight,
                 stratum=stratum,
                 psu=psu,
                 ssu=ssu,
-                domain=vars_np,
                 fpc=fpc,
                 deff=deff,
                 coef_variation=coef_variation,
[…]
