+++ tests/categorical/test_oneway.py
@@ -13,19 +13,21 @@ birth_cat = birthcat["birthcat"]
 pop = birthcat["pop"]
 
 
-# tbl_prop = OneWay("proportion")
-# tbl_prop.tabulate(birth_cat, remove_nan=True)
+# tbl2_count = OneWay("count")
 
+# tbl2_count.tabulate(birthcat[["region", "birthcat"]], remove_nan=True)
 
-# def test_oneway_prop_one_var_count():
-#     assert tbl_prop.table["birthcat"][1] == 0.26
-#     assert tbl_prop.table["birthcat"][2] == 0.4875
-#     assert tbl_prop.table["birthcat"][3] == 0.2524
+# breakpoint()
 # exit()
 
 
+########
+
 tbl_count = OneWay("count")
 # print(tbl_count)
+
+# print(birthcat[["region", "birthcat"]])
+# exit()
 
 
 @pytest.mark.xfail(strict=True, reason="Parameter not valid")
@@ -80,7 +82,6 @@ def test_oneway_count_one_var_deff_false():
 
 tbl_prop = OneWay("proportion")
 tbl_prop.tabulate(birth_cat, remove_nan=True)
-print(tbl_prop.__dict__)
 
 
 def test_oneway_prop_one_var_prop():
@@ -111,5 +112,126 @@ def test_oneway_prop_one_var_deff_false():
     assert tbl_prop.deff["birthcat"] == {}
 
 
-# def test_two():
-#     tbl_count.tabulate([birth_cat, region], remove_nan=True)
+tbl2_count = OneWay("count")
+
+
+def test_oneway_count_two_vars_list():
+    tbl2_count.tabulate([region, birth_cat], remove_nan=True)
+
+    assert tbl2_count.table["birthcat"][1] == 240
+    assert tbl2_count.table["birthcat"][2] == 450
+    assert tbl2_count.table["birthcat"][3] == 233
+
+    assert tbl2_count.table["region"][1] == 166
+    assert tbl2_count.table["region"][2] == 284
+    assert tbl2_count.table["region"][3] == 250
+    assert tbl2_count.table["region"][4] == 256
+
+
+def test_oneway_count_two_vars_tuple():
+    tbl2_count.tabulate((region, birth_cat), remove_nan=True)
+
+    assert tbl2_count.table["birthcat"][1] == 240
+    assert tbl2_count.table["birthcat"][2] == 450
+    assert tbl2_count.table["birthcat"][3] == 233
+
+    assert tbl2_count.table["region"][1] == 166
+    assert tbl2_count.table["region"][2] == 284
+    assert tbl2_count.table["region"][3] == 250
+    assert tbl2_count.table["region"][4] == 256
+
+
+def test_oneway_count_two_vars_pandas():
+    tbl2_count.tabulate(birthcat[["region", "birthcat"]], remove_nan=True)
+
+    assert tbl2_count.table["birthcat"][1] == 240
+    assert tbl2_count.table["birthcat"][2] == 450
+    assert tbl2_count.table["birthcat"][3] == 233
+
+    assert tbl2_count.table["region"][1] == 166
+    assert tbl2_count.table["region"][2] == 284
+    assert tbl2_count.table["region"][3] == 250
+    assert tbl2_count.table["region"][4] == 256
+
+
+tbl22_count = OneWay("count")
+tbl22_numpy = birthcat[["region", "birthcat"]].to_numpy()
+tbl22_count.tabulate(tbl22_numpy, remove_nan=True)
+
+
+def test_oneway_count_two_vars_numpy():
[…]
