### 2021-01-03 e602446
adding twoway class
