+++ tests/categorical/test_twoway.py
@@ -44,6 +44,7 @@ tbl_prop.tabulate([region, birth_cat], varnames=["region", "birth_cat"], remove_
 tbl_count = CrossTabulation("count")
 tbl_count.tabulate([age_cat, birth_cat], varnames=["age_cat", "birth_cat"], remove_nan=True)
 
+breakpoint()
 
 # def test_oneway_count_one_var_count():
 #     assert tbl_count.table["birthcat"][1] == 240
