+++ src/samplics/categorical/tabulation.py
@@ -18,6 +18,7 @@ from scipy.stats import chi2, f
 from samplics.utils.basic_functions import set_variables_names
 from samplics.utils.formats import (
     concatenate_series_to_str,
+    dict_to_dataframe,
     numpy_array,
     numpy_to_dummies,
     remove_nans,
@@ -244,9 +245,39 @@ class CrossTabulation:
         self.deff = {}  # Dict[str, Dict[str, Number]]
         self.alpha = alpha
         self.ciprop_method = ciprop_method
+        self.design_info = {}
+        self.row_levels = []
+        self.col_levels = []
+        self.vars_names = []
 
     def __repr__(self):
         return f"CrossTabulation(parameter={self.parameter}, alpha={self.alpha})"
+
+    def __str__(self):
+
+        if self.vars_names == []:
+            return "Not variables to tabulate"
+        else:
+            table_header = f"Cross-tabulation of {self.vars_names[0]} and {self.vars_names[1]}"
+            table_subheader1 = f"Number of strata: {self.design_info['number_strata']}"
+            table_subheader2 = f"Number of PSUs: {self.design_info['number_psus']}"
+            table_subheader3 = f"Number of observations: {self.design_info['number_obs']}"
+            table_subheader4 = f"Design effect: {self.design_info['design_effect']}"
+
+            chisq_dist = f"chi2({self.stats['Pearson-Unadj']['df']})"
+            f_dist = f"F({self.stats['Pearson-Adj']['df_num']:.2f}, {self.stats['Pearson-Adj']['df_den']:.2f}"
+
+            pearson_unadj = f"Unadjusted - {chisq_dist}: {self.stats['Pearson-Unadj']['chisq_value']:.4f} with p-value of {self.stats['Pearson-Unadj']['p_value']:.4f}"
+            pearson_adj = f"Adjusted - {f_dist}): {self.stats['Pearson-Adj']['f_value']:.4f}  with p-value of {self.stats['Pearson-Adj']['p_value']:.4f}"
+            pearson_test = (
+                f"Pearson (with Rao-Scott adjustment):\n\t{pearson_unadj}\n\t{pearson_adj}"
+            )
+
+            lr_unadj = f"Unadjusted - {chisq_dist}: {self.stats['LR-Unadj']['chisq_value']:.4f} with p-value of {self.stats['LR-Unadj']['p_value']:.4f}"
+            lr_adj = f"Adjusted - {f_dist}): {self.stats['LR-Adj']['f_value']:.4f}  with p-value of {self.stats['LR-Adj']['p_value']:.4f}"
+            lr_test = f"Likelihood ratio (with Rao-Scott adjustment):\n\t{lr_unadj}\n\t{lr_adj}"
+
+            return f"\n{table_header}\n {table_subheader1}\n {table_subheader2}\n {table_subheader3}\n {table_subheader4}\n\n {self.to_dataframe()}\n\n {pearson_test} \n\n {lr_test}"
 
     def tabulate(
         self,
@@ -478,26 +509,63 @@ class CrossTabulation:
         df_den = (tbl_est.number_psus - tbl_est.number_strata) * df_num
 
         self.stats = {
-            "Pearson-Chisq": {
+            "Pearson-Unadj": {
                 "df": (nrows - 1) * (ncols - 1),
                 "chisq_value": chisq_p,
                 "p_value": chi2.pdf(chisq_p, (nrows - 1) * (ncols - 1)),
             },
-            "Pearson-F": {
+            "Pearson-Adj": {
                 "df_num": df_num,
                 "df_den": df_den,
-                "F_value": f_p,
+                "f_value": f_p,
                 "p_value": f.pdf(f_p, df_num, df_den),
[…]
