+++ tests/categorical/test_twoway.py
@@ -43,6 +43,7 @@ tbl_prop.tabulate([region, birth_cat], varnames=["region", "birth_cat"], remove_
 
 tbl_count = CrossTabulation("count")
 tbl_count.tabulate([age_cat, birth_cat], varnames=["age_cat", "birth_cat"], remove_nan=True)
+tbl_df = tbl_count.to_dataframe()
 
 breakpoint()
 
