+++ tests/categorical/test_crosstabulation.py
@@ -3,7 +3,7 @@ import numpy as np
 import pandas as pd
 
 from samplics.estimation import TaylorEstimator
-from samplics.categorical import OneWay, CrossTabulation
+from samplics.categorical import CrossTabulation
 
 
 birthcat = pd.read_csv("./tests/categorical/birthcat.csv")
