+++ src/samplics/categorical/tabulation.py
@@ -41,7 +41,7 @@ class Tabulation:
         else:
             raise ValueError("parameter must be 'count' or 'proportion'")
         self.table_type = "oneway"
-        self.table = {}  # Dict[str, Dict[StringNumber, Number]]
+        self.point_est = {}  # Dict[str, Dict[StringNumber, Number]]
         self.stats = {}  # Dict[str, Dict[str, Number]]
         self.stderror = {}  # Dict[str, Dict[str, Number]]
         self.lower_ci = {}  # Dict[str, Dict[str, Number]]
@@ -49,9 +49,25 @@ class Tabulation:
         self.deff = {}  # Dict[str, Dict[str, Number]]
         self.alpha = alpha
         self.ciprop_method = ciprop_method
+        self.design_info = {}
+        self.vars_names = []
+        self.vars_levels = {}
 
     def __repr__(self):
         return f"Tabulation(parameter={self.parameter}, alpha={self.alpha})"
+
+    def __str__(self):
+
+        if self.vars_names == []:
+            return "No categorical variables to tabulate"
+        else:
+            tbl_head = f"Tabulation of {self.vars_names[0]}"
+            tbl_subhead1 = f"Number of strata: {self.design_info['number_strata']}"
+            tbl_subhead2 = f"Number of PSUs: {self.design_info['number_psus']}"
+            tbl_subhead3 = f"Number of observations: {self.design_info['number_obs']}"
+            tbl_subhead4 = f"Degrees of freedom: {self.design_info['degrees_of_freedom']}"
+
+            return f"\n{tbl_head}\n {tbl_subhead1}\n {tbl_subhead2}\n {tbl_subhead3}\n {tbl_subhead4}\n\n {self.to_dataframe()}"
 
     def _estimate(
         self,
@@ -65,7 +81,7 @@ class Tabulation:
         deff: bool = False,
         coef_variation: bool = False,
         remove_nan: bool = False,
-    ) -> TaylorEstimator:
+    ) -> Tuple[TaylorEstimator, list, int]:
 
         if remove_nan:
             excluded_units = np.isnan(var) | np.isnan(samp_weight)
@@ -103,7 +119,7 @@ class Tabulation:
         else:
             raise ValueError("parameter must be 'count' or 'proportion'")
 
-        return tbl_est
+        return tbl_est, list(np.unique(var)), var_of_ones.shape[0]
 
     def tabulate(
         self,
@@ -156,7 +172,7 @@ class Tabulation:
             samp_weight = numpy_array(samp_weight)
 
         if nb_vars == 1:
-            tbl_est = self._estimate(
+            tbl_est, var_levels, nb_obs = self._estimate(
                 var_of_ones=np.ones(vars_np.shape[0]),
                 var=vars_np,
                 samp_weight=samp_weight,
@@ -168,15 +184,16 @@ class Tabulation:
                 coef_variation=coef_variation,
                 remove_nan=remove_nan,
             )
+            self.vars_levels[vars_names[0]] = var_levels
             if self.parameter == "count":
+                self.point_est[vars_names[0]] = tbl_est.point_est
                 self.stderror[vars_names[0]] = tbl_est.stderror
-                self.table[vars_names[0]] = tbl_est.point_est
                 self.lower_ci[vars_names[0]] = tbl_est.lower_ci
                 self.upper_ci[vars_names[0]] = tbl_est.upper_ci
                 self.deff[vars_names[0]] = {}  # todo: tbl_est.deff
             elif self.parameter == "proportion":
+                self.point_est[vars_names[0]] = tbl_est.point_est["__none__"]
                 self.stderror[vars_names[0]] = tbl_est.stderror["__none__"]
[…]
