+++ tests/categorical/test_tabulation.py
@@ -27,9 +27,9 @@ tbl_count.tabulate(birth_cat, remove_nan=True)
 
 
 def test_oneway_count_one_var_count():
-    assert tbl_count.table["birthcat"][1] == 240
-    assert tbl_count.table["birthcat"][2] == 450
-    assert tbl_count.table["birthcat"][3] == 233
+    assert tbl_count.point_est["birthcat"][1] == 240
+    assert tbl_count.point_est["birthcat"][2] == 450
+    assert tbl_count.point_est["birthcat"][3] == 233
 
 
 def test_oneway_count_one_var_stderr():
@@ -67,9 +67,9 @@ tbl_prop.tabulate(birth_cat, remove_nan=True)
 
 
 def test_oneway_prop_one_var_prop():
-    assert np.isclose(tbl_prop.table["birthcat"][1], 0.2600, atol=1e-4)
-    assert np.isclose(tbl_prop.table["birthcat"][2], 0.4875, atol=1e-4)
-    assert np.isclose(tbl_prop.table["birthcat"][3], 0.2524, atol=1e-4)
+    assert np.isclose(tbl_prop.point_est["birthcat"][1], 0.2600, atol=1e-4)
+    assert np.isclose(tbl_prop.point_est["birthcat"][2], 0.4875, atol=1e-4)
+    assert np.isclose(tbl_prop.point_est["birthcat"][3], 0.2524, atol=1e-4)
 
 
 def test_oneway_prop_one_var_stderr():
@@ -100,40 +100,40 @@ tbl2_count = Tabulation("count")
 def test_oneway_count_two_vars_list():
     tbl2_count.tabulate([region, birth_cat], remove_nan=True)
 
-    assert tbl2_count.table["var_2"][1] == 240
-    assert tbl2_count.table["var_2"][2] == 450
-    assert tbl2_count.table["var_2"][3] == 233
+    assert tbl2_count.point_est["var_2"][1] == 240
+    assert tbl2_count.point_est["var_2"][2] == 450
+    assert tbl2_count.point_est["var_2"][3] == 233
 
-    assert tbl2_count.table["var_1"][1] == 166
-    assert tbl2_count.table["var_1"][2] == 284
-    assert tbl2_count.table["var_1"][3] == 250
-    assert tbl2_count.table["var_1"][4] == 256
+    assert tbl2_count.point_est["var_1"][1] == 166
+    assert tbl2_count.point_est["var_1"][2] == 284
+    assert tbl2_count.point_est["var_1"][3] == 250
+    assert tbl2_count.point_est["var_1"][4] == 256
 
 
 def test_oneway_count_two_vars_tuple():
     tbl2_count.tabulate((region, birth_cat), remove_nan=True)
 
-    assert tbl2_count.table["var_2"][1] == 240
-    assert tbl2_count.table["var_2"][2] == 450
-    assert tbl2_count.table["var_2"][3] == 233
+    assert tbl2_count.point_est["var_2"][1] == 240
+    assert tbl2_count.point_est["var_2"][2] == 450
+    assert tbl2_count.point_est["var_2"][3] == 233
 
-    assert tbl2_count.table["var_1"][1] == 166
-    assert tbl2_count.table["var_1"][2] == 284
-    assert tbl2_count.table["var_1"][3] == 250
-    assert tbl2_count.table["var_1"][4] == 256
+    assert tbl2_count.point_est["var_1"][1] == 166
+    assert tbl2_count.point_est["var_1"][2] == 284
+    assert tbl2_count.point_est["var_1"][3] == 250
+    assert tbl2_count.point_est["var_1"][4] == 256
 
 
[…]
