### 2021-01-05 26171de
adding tests for twoway class
