+++ tests/categorical/test_crosstabulation.py
@@ -22,10 +22,10 @@ def test_not_valid_parameter(param):
     tbl = CrossTabulation(param)
 
 
-# @pytest.mark.xfail(strict=True, reason="2way tables needs two variables")
-# def test_twoway_count_one_var_count():
-#     tbl = CrossTabulation("count")
-#     tbl.tabulate(region, remove_nan=True)
+@pytest.mark.xfail(strict=True, reason="2way tables needs two variables")
+def test_twoway_count_one_var_count():
+    tbl = CrossTabulation("count")
+    tbl.tabulate(region, remove_nan=True)
 
 
 @pytest.mark.xfail(strict=True, reason="For now, the method will fail if there are missing values")
@@ -34,20 +34,171 @@ def test_for_missing_values_in_the_design_matrix():
     tbl_prop.tabulate([region, birth_cat], varnames=["region", "birth_cat"], remove_nan=False)
 
 
-# tbl_count = CrossTabulation("count")
-# tbl_count.tabulate([region, birth_cat], remove_nan=True)
-
-tbl_prop = CrossTabulation("proportion")
-tbl_prop.tabulate([region, birth_cat], varnames=["region", "birth_cat"], remove_nan=True)
-
-
 tbl_count = CrossTabulation("count")
 tbl_count.tabulate([age_cat, birth_cat], varnames=["age_cat", "birth_cat"], remove_nan=True)
-tbl_df = tbl_count.to_dataframe()
 
-breakpoint()
 
-# def test_oneway_count_one_var_count():
-#     assert tbl_count.table["birthcat"][1] == 240
-#     assert tbl_count.table["birthcat"][2] == 450
-#     assert tbl_count.table["birthcat"][3] == 233
+def test_twoway_count_to_dataframe():
+    tbl_df = tbl_count.to_dataframe()
+
+
+def test_twoway_count_point_est():
+    assert tbl_count.point_est["1.0"]["1.0"] == 56
+    assert tbl_count.point_est["1.0"]["2.0"] == 239
+    assert tbl_count.point_est["1.0"]["3.0"] == 192
+    assert tbl_count.point_est["2.0"]["1.0"] == 79
+    assert tbl_count.point_est["2.0"]["2.0"] == 193
+    assert tbl_count.point_est["2.0"]["3.0"] == 37
+    assert tbl_count.point_est["3.0"]["1.0"] == 105
+    assert tbl_count.point_est["3.0"]["2.0"] == 18
+    assert tbl_count.point_est["3.0"]["3.0"] == 4
+
+
+def test_twoway_count_sderror():
+    assert np.isclose(tbl_count.stderror["1.0"]["1.0"], 7.2567, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["1.0"]["2.0"], 13.3156, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["1.0"]["3.0"], 12.3380, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["2.0"]["1.0"], 8.5039, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["2.0"]["2.0"], 12.3616, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["2.0"]["3.0"], 5.9628, atol=1e-4)
[…]
