### 2021-01-05 ade19bd
adding tests for twoway class
