### 2021-01-06 5d7b465
refactor covariance calculation
