+++ src/samplics/estimation/expansion_old.py
@@ -273,18 +273,15 @@ class TaylorEstimator(_SurveyEstimator):
     ) -> np.ndarray:
         """Provides the scores used to calculate the variance"""
 
-        ncols = 1 if len(y.shape) == 1 else y.shape[1]
-        y = numpy_array(y)
-        y = y.reshape(y.shape[0], ncols)
-        y_weighted = y * samp_weight[:, None]
+        y_weighted = y * samp_weight
         if self.parameter in ("proportion", "mean"):
             scale_weights = np.sum(samp_weight)
-            location_weights = np.sum(y_weighted, axis=0) / scale_weights
-            return (y - location_weights) * samp_weight[:, None] / scale_weights
+            location_weights = np.sum(y_weighted) / scale_weights
+            return samp_weight * (y - location_weights) / scale_weights
         elif self.parameter == "ratio":
             weighted_sum_x = np.sum(x * samp_weight)
-            weighted_ratio = np.sum(y_weighted, axis=0) / weighted_sum_x
-            return samp_weight[:, None] * (y - x[:, None] * weighted_ratio) / weighted_sum_x
+            weighted_ratio = np.sum(y_weighted) / weighted_sum_x
+            return samp_weight * (y - x * weighted_ratio) / weighted_sum_x
         elif self.parameter == "total":
             return y_weighted
         else:
@@ -296,28 +293,23 @@ class TaylorEstimator(_SurveyEstimator):
     ) -> np.float64:
         """Computes the variance for one stratum """
 
+        variance = 0.0
         if number_psus_in_s > 1:
-            scores_s_mean = y_score_s.sum(axis=0) / number_psus_in_s  # new
+            scores_s_mean = y_score_s.sum() / number_psus_in_s
             psus = np.unique(psu_s)
-            scores_psus_sums = np.zeros((number_psus_in_s, scores_s_mean.shape[0]))
+            scores_psus_sums = np.zeros(number_psus_in_s)
             for k, psu in enumerate(np.unique(psus)):
-                scores_psus_sums[k, :] = y_score_s[psu_s == psu].sum(axis=0)
-            covariance = np.transpose(scores_psus_sums - scores_s_mean) @ (
-                scores_psus_sums - scores_s_mean
-            )
-            covariance = (number_psus_in_s / (number_psus_in_s - 1)) * covariance
+                scores_psus_sums[k] = y_score_s[psu_s == psu].sum()
+            variance = ((scores_psus_sums - scores_s_mean) ** 2).sum()
+            variance = (number_psus_in_s / (number_psus_in_s - 1)) * variance
         elif number_psus_in_s in (0, 1):
-            number_obs = y_score_s.shape[0]
-            y_score_s_mean = y_score_s.sum(axis=0) / number_obs
-            covariance = (
-                (number_obs / (number_obs - 1))
-                * np.transpose(y_score_s - y_score_s_mean)
-                @ (y_score_s - y_score_s_mean)
-            )
+            number_obs = y_score_s.size
+            y_score_s_mean = y_score_s.sum() / number_obs
+            variance = (number_obs / (number_obs - 1)) * ((y_score_s - y_score_s_mean) ** 2).sum()
         else:
             raise ValueError("Number of psus cannot be negative.")
 
-        return covariance
+        return variance
 
     @staticmethod
     def _variance_stratum_within(
@@ -358,28 +350,27 @@ class TaylorEstimator(_SurveyEstimator):
 
         if stratum is None:
             number_psus = np.unique(psu).size
-            covariance = fpc["__none__"] * self._variance_stratum_between(
+            var_est = fpc["__none__"] * self._variance_stratum_between(
                 y_score, samp_weight, number_psus, psu
[…]
