### 2021-01-06 ee5ed65
refactor covariance calculation
