### 2021-01-06 430c843
adding estimation of factor-type variables
