+++ src/samplics/categorical/tabulation.py
@@ -282,6 +282,7 @@ class CrossTabulation:
         self.point_est = {}  # Dict[str, Dict[StringNumber, Number]]
         self.stats = {}  # Dict[str, Dict[str, Number]]
         self.stderror = {}  # Dict[str, Dict[str, Number]]
+        self.covariance = {}
         self.lower_ci = {}  # Dict[str, Dict[str, Number]]
         self.upper_ci = {}  # Dict[str, Dict[str, Number]]
         self.deff = {}  # Dict[str, Dict[str, Number]]
@@ -360,6 +361,11 @@ class CrossTabulation:
         else:
             raise AssertionError("varnames should be a string or a list of string")
 
+        samp_weight = numpy_array(samp_weight)
+        stratum = numpy_array(stratum)
+        psu = numpy_array(psu)
+        ssu = numpy_array(ssu)
+
         if remove_nan:
             vars_nans = vars.isna()
             excluded_units = vars_nans.iloc[:, 0] | vars_nans.iloc[:, 1]
@@ -370,6 +376,14 @@ class CrossTabulation:
 
         vars = vars.astype(str)
         vars_names = set_variables_names(vars, varnames, prefix)
+
+        vars.reset_index(inplace=True, drop=True)
+        vars.sort_values(by=vars_names, inplace=True)
+        samp_weight = samp_weight[vars.index]
+        stratum = stratum[vars.index] if stratum is not None else None
+        psu = psu[vars.index] if psu is not None else None
+        ssu = ssu[vars.index] if ssu is not None else None
+
         two_way_full_model = saturated_two_ways_model(vars_names)
         # vars.sort_values(by=vars_names, inplace=True)
         vars.columns = vars_names
@@ -394,14 +408,6 @@ class CrossTabulation:
         )
 
         if self.parameter == "count":
-            tbl_est_srs = TaylorEstimator(parameter="total", alpha=self.alpha)
-            tbl_est_srs.estimate(
-                y=np.ones(vars_for_oneway.shape[0]),
-                samp_weight=np.ones(vars_for_oneway.shape[0]),
-                domain=vars_for_oneway,
-                fpc=fpc,
-            )
-
             tbl_est = TaylorEstimator(parameter="total", alpha=self.alpha)
             tbl_est.estimate(
                 y=np.ones(vars_for_oneway.shape[0]),
@@ -413,16 +419,13 @@ class CrossTabulation:
                 fpc=fpc,
             )
 
-            tbl_keys = list(tbl_est_srs.point_est.keys())
-            cell_est_srs = np.zeros(vars_levels.shape[0])
+            tbl_keys = list(tbl_est.point_est.keys())
             cell_est = np.zeros(vars_levels.shape[0])
             cell_stderror = np.zeros(vars_levels.shape[0])
             cell_lower_ci = np.zeros(vars_levels.shape[0])
             cell_upper_ci = np.zeros(vars_levels.shape[0])
             for k in range(vars_levels.shape[0]):
                 if vars_levels_concat[k] in tbl_keys:
-                    cell_est_srs[k] = tbl_est_srs.point_est[vars_levels_concat[k]]
-
                     cell_est[k] = tbl_est.point_est[vars_levels_concat[k]]
                     cell_est[k] = tbl_est.point_est[vars_levels_concat[k]]
                     cell_stderror[k] = tbl_est.stderror[vars_levels_concat[k]]
@@ -430,13 +433,6 @@ class CrossTabulation:
                     cell_upper_ci[k] = tbl_est.upper_ci[vars_levels_concat[k]]
 
[…]
