### 2021-01-06 acfebf8
adding estimation of factor-type variables
