### 2021-01-07 007ee97
adding estimation of factor-type variables
