+++ src/samplics/categorical/tabulation.py
@@ -376,6 +376,8 @@ class CrossTabulation:
 
         vars = vars.astype(str)
         vars_names = set_variables_names(vars, varnames, prefix)
+        # breakpoint()
+        vars.columns = vars_names
 
         vars.reset_index(inplace=True, drop=True)
         vars.sort_values(by=vars_names, inplace=True)
@@ -386,7 +388,6 @@ class CrossTabulation:
 
         two_way_full_model = saturated_two_ways_model(vars_names)
         # vars.sort_values(by=vars_names, inplace=True)
-        vars.columns = vars_names
         row_levels = vars[vars_names[0]].unique()
         col_levels = vars[vars_names[1]].unique()
 
@@ -408,62 +409,51 @@ class CrossTabulation:
         )
 
         if self.parameter == "count":
-            tbl_est = TaylorEstimator(parameter="total", alpha=self.alpha)
-            tbl_est.estimate(
-                y=np.ones(vars_for_oneway.shape[0]),
-                samp_weight=samp_weight,
-                stratum=stratum,
-                psu=psu,
-                ssu=ssu,
-                domain=vars_for_oneway,
-                fpc=fpc,
-            )
-
-            tbl_keys = list(tbl_est.point_est.keys())
-            cell_est = np.zeros(vars_levels.shape[0])
-            cell_stderror = np.zeros(vars_levels.shape[0])
-            cell_lower_ci = np.zeros(vars_levels.shape[0])
-            cell_upper_ci = np.zeros(vars_levels.shape[0])
-            for k in range(vars_levels.shape[0]):
-                if vars_levels_concat[k] in tbl_keys:
-                    cell_est[k] = tbl_est.point_est[vars_levels_concat[k]]
-                    cell_est[k] = tbl_est.point_est[vars_levels_concat[k]]
-                    cell_stderror[k] = tbl_est.stderror[vars_levels_concat[k]]
-                    cell_lower_ci[k] = tbl_est.lower_ci[vars_levels_concat[k]]
-                    cell_upper_ci[k] = tbl_est.upper_ci[vars_levels_concat[k]]
-
+            parameter = "total"
         elif self.parameter == "proportion":
-            tbl_est = TaylorEstimator(parameter=self.parameter, alpha=self.alpha)
-            tbl_est.estimate(
-                y=vars_for_oneway,
-                samp_weight=samp_weight,
-                stratum=stratum,
-                psu=psu,
-                ssu=ssu,
-                fpc=fpc,
-            )
-
-            tbl_keys = list(tbl_est.point_est["__none__"].keys())
-            cell_est = np.zeros(vars_levels.shape[0])
-            cell_stderror = np.zeros(vars_levels.shape[0])
-            cell_lower_ci = np.zeros(vars_levels.shape[0])
-            cell_upper_ci = np.zeros(vars_levels.shape[0])
-            for k in range(vars_levels.shape[0]):
-                if vars_levels_concat[k] in tbl_keys:
-                    cell_est[k] = tbl_est.point_est["__none__"][vars_levels_concat[k]]
-                    cell_stderror[k] = tbl_est.stderror["__none__"][vars_levels_concat[k]]
-                    cell_lower_ci[k] = tbl_est.lower_ci["__none__"][vars_levels_concat[k]]
-                    cell_upper_ci[k] = tbl_est.upper_ci["__none__"][vars_levels_concat[k]]
+            parameter = "mean"
         else:
             raise ValueError("parameter must be 'count' or 'proportion'")
 
-        cov_srs = (
[…]
