+++ tests/categorical/test_crosstabulation.py
@@ -6,202 +6,202 @@ from samplics.estimation import TaylorEstimator
 from samplics.categorical import CrossTabulation
 
 
-# birthcat = pd.read_csv("./tests/categorical/birthcat.csv")
+birthcat = pd.read_csv("./tests/categorical/birthcat.csv")
 
-# # birthcat.loc[(birthcat["birthcat"] == 2) & (birthcat["region"]==3), "birthcat"] = 3
+# birthcat.loc[(birthcat["birthcat"] == 2) & (birthcat["region"]==3), "birthcat"] = 3
 
-# region = birthcat["region"].to_numpy().astype(int)
-# age_cat = birthcat["agecat"].to_numpy()
-# birth_cat = birthcat["birthcat"].to_numpy()
-# pop = birthcat["pop"]
-
-
-# @pytest.mark.xfail(strict=True, reason="Parameter not valid")
-# @pytest.mark.parametrize("param", ["total", "mean", "ratio", "other"])
-# def test_not_valid_parameter(param):
-#     tbl = CrossTabulation(param)
+region = birthcat["region"].to_numpy().astype(int)
+age_cat = birthcat["agecat"].to_numpy()
+birth_cat = birthcat["birthcat"].to_numpy()
+pop = birthcat["pop"]
 
 
-# @pytest.mark.xfail(strict=True, reason="2way tables needs two variables")
-# def test_twoway_count_one_var_count():
-#     tbl = CrossTabulation("count")
-#     tbl.tabulate(region, remove_nan=True)
+@pytest.mark.xfail(strict=True, reason="Parameter not valid")
+@pytest.mark.parametrize("param", ["total", "mean", "ratio", "other"])
+def test_not_valid_parameter(param):
+    tbl = CrossTabulation(param)
 
 
-# @pytest.mark.xfail(strict=True, reason="For now, the method will fail if there are missing values")
-# def test_for_missing_values_in_the_design_matrix():
-#     tbl_prop = CrossTabulation("proportion")
-#     tbl_prop.tabulate([region, birth_cat], varnames=["region", "birth_cat"], remove_nan=False)
+@pytest.mark.xfail(strict=True, reason="2way tables needs two variables")
+def test_twoway_count_one_var_count():
+    tbl = CrossTabulation("count")
+    tbl.tabulate(region, remove_nan=True)
 
 
-# tbl_count = CrossTabulation("count")
-# tbl_count.tabulate([age_cat, birth_cat], varnames=["age_cat", "birth_cat"], remove_nan=True)
+@pytest.mark.xfail(strict=True, reason="For now, the method will fail if there are missing values")
+def test_for_missing_values_in_the_design_matrix():
+    tbl_prop = CrossTabulation("proportion")
+    tbl_prop.tabulate([region, birth_cat], varnames=["region", "birth_cat"], remove_nan=False)
 
 
-# def test_twoway_count_to_dataframe():
-#     tbl_df = tbl_count.to_dataframe()
+tbl_count = CrossTabulation("count")
[…]

[…]
