### 2021-01-07 e81b854
adding estimation of factor-type variables
