+++ tests/estimation/test_expansion.py
@@ -35,6 +35,7 @@ y = yrbs["y"]
 # fpc_array = yrbs["fpc"]
 # fpc_dict = dict(zip(stratum, fpc_array))
 
+
 """Taylor Approximation WITHOUT Stratification for TOTAL"""
 svy_total_without_str = TaylorEstimator("total")
 
@@ -619,3 +620,387 @@ def test_prop_estimator_with_str_nor_psu_domain():
     assert np.isclose(svy_prop_with_str_domain.upper_ci["d2"][1.0], 0.8206556)
     assert np.isclose(svy_prop_with_str_domain.upper_ci["d3"][0.0], 0.1981554)
     assert np.isclose(svy_prop_with_str_domain.upper_ci["d3"][1.0], 0.8350427)
+
+
+"""Taylor Approximation WITHOUT Stratification for FACTOR-MEAN"""
+svy_factor_mean_without_str = TaylorEstimator("mean")
+
+
+def test_factor_mean_estimator_without_str():
+    svy_factor_mean_without_str.estimate(y, weight, psu=psu, as_factor=True, remove_nan=True)
+
+    assert np.isclose(svy_factor_mean_without_str.point_est["__none__"][0.0], 0.186_377_5)
+    assert np.isclose(svy_factor_mean_without_str.point_est["__none__"][1.0], 0.813_622_5)
+    assert np.isclose(svy_factor_mean_without_str.stderror["__none__"][0.0], 0.020_285_6)
+    assert np.isclose(svy_factor_mean_without_str.stderror["__none__"][1.0], 0.020_285_6)
+    assert np.isclose(svy_factor_mean_without_str.lower_ci["__none__"][0.0], 0.149_023)
+    assert np.isclose(svy_factor_mean_without_str.lower_ci["__none__"][1.0], 0.769_441_4)
+    assert np.isclose(svy_factor_mean_without_str.upper_ci["__none__"][0.0], 0.230_558_6)
+    assert np.isclose(svy_factor_mean_without_str.upper_ci["__none__"][1.0], 0.850_977)
+
+
+def test_factor_mean_estimator_without_str_nor_psu():
+    svy_factor_mean_without_str.estimate(y, weight, as_factor=True, remove_nan=True)
+
+    assert np.isclose(svy_factor_mean_without_str.point_est["__none__"][0.0], 0.1863775)
+    assert np.isclose(svy_factor_mean_without_str.point_est["__none__"][1.0], 0.8136225)
+    assert np.isclose(svy_factor_mean_without_str.stderror["__none__"][0.0], 0.0066567)
+    assert np.isclose(svy_factor_mean_without_str.stderror["__none__"][1.0], 0.0066567)
+    assert np.isclose(svy_factor_mean_without_str.lower_ci["__none__"][0.0], 0.1736793)
+    assert np.isclose(svy_factor_mean_without_str.lower_ci["__none__"][1.0], 0.8002204)
+    assert np.isclose(svy_factor_mean_without_str.upper_ci["__none__"][0.0], 0.1997796)
+    assert np.isclose(svy_factor_mean_without_str.upper_ci["__none__"][1.0], 0.8263207)
+
+
+svy_factor_mean_without_str_domain = TaylorEstimator("mean")
+
+
+def test_factor_mean_estimator_without_str_domain():
+    svy_factor_mean_without_str_domain.estimate(
+        y, weight, psu=psu, domain=domain, as_factor=True, remove_nan=True
+    )
+
+    assert np.isclose(svy_factor_mean_without_str_domain.point_est["d1"][0.0], 0.1688402)
[…]
