### 2021-01-07 4c1f7b5
adding estimation of factor-type variables
