### 2021-01-07 bacd251
refactor to remove method _remove_nans from _SurveyEstimator class
