+++ src/samplics/estimation/replication.py
@@ -19,27 +19,27 @@ class ReplicateEstimator(_SurveyEstimator):
         | point_est (dict): point estimate of the parameter of interest.
         | variance (dict): variance estimate of the parameter of interest.
         | stderror (dict): standard error of the parameter of interest.
-        | coef_var (dict): estimate of the coefficient of variation.  
+        | coef_var (dict): estimate of the coefficient of variation.
         | deff (dict): estimate of the design effect due to weighting.
-        | lower_ci (dict): estimate of the lower bound of the confidence interval. 
-        | upper_ci (dict): estimate of the upper bound of the confidence interval. 
+        | lower_ci (dict): estimate of the lower bound of the confidence interval.
+        | upper_ci (dict): estimate of the upper bound of the confidence interval.
         | degree_of_freedom (int): degree of freedom for the confidence interval.
         | alpha (float): significant level for the confidence interval
         | strata (list): list of the strata in the sample.
-        | domains (list): list of the domains in the sample. 
-        | method (str): variance estimation method. 
-        | parameter (str): the parameter of the population to estimate e.g. total. 
+        | domains (list): list of the domains in the sample.
+        | method (str): variance estimation method.
+        | parameter (str): the parameter of the population to estimate e.g. total.
         | number_strata (int): number of strata.
         | number_psus (int): number of primary sampling units (psus).
-        | conservative (bool): indicate whether to produce conservative variance estimates. 
+        | conservative (bool): indicate whether to produce conservative variance estimates.
         | number_reps (int): number of replicate weights.
-        | rep_coefs (array): coefficients associated to the replicate weights. 
+        | rep_coefs (array): coefficients associated to the replicate weights.
         | fay_coef (float): Fay coefficient for the the BRR-Fay algorithm.
-        | random_seed (int): random seed for reproducibility. 
+        | random_seed (int): random seed for reproducibility.
 
     Methods
         | estimate(): produces the point estimate of the parameter of interest with the associated
-        |   measures of precision. 
+        |   measures of precision.
     """
 
     def __init__(
@@ -83,7 +83,11 @@ class ReplicateEstimator(_SurveyEstimator):
             )
 
     def _bias(
-        self, y: np.ndarray, samp_weight: np.ndarray, rep_weights: np.ndarray, x: np.ndarray,
+        self,
+        y: np.ndarray,
+        samp_weight: np.ndarray,
+        rep_weights: np.ndarray,
+        x: np.ndarray,
     ) -> float:
 
         estimate = self._get_point(y, samp_weight, x).get("__none__")
@@ -132,7 +136,10 @@ class ReplicateEstimator(_SurveyEstimator):
                 variance = float(
                     np.sum(
                         rep_coefs
-                        * pow((pseudo_estimates - np.mean(pseudo_estimates)) / (jk_factor - 1), 2,)
+                        * pow(
+                            (pseudo_estimates - np.mean(pseudo_estimates)) / (jk_factor - 1),
+                            2,
+                        )
                     )
                 )
         else:
@@ -202,7 +209,10 @@ class ReplicateEstimator(_SurveyEstimator):
                         cat_dict_d_k = dict(
                             {
                                 categories[k]: self._bias(
-                                    y_dummies_d[:, k], samp_weight_d, rep_weights_d, None,
+                                    y_dummies_d[:, k],
+                                    samp_weight_d,
+                                    rep_weights_d,
+                                    None,
                                 )
                             }
                         )
@@ -291,7 +301,12 @@ class ReplicateEstimator(_SurveyEstimator):
                     y_d = y * (domain == d)
[…]

[…]
