### 2021-01-07 d3e67b0
refactor to remove method _remove_nans from _SurveyEstimator class
