### 2021-01-07 9787143
refactor to remove method _remove_nans from _SurveyEstimator class
