+++ tests/categorical/test_comparison.py
@@ -1,0 +1,9 @@
+
+
+import pandas as pd 
+
+from samplics.categorical.comparison import Ttest
+
+auto = pd.read_csv("./tests/categorical/auto.csv")
+
+print(auto)
