+++ src/samplics/estimation/expansion_old.py
@@ -1,563 +1,0 @@
-"""Expansion module 
-
-The module implements the taylor-based variance estimation methods. There are several good books
-that provide the the taylor-based approximations of the sample variance. Some of these reference 
-books are Cochran, W.G. (1977) [#c1977]_, Kish, L. (1965) [#k1965]_, Lohr, S.L. (2010) [#l2010]_, 
-and Wolter, K.M. (2007) [#w2007]_. 
-
-.. [#c1977] Cochran, W.G. (1977), *Sampling Techniques, 3rd edn.*, Jonh Wiley & Sons, Inc.
-.. [#k1965] Kish, L. (1965), *Survey Sampling*, Jonh Wiley & Sons, Inc.
-.. [#l2010] Lohr, S.L. (2010), *Sampling: Design and Analysis, 2nd edn.*, Cengage Learning, Inc.
-.. [#w2007] Wolter, K.M. (2007), *Introduction to Variance Estimation, 2nd edn.*, 
-   Springer-Verlag New York, Inc
-"""
-
-import math
-from typing import TypeVar, Any, Dict, List, Optional, Union, Tuple
-
-import numpy as np
-import pandas as pd
-from samplics.utils.formats import fpc_as_dict, numpy_array
-from samplics.utils.types import Array, Number, StringNumber
-from scipy.stats import t as student
-
-TypeTaylorEst = TypeVar("TypeTaylorEst", bound="TaylorEstimator")
-
-
-class _SurveyEstimator:
-    """General approach for sample estimation of linear parameters."""
-
-    def __init__(self, parameter: str, alpha: float = 0.05, random_seed: int = None) -> None:
-        """Initializes the instance """
-
-        if random_seed is not None:
-            self.random_seed = random_seed
-            np.random.seed(self.random_seed)
-        else:
-            self.random_seed = None
-
-        if parameter.lower() in ("proportion", "mean", "total", "ratio"):
-            self.parameter = parameter.lower()
-        else:
-            raise AssertionError("parameter must be 'proportion', 'mean', 'total' or 'ratio'")
-
-        self.point_est: Dict[StringNumber, Any] = {}
-        self.variance: Dict[StringNumber, Any] = {}
-        self.stderror: Dict[StringNumber, Any] = {}
-        self.coef_var: Dict[StringNumber, Any] = {}
-        self.deff: Dict[StringNumber, Any] = {}
-        self.lower_ci: Dict[StringNumber, Any] = {}
-        self.upper_ci: Dict[StringNumber, Any] = {}
-        self.fpc: Dict[StringNumber, Any] = {}
-        self.strata: List[StringNumber] = ["__none__"]
-        self.domains: List[StringNumber] = ["__none__"]
-        self.method: str = "taylor"
-        self.number_strata: Optional[int] = None
-        self.number_psus: Optional[int] = None
-        self.degree_of_freedom: Optional[int] = None
-        self.alpha: float = alpha
-
-    def __str__(self) -> Any:
-        print(f"SAMPLICS - Estimation of {self.parameter.title()}\n")
-        print(f"Number of strata: {self.number_strata}")
-        print(f"Number of psus: {self.number_psus}")
-        print(f"Degree of freedom: {self.degree_of_freedom}\n")
-        parameter = self.parameter.upper()
-        estimation = pd.DataFrame()
-        estimation["DOMAINS"] = self.domains
-        estimation[parameter] = self.point_est.values()
-        estimation["SE"] = self.stderror.values()
-        estimation["LCI"] = self.lower_ci.values()
-        estimation["UCI"] = self.upper_ci.values()
-        estimation["CV"] = self.coef_var.values()
-
-        return "%s" % estimation
-
[…]
