+++ src/samplics/utils/formats.py
@@ -115,21 +115,26 @@ def sample_units(all_units: Array, unique: bool = True) -> np.ndarray:
 
 def dict_to_dataframe(col_names: List[str], *args: Dict[Any, Number]) -> pd.DataFrame:
 
-    keys = list(args[0].keys())
-    number_keys = len(keys)
-    values = list()
-    for k, arg in enumerate(args):
-        argk_keys = list(args[k].keys())
-        if not isinstance(arg, dict) or (keys != argk_keys) or number_keys != len(argk_keys):
-            raise AssertionError("All input parameters must be dictionaries with the same keys.")
+    if isinstance(args[0], dict):
+        keys = list(args[0].keys())
+        number_keys = len(keys)
+        values = list()
+        for k, arg in enumerate(args):
+            argk_keys = list(args[k].keys())
+            if not isinstance(arg, dict) or (keys != argk_keys) or number_keys != len(argk_keys):
+                raise AssertionError("All input parameters must be dictionaries with the same keys.")
 
-        values.append(list(arg.values()))
+            values.append(list(arg.values()))
 
-    values_df = pd.DataFrame(
-        values,
-    ).T
-    values_df.insert(0, "00", keys)
+        values_df = pd.DataFrame(
+            values,
+        ).T
+        values_df.insert(0, "00", keys)
+    else:
+        values_df = pd.DataFrame({args})
+    
     values_df.columns = col_names
+
 
     return values_df
 
