+++ src/samplics/sampling/size.py
@@ -2,7 +2,7 @@
 
 """
 
-from typing import Any, Dict, List, Union, Optional
+from typing import Any, Dict, List, Union, Optional, Generic, overload
 
 import math
 
@@ -12,12 +12,11 @@ import pandas as pd
 from scipy.stats import norm as normal
 
 from samplics.utils import formats
-from samplics.utils.types import Array, Number
+from samplics.utils.types import Array, Number, StringNumber
 
 
-class SampleSize:
-    """*SampleSize* implements sample size calculation methods
-    """
+class SampleSize(Generic[Number, StringNumber]):
+    """*SampleSize* implements sample size calculation methods"""
 
     def __init__(
         self, parameter: str = "proportion", method: str = "wald", stratification: bool = False
@@ -29,24 +28,26 @@ class SampleSize:
             raise AssertionError("Sample size calculation method not valid.")
 
         self.stratification = stratification
-        self.deff_c: Dict[Any, float] = {"__none__": 1.0}
-        self.deff_w: Dict[Any, float] = {"__none__": 1.0}
+        self.deff_c: Union[Number, Dict[StringNumber, Number]] = 1
+        self.deff_w: Union[Number, Dict[StringNumber, Number]] = 1
 
-        self.samp_size: Dict[Any, int]
+        self.samp_size: Union[int, Dict[StringNumber, int]]
 
-    def icc(self) -> Union[Dict[Any, Number], Number]:
+    def icc(self) -> Union[Dict[StringNumber, Number], Number]:
         pass
 
     def deff(
-        self, cluster_size: Union[Dict[Any, Number], Number], icc: Union[Dict[Any, Number], Number]
-    ) -> Union[Dict[Any, Number], Number]:
+        self,
+        cluster_size: Union[Dict[StringNumber, Number], Number],
+        icc: Union[Dict[StringNumber, Number], Number],
+    ) -> Union[Dict[StringNumber, Number], Number]:
 
         if isinstance(cluster_size, (int, float)) and isinstance(icc, (int, float)):
             return max(1 + (cluster_size - 1) * icc, 0)
         elif isinstance(cluster_size, dict) and isinstance(icc, dict):
             if cluster_size.keys() != icc.keys():
                 raise AssertionError("Parameters do not have the same dictionary keys.")
-            deff_c: Dict[Any, Number] = {}
+            deff_c: Dict[StringNumber, Number] = {}
             for s in cluster_size:
                 deff_c[s] = max(1 + (cluster_size[s] - 1) * icc[s], 0)
             return deff_c
@@ -54,31 +55,40 @@ class SampleSize:
             raise ValueError("Combination of types not supported.")
 
     def _calculate_wald(
-        self, target: Dict[Any, Number], precision: Dict[Any, Number], stratum: Optional[Array],
-    ) -> Dict[Any, int]:
+        self,
+        target: Union[Dict[StringNumber, Number], Number],
+        precision: Union[Dict[StringNumber, Number], Number],
+        stratum: Optional[Array],
+    ) -> Union[int, Dict[StringNumber, int]]:
 
         z_value = normal().ppf(1 - self.alpha / 2)
 
-        samp_size: Dict[Any, int] = {}
-        if self.stratification:
+        if (
+            isinstance(target, dict)
+            and isinstance(precision, dict)
+            and isinstance(self.deff_c, dict)
+        ):
+            samp_size: Dict[StringNumber, int] = {}
             for s in stratum:
                 samp_size[s] = math.ceil(
                     self.deff_c[s] * z_value ** 2 * target[s] * (1 - target[s]) / precision[s] ** 2
                 )
+            return samp_size
+        elif (
+            isinstance(target, (int, float))
[…]
