+++ src/samplics/sampling/selection.py
@@ -25,32 +25,35 @@ from samplics.utils.types import Array, StringNumber
 
 
 class SampleSelection:
-    """*SampleSelection* implements a number of sampling selection algorithms. 
+    """*SampleSelection* implements a number of sampling selection algorithms.
 
-    The implemented sampling algorithms are the simple random sampling (srs), the 
-    systematic selection (sys), five different algorithms of probability proportional 
-    to size (pps), and the generic selection based on probabilities of selection. 
+    The implemented sampling algorithms are the simple random sampling (srs), the
+    systematic selection (sys), five different algorithms of probability proportional
+    to size (pps), and the generic selection based on probabilities of selection.
 
     Attributes of the class:
-        | method (str): a string specifying the sampling algorithm. The available 
+        | method (str): a string specifying the sampling algorithm. The available
         |   sampling algorithms are: simple random sampling (srs), systematic selection (sys),
-        |   Brewer's procedure (pps-brewer), Hanurav-Vijayan procedure (pps-hv), 
-        |   Murphy's procedure (pps-murphy), Rao-Sampford procedure (pps-rs), systematic 
-        |   pps (pps-sys), and a generic selection from arbitrary probabilities using 
-        |   *numpy.random.choice()* method. 
-        | stratification (bool): indicates if the sampling procedure is stratified or not. 
-        | with_replacement (bool): indicates whether the selection is with replacement. 
+        |   Brewer's procedure (pps-brewer), Hanurav-Vijayan procedure (pps-hv),
+        |   Murphy's procedure (pps-murphy), Rao-Sampford procedure (pps-rs), systematic
+        |   pps (pps-sys), and a generic selection from arbitrary probabilities using
+        |   *numpy.random.choice()* method.
+        | stratification (bool): indicates if the sampling procedure is stratified or not.
+        | with_replacement (bool): indicates whether the selection is with replacement.
 
     Main functions:
-        | inclusion_probs(): provides the inclusion probabilities. 
+        | inclusion_probs(): provides the inclusion probabilities.
         | joint_inclusion_probs(): provides the joint probalities of selection.
-        | select(): indicates the selected sample. 
+        | select(): indicates the selected sample.
 
     TODO: handling of certainties and implementation of joint_inclusion_probs().
     """
 
     def __init__(
-        self, method: str, stratification: bool = False, with_replacement: bool = True,
+        self,
+        method: str,
+        stratification: bool = False,
+        with_replacement: bool = True,
     ) -> None:
         if method.lower() in (
             "srs",
@@ -70,16 +73,6 @@ class SampleSelection:
         self.stratification = True if stratification else False
         self.with_replacement = with_replacement
         self.fpc: Dict[Any, float] = {}
-
-    @staticmethod
-    def _convert_to_dict(obj_any: Any, obj_type: type) -> Dict[Any, Any]:
-
-        if obj_any is not None and isinstance(obj_any, obj_type):
-            return {"__none__": obj_any}
-        elif obj_any is not None and isinstance(obj_any, Dict):
-            return obj_any
-        elif not isinstance(obj_any, dict):
-            raise TypeError(f"{str(obj_any)} must be a dictionary or {obj_type.__name__}.")
 
     @staticmethod
     def _to_dataframe(
@@ -110,22 +103,23 @@ class SampleSelection:
         return df
 
     def _calculate_fpc(
-        self, samp_unit: np.ndarray, samp_size: Union[Dict[Any, int], int], stratum: np.ndarray,
+        self,
+        samp_unit: np.ndarray,
+        samp_size: Union[Dict[Any, int], int],
+        stratum: np.ndarray,
     ) -> None:
 
         samp_unit = formats.sample_units(samp_unit, unique=True)
         samp_size = formats.sample_size_dict(samp_size, self.stratification, stratum)
[…]
