### 2021-01-08 68c0315
Refactoring to remove use of '__none_'
