### 2021-01-08 2c75a51
Refactoring to remove use of '__none_'
