+++ src/samplics/estimation/expansion.py
@@ -40,6 +40,7 @@ class _SurveyEstimator:
             self.parameter = parameter.lower()
         else:
             raise AssertionError("parameter must be 'proportion', 'mean', 'total' or 'ratio'")
+        self.alpha: float = alpha
 
         self.point_est: Dict[StringNumber, Any] = {}
         self.variance: Dict[StringNumber, Any] = {}
@@ -49,13 +50,12 @@ class _SurveyEstimator:
         self.lower_ci: Dict[StringNumber, Any] = {}
         self.upper_ci: Dict[StringNumber, Any] = {}
         self.fpc: Dict[StringNumber, Any] = {}
-        self.strata: List[StringNumber] = ["__none__"]
-        self.domains: List[StringNumber] = ["__none__"]
+        self.strata: List[StringNumber] = []
+        self.domains: List[StringNumber] = []
         self.method: str = "taylor"
         self.number_strata: Optional[int] = None
         self.number_psus: Optional[int] = None
         self.degree_of_freedom: Optional[int] = None
-        self.alpha: float = alpha
 
     def __str__(self) -> Any:
         print(f"SAMPLICS - Estimation of {self.parameter.title()}\n")
@@ -153,7 +153,6 @@ class _SurveyEstimator:
             categories = None
             y_dummies = None
 
-        estimate: Dict[StringNumber, Any] = {}
         if domain is None:
             if self.parameter == "proportion" or as_factor:
                 cat_dict = dict()
@@ -161,10 +160,11 @@ class _SurveyEstimator:
                     y_k = y_dummies[:, k]
                     cat_dict_k = dict({categories[k]: self._get_point_d(y_k, samp_weight)})
                     cat_dict.update(cat_dict_k)
-                estimate["__none__"] = cat_dict
+                estimate = cat_dict
             else:
-                estimate["__none__"] = self._get_point_d(y, samp_weight, x)
+                estimate = self._get_point_d(y, samp_weight, x)
         else:
+            estimate: Dict[StringNumber, Any] = {}
             domain_ids = np.unique(domain)
             for d in domain_ids:
                 weight_d = samp_weight[domain == d]
@@ -307,16 +307,16 @@ class TaylorEstimator(_SurveyEstimator):
         stratum: np.ndarray,
         psu: np.ndarray,
         ssu: Optional[np.ndarray] = None,
-        fpc: Dict[StringNumber, Number] = {"__none__": 1},
+        fpc: Union[Dict[StringNumber, Number], Number] = 1,
     ) -> np.float64:
         """Computes the variance across stratum """
 
         if stratum is None:
             number_psus = np.unique(psu).size
-            covariance = fpc["__none__"] * self._variance_stratum_between(
+            covariance = fpc * self._variance_stratum_between(
                 y_score, samp_weight, number_psus, psu
             )
-            # + (1 - fpc["__none__"]) * self._variance_stratum_within(
+            # + (1 - fpc) * self._variance_stratum_within(
             #     y_score, number_psus, psu, ssu
             # )
         else:
@@ -345,10 +345,10 @@ class TaylorEstimator(_SurveyEstimator):
         psu: Optional[Array] = None,
         ssu: Optional[Array] = None,
         domain: Optional[Array] = None,
-        fpc: Dict[StringNumber, Number] = {"__none__": 1},
+        fpc: Dict[StringNumber, Number] = 1,
         as_factor: bool = False,
         remove_nan: bool = False,
-    ) -> Tuple[dict, dict]:
+    ) -> Tuple[Union[float, dict], Union[float, dict]]:
 
         if self.parameter == "ratio" and x is None:
             raise AssertionError("Parameter x must be provided for ratio estimation.")
[…]
