### 2021-01-08 fd9caae
Refactoring to remove use of '__none_'
