+++ tests/estimation/test_expansion_fpc.py
@@ -26,19 +26,19 @@ svy_total_without_str = TaylorEstimator("total")
 def test_total_estimator_without_str():
     svy_total_without_str.estimate(y, weight, psu=psu, fpc=fpc_psu)
 
-    assert np.isclose(svy_total_without_str.point_est["__none__"], 853.12)
-    assert np.isclose(svy_total_without_str.stderror["__none__"], 153.2627)
-    assert np.isclose(svy_total_without_str.lower_ci["__none__"], 427.5946)
-    assert np.isclose(svy_total_without_str.upper_ci["__none__"], 1278.645)
+    assert np.isclose(svy_total_without_str.point_est, 853.12)
+    assert np.isclose(svy_total_without_str.stderror, 153.2627)
+    assert np.isclose(svy_total_without_str.lower_ci, 427.5946)
+    assert np.isclose(svy_total_without_str.upper_ci, 1278.645)
 
 
 def test_total_estimator_without_str_nor_psu():
     svy_total_without_str.estimate(y, weight, fpc=fpc_no_psu)
 
-    assert np.isclose(svy_total_without_str.point_est["__none__"], 853.12)
-    assert np.isclose(svy_total_without_str.stderror["__none__"], 67.14785)
-    assert np.isclose(svy_total_without_str.lower_ci["__none__"], 713.052)
-    assert np.isclose(svy_total_without_str.upper_ci["__none__"], 993.188)
+    assert np.isclose(svy_total_without_str.point_est, 853.12)
+    assert np.isclose(svy_total_without_str.stderror, 67.14785)
+    assert np.isclose(svy_total_without_str.lower_ci, 713.052)
+    assert np.isclose(svy_total_without_str.upper_ci, 993.188)
 
 
 svy_total_without_str_domain = TaylorEstimator("total")
@@ -77,19 +77,19 @@ svy_total_with_str = TaylorEstimator("total")
 def test_total_estimator_with_str():
     svy_total_with_str.estimate(y, weight, stratum=stratum, psu=psu, fpc=fpc_dict_psu)
 
-    assert np.isclose(svy_total_with_str.point_est["__none__"], 853.12)
-    assert np.isclose(svy_total_with_str.stderror["__none__"], 97.41386)
-    assert np.isclose(svy_total_with_str.lower_ci["__none__"], 632.75452)
-    assert np.isclose(svy_total_with_str.upper_ci["__none__"], 1073.48547)
+    assert np.isclose(svy_total_with_str.point_est, 853.12)
+    assert np.isclose(svy_total_with_str.stderror, 97.41386)
+    assert np.isclose(svy_total_with_str.lower_ci, 632.75452)
+    assert np.isclose(svy_total_with_str.upper_ci, 1073.48547)
 
 
 def test_total_estimator_with_str_without_psu():
     svy_total_with_str.estimate(y, weight, stratum=stratum, fpc=fpc_dict_no_psu)
 
-    assert np.isclose(svy_total_with_str.point_est["__none__"], 853.12)
-    assert np.isclose(svy_total_with_str.stderror["__none__"], 50.8888)
-    assert np.isclose(svy_total_with_str.lower_ci["__none__"], 746.2066)
-    assert np.isclose(svy_total_with_str.upper_ci["__none__"], 960.0334)
+    assert np.isclose(svy_total_with_str.point_est, 853.12)
+    assert np.isclose(svy_total_with_str.stderror, 50.8888)
[…]
