### 2021-01-09 71b53ef
Refactoring to remove use of '__none_'
