+++ src/samplics/estimation/replication.py
@@ -90,7 +90,7 @@ class ReplicateEstimator(_SurveyEstimator):
         x: np.ndarray,
     ) -> float:
 
-        estimate = self._get_point(y, samp_weight, x).get("__none__")
+        estimate = self._get_point(y, samp_weight, x)
         rep_estimates = self._rep_point(y, rep_weights, x)
 
         return float(np.sum(np.mean(rep_estimates) - estimate))
@@ -191,9 +191,9 @@ class ReplicateEstimator(_SurveyEstimator):
                         }
                     )
                     cat_dict.update(cat_dict_k)
-                bias["__none__"] = cat_dict
+                bias = cat_dict
             else:
-                bias["__none__"] = self._bias(y, samp_weight, rep_weights, x)
+                bias = self._bias(y, samp_weight, rep_weights, x)
         else:
             for d in np.unique(domain):
                 samp_weight_d = samp_weight * (domain == d)
@@ -248,7 +248,7 @@ class ReplicateEstimator(_SurveyEstimator):
             if self.parameter == "proportion":
                 cat_dict = dict()
                 for k in range(categories.size):
-                    estimate_k = self._get_point(y_dummies[:, k], samp_weight, x)["__none__"][1]
+                    estimate_k = self._get_point(y_dummies[:, k], samp_weight, x)[1]
                     cat_dict_k = dict(
                         {
                             categories[k]: self._variance(
@@ -262,12 +262,10 @@ class ReplicateEstimator(_SurveyEstimator):
                         }
                     )
                     cat_dict.update(cat_dict_k)
-                variance["__none__"] = cat_dict
+                variance = cat_dict
             else:
-                estimate = self._get_point(y, samp_weight, x).get("__none__")
-                variance["__none__"] = self._variance(
-                    y, rep_weights, rep_coefs, x, estimate, conservative
-                )
+                estimate = self._get_point(y, samp_weight, x)
+                variance = self._variance(y, rep_weights, rep_coefs, x, estimate, conservative)
         else:
             for d in np.unique(domain):
                 samp_weight_d = samp_weight * (domain == d)
@@ -280,9 +278,7 @@ class ReplicateEstimator(_SurveyEstimator):
                     y_dummies_d = y_dummies * (domain == d)[:, None]
                     cat_dict = dict()
                     for k in range(categories.size):
-                        estimate_d_k = self._get_point(y_dummies_d[:, k], samp_weight_d, x_d).get(
-                            "__none__"
-                        )[1]
+                        estimate_d_k = self._get_point(y_dummies_d[:, k], samp_weight_d, x_d)[1]
                         cat_dict_d_k = dict(
                             {
                                 categories[k]: self._variance(
@@ -299,7 +295,7 @@ class ReplicateEstimator(_SurveyEstimator):
                     variance[d] = cat_dict
                 else:
                     y_d = y * (domain == d)
-                    estimate_d = self._get_point(y_d, samp_weight_d, x_d).get("__none__")
+                    estimate_d = self._get_point(y_d, samp_weight_d, x_d)
                     variance[d] = self._variance(
                         y_d,
                         rep_weights_d,
@@ -311,54 +307,78 @@ class ReplicateEstimator(_SurveyEstimator):
 
         return variance
 
-    @staticmethod
     def _get_confint(
+        self,
         parameter: str,
-        estimate: Dict[StringNumber, Any],
-        variance: Dict[StringNumber, Any],
+        estimate: Union[Dict[StringNumber, Number], Number],
+        variance: Union[Dict[StringNumber, Number], Number],
         quantile: float,
     ) -> Tuple[Dict[StringNumber, Any], Dict[StringNumber, Any]]:
 
-        lower_ci: Dict[StringNumber, Any] = {}
-        upper_ci: Dict[StringNumber, Any] = {}
[…]
