### 2021-01-09 4d92de1
Refactoring to remove use of '__none_'
