### 2021-01-09 1852646
Refactoring to remove use of '__none_'
