### 2021-01-09 75cbfcf
Refactoring to remove use of '__none_'
