+++ src/samplics/estimation/expansion.py
@@ -13,7 +13,7 @@ and Wolter, K.M. (2007) [#w2007]_.
 """
 
 import math
-from typing import TypeVar, Any, Dict, List, Optional, Union, Tuple
+from typing import TypeVar, Generic, Any, Dict, List, Optional, Union, Tuple
 
 import numpy as np
 import pandas as pd
@@ -21,13 +21,11 @@ from samplics.utils.formats import fpc_as_dict, numpy_array, remove_nans, sample
 from samplics.utils.types import Array, Number, Series, StringNumber
 from scipy.stats import t as student
 
-TypeTaylorEst = TypeVar("TypeTaylorEst", bound="TaylorEstimator")
 
-
-class _SurveyEstimator:
+class _SurveyEstimator(Generic[Number, StringNumber]):
     """General approach for sample estimation of linear parameters."""
 
-    def __init__(self, parameter: str, alpha: float = 0.05, random_seed: int = None) -> None:
+    def __init__(self, parameter: str, alpha: float = 0.05, random_seed: Optional[int] = None):
         """Initializes the instance """
 
         if random_seed is not None:
@@ -40,22 +38,23 @@ class _SurveyEstimator:
             self.parameter = parameter.lower()
         else:
             raise AssertionError("parameter must be 'proportion', 'mean', 'total' or 'ratio'")
-        self.alpha: float = alpha
+        self.alpha = alpha
 
-        self.point_est: Dict[StringNumber, Any] = {}
-        self.variance: Dict[StringNumber, Any] = {}
-        self.stderror: Dict[StringNumber, Any] = {}
-        self.coef_var: Dict[StringNumber, Any] = {}
-        self.deff: Dict[StringNumber, Any] = {}
-        self.lower_ci: Dict[StringNumber, Any] = {}
-        self.upper_ci: Dict[StringNumber, Any] = {}
-        self.fpc: Dict[StringNumber, Any] = {}
-        self.strata: List[StringNumber] = []
-        self.domains: Optional[List[StringNumber]] = None
-        self.method: str = "taylor"
-        self.number_strata: Optional[int] = None
-        self.number_psus: Optional[int] = None
-        self.degree_of_freedom: Optional[int] = None
+        self.point_est: Any = {}
+        self.variance: Any = {}
+        self.covariance: Any = {}
+        self.stderror: Any = {}
+        self.coef_var: Any = {}
+        self.deff: Any = {}
+        self.lower_ci: Any = {}
+        self.upper_ci: Any = {}
+        self.fpc: Any = {}
+        self.strata: Any = []
+        self.domains: Any = None
+        self.method: Any = "taylor"
+        self.number_strata: Any = None
+        self.number_psus: Any = None
+        self.degree_of_freedom: Any = None
 
     def __str__(self) -> Any:
         print(f"SAMPLICS - Estimation of {self.parameter.title()}\n")
@@ -64,12 +63,25 @@ class _SurveyEstimator:
         print(f"Degree of freedom: {self.degree_of_freedom}\n")
         parameter = self.parameter.upper()
         estimation = pd.DataFrame()
-        estimation["DOMAINS"] = self.domains
-        estimation[parameter] = self.point_est.values()
-        estimation["SE"] = self.stderror.values()
-        estimation["LCI"] = self.lower_ci.values()
-        estimation["UCI"] = self.upper_ci.values()
-        estimation["CV"] = self.coef_var.values()
+        if (
+            isinstance(self.point_est, dict)
+            and isinstance(self.stderror, dict)
+            and isinstance(self.lower_ci, dict)
+            and isinstance(self.upper_ci, dict)
+            and isinstance(self.coef_var, dict)
+        ):
+            estimation["DOMAINS"] = self.domains
[…]
