+++ src/samplics/sampling/selection.py
@@ -13,7 +13,8 @@ sampling algorithms.
    Springer-Verlag New York, Inc
 """
 
-from typing import Any, Dict, Tuple, Union, Optional
+from ast import Str
+from typing import Dict, Generic, Tuple, Union, Optional, overload
 
 import numpy as np
 import pandas as pd
@@ -21,10 +22,10 @@ import pandas as pd
 import math
 
 from samplics.utils import formats
-from samplics.utils.types import Array, StringNumber
+from samplics.utils.types import Array, Number, Series, StringNumber
 
 
-class SampleSelection:
+class SampleSelection(Generic[Number, StringNumber]):
     """*SampleSelection* implements a number of sampling selection algorithms.
 
     The implemented sampling algorithms are the simple random sampling (srs), the
@@ -43,18 +44,18 @@ class SampleSelection:
 
     Main functions:
         | inclusion_probs(): provides the inclusion probabilities.
-        | joint_inclusion_probs(): provides the joint probalities of selection.
+        | joNumber_inclusion_probs(): provides the joNumber probalities of selection.
         | select(): indicates the selected sample.
 
-    TODO: handling of certainties and implementation of joint_inclusion_probs().
+    TODO: handling of certaNumberies and implementation of joNumber_inclusion_probs().
     """
 
     def __init__(
         self,
-        method: str,
-        stratification: bool = False,
-        with_replacement: bool = True,
-    ) -> None:
+        method,
+        stratification=False,
+        with_replacement=True,
+    ):
         if method.lower() in (
             "srs",
             "sys",
@@ -72,16 +73,16 @@ class SampleSelection:
             )
         self.stratification = True if stratification else False
         self.with_replacement = with_replacement
-        self.fpc: Dict[Any, float] = {}
+        self.fpc = {}
 
     @staticmethod
     def _to_dataframe(
-        samp_unit: np.ndarray,
-        stratum: np.ndarray,
-        mos: np.ndarray,
-        sample: np.ndarray,
-        hits: np.ndarray,
-        probs: np.ndarray,
+        samp_unit,
+        stratum,
+        mos,
+        sample,
+        hits,
+        probs,
     ) -> pd.DataFrame:
 
         df = pd.DataFrame(
@@ -104,15 +105,15 @@ class SampleSelection:
 
     def _calculate_fpc(
         self,
-        samp_unit: np.ndarray,
-        samp_size: Union[Dict[Any, int], int],
-        stratum: np.ndarray,
-    ) -> None:
+        samp_unit,
+        samp_size,
+        stratum,
+    ):
 
         samp_unit = formats.sample_units(samp_unit, unique=True)
         samp_size = formats.sample_size_dict(samp_size, self.stratification, stratum)
 
-        if self.stratification:
+        if isinstance(samp_size, dict):
             self.fpc = dict()
             strata = np.unique(stratum)
             for k, s in enumerate(strata):
@@ -123,16 +124,16 @@ class SampleSelection:
 
     def _grs_select(
         self,
-        probs: np.ndarray,
-        samp_unit: np.ndarray,
-        samp_size: Union[Dict[Any, int], int],
-        stratum: Optional[np.ndarray] = None,
-    ) -> Tuple[np.ndarray, np.ndarray]:
+        probs,
+        samp_unit,
+        samp_size,
+        stratum=None,
+    ):
 
         samp_size = formats.sample_size_dict(samp_size, self.stratification, stratum)
-        sample = hits = np.zeros(samp_unit.size).astype("int")
+        sample = hits = np.zeros(samp_unit.size).astype(int)
         self._calculate_fpc(samp_unit, samp_size, stratum)
-        if self.stratification:
[…]
