+++ src/samplics/estimation/expansion.py
@@ -229,13 +229,15 @@ class TaylorEstimator(_SurveyEstimator[Number, StringNumber]):
         self,
         parameter: str,
         alpha: float = 0.05,
-        random_seed: int = None,
-        ciprop_method: str = "logit",
+        random_seed: Optional[int] = None,
+        ciprop_method: Optional[str] = "logit",
     ) -> None:
         """Initializes the instance """
         _SurveyEstimator.__init__(self, parameter)
         if self.parameter == "proportion":
             self.ciprop_method = ciprop_method
+        else:
+            self.ciprop_method = None
 
     def _score_variable(
         self, y: np.ndarray, samp_weight: np.ndarray, x: np.ndarray = None
@@ -340,7 +342,7 @@ class TaylorEstimator(_SurveyEstimator[Number, StringNumber]):
             for s in np.unique(stratum):
                 y_score_s = y_score[stratum == s]
                 samp_weight_s = samp_weight[stratum == s]
-                psu_s = psu[stratum == s] if psu is not None else np.array([])
+                psu_s = psu[stratum == s] if psu is not None else None  # np.array([])
                 number_psus_in_s = np.size(np.unique(psu_s))
                 ssu_s = ssu[stratum == s] if ssu is not None else None
                 covariance += fpc[s] * self._variance_stratum_between(
@@ -442,86 +444,44 @@ class TaylorEstimator(_SurveyEstimator[Number, StringNumber]):
                     variance3[d] = float(cov_score_d)
                 return variance3, variance3
 
-    def estimate(
+    def _estimate(
         self,
         y: Array,
-        samp_weight: Optional[Array] = None,
-        x: Optional[Array] = None,
-        stratum: Optional[Series] = None,
-        psu: Optional[Series] = None,
-        ssu: Optional[Series] = None,
-        domain: Optional[Series] = None,
-        fpc: Union[Dict[StringNumber, Number], Series, Number] = 1.0,
-        deff: bool = False,
-        coef_variation: bool = False,
-        as_factor: bool = False,
-        remove_nan: bool = False,
+        samp_weight: np.ndarray,
+        x: Optional[np.ndarray],
+        stratum: Optional[np.ndarray],
+        psu: Optional[np.ndarray],
+        ssu: Optional[np.ndarray],
+        domain: Optional[np.ndarray],
+        fpc: Union[Dict[StringNumber, Number], Number],
+        deff: bool,
+        coef_variation: bool,
+        as_factor: bool,
+        remove_nan: bool,
     ) -> None:
 
-        if as_factor and self.parameter not in ("mean", "total"):
-            raise AssertionError("When as_factor is True, parameter must be mean or total!")
-
-        if self.parameter == "ratio" and x is None:
-            raise AssertionError("x must be provided for ratio estimation.")
-
-        y = numpy_array(y)
-        z = y.copy()
-
-        if samp_weight is None:
-            samp_weight = np.ones(z.shape[0])
-        elif isinstance(samp_weight, (float, int)):
-            samp_weight = samp_weight * np.ones(z.shape[0])
-        else:
-            samp_weight = np.asarray(samp_weight)
-
-        if remove_nan:
-            if self.parameter == "ratio":
-                excluded_units = np.isnan(y) | np.isnan(x)
-            else:
-                excluded_units = np.isnan(y)
-            z, samp_weight, x, stratum, domain, psu, ssu = remove_nans(
-                excluded_units, z, samp_weight, x, stratum, domain, psu, ssu
-            )
-
-        if stratum is not None:
-            stratum = numpy_array(stratum)
-            self.strata = np.unique(stratum).tolist()
-
-        if domain is not None:
[…]
