+++ src/samplics/categorical/comparison.py
@@ -5,7 +5,7 @@ The module implements comparisons of groups.
 """
 
 from __future__ import annotations
-from typing import Any, Generic, Dict, List, Optional, Union
+from typing import Any, Generic, Dict, List, Optional, Union, Tuple
 
 import math
 import numpy as np
@@ -21,7 +21,7 @@ from samplics.estimation import TaylorEstimator
 
 
 class Ttest(Generic[Number, StringNumber]):
-    def __init__(self, type: str, paired: Optional[bool] = None, alpha: float = 0.05) -> None:
+    def __init__(self, type: str, paired: bool = False, alpha: float = 0.05) -> None:
 
         if type.lower() not in ("one-sample", "two-sample"):
             raise ValueError("Parameter 'type' must be equal to 'one-sample', 'two-sample'!")
@@ -92,39 +92,27 @@ class Ttest(Generic[Number, StringNumber]):
         self.upper_ci = one_sample.upper_ci
         self.stddev = samp_std_dev
 
-    def _one_sample_two_groups(
-        self,
-        y: np.ndarray,
-        group: Array = None,
-        samp_weight: Array = None,
-        stratum: Optional[Array] = None,
-        psu: Optional[Array] = None,
-        ssu: Optional[Array] = None,
-        fpc: Union[Dict, float] = 1,
-    ) -> None:
+    def _two_groups_unpaired(
+        self, mean_est: TaylorEstimator, group: np.ndarray
+    ) -> Tuple[Any, Any, Any, Any, Any, Any]:
 
-        one_sample = TaylorEstimator(parameter="mean", alpha=self.alpha)
-        one_sample.estimate(
-            y=y,
-            domain=group,
-            samp_weight=samp_weight,
-            stratum=stratum,
-            psu=psu,
-            ssu=ssu,
-            fpc=fpc,
-        )
+        if self.type == "one-sample":
+            group1 = mean_est.domains[0]
+            group2 = mean_est.domains[1]
+        elif self.type == "two-sample":
+            group1 = mean_est.by[0]
+            group2 = mean_est.by[1]
+        else:
+            raise ValueError("Parameter type is not valid!")
 
-        group1 = one_sample.domains[0]
-        group2 = one_sample.domains[1]
-
-        mean_group1 = one_sample.point_est[group1]
-        mean_group2 = one_sample.point_est[group2]
+        mean_group1 = mean_est.point_est[group1]
+        mean_group2 = mean_est.point_est[group2]
 
         nb_obs_group1 = np.sum(group == group1)
         nb_obs_group2 = np.sum(group == group2)
 
-        stddev_group1 = math.sqrt(nb_obs_group1) * one_sample.stderror[group1]
-        stddev_group2 = math.sqrt(nb_obs_group2) * one_sample.stderror[group2]
+        stddev_group1 = math.sqrt(nb_obs_group1) * mean_est.stderror[group1]
+        stddev_group2 = math.sqrt(nb_obs_group2) * mean_est.stderror[group2]
 
         t_equal_variance = (mean_group1 - mean_group2) / (
             math.sqrt(
@@ -156,7 +144,7 @@ class Ttest(Generic[Number, StringNumber]):
         left_p_value_unequal_variance = t.cdf(t_unequal_variance, t_df_unequal_variance)
         both_p_value_unequal_variance = 2 * t.cdf(-abs(t_unequal_variance), t_df_unequal_variance)
 
-        self.stats = {
+        stats = {
             "number_obs": {group1: nb_obs_group1, group2: nb_obs_group2},
             "t_eq_variance": t_equal_variance,
             "df_eq_variance": t_df_equal_variance,
[…]
