+++ src/samplics/categorical/comparison.py
@@ -13,6 +13,7 @@ import pandas as pd
 
 from scipy.stats import t
 
+from samplics.utils.basic_functions import set_variables_names
 from samplics.utils.checks import assert_probabilities
 from samplics.utils.formats import numpy_array, remove_nans
 from samplics.utils.types import Array, Number, Series, StringNumber
@@ -44,6 +45,59 @@ class Ttest(Generic[Number, StringNumber]):
 
     def __repr__(self) -> str:
         return f"T-test(samp_type={self.samp_type}, paired={self.paired}, alpha={self.alpha})"
+
+    def __str__(self) -> str:
+
+        if self.point_est == {}:
+            return "No table to display"
+        else:
+            tbl_head = f"Design-based {self.samp_type.title()} T-test"
+            if (self.samp_type == "one-sample" and self.group_names == []) or self.paired:
+                if self.samp_type == "one-sample":
+                    tbl_subhead1 = f" Null hypothesis (Ho): mean = {self.stats['known_mean']}"
+                else:
+                    tbl_subhead1 = f" Null hypothesis (Ho): mean(Diff = {self.vars_names[0]} - {self.vars_names[1]}) = 0"
+                tbl_subhead2 = f" t statictics: {self.stats['t']:.4f}"
+                tbl_subhead3 = f" Degrees of freedom: {self.stats['df']:.2f}"
+                tbl_subhead4 = f" Alternative hypothesis (Ha):"
+                tbl_subhead4a = f"  Prob(T < t) = {self.stats['p_value']['less_than']:.4f}"
+                tbl_subhead4b = f"  Prob(|T| > |t|) = {self.stats['p_value']['not_equal']:.4f}"
+                tbl_subhead4c = f"  Prob(T > t) = {self.stats['p_value']['greater_than']:.4f}"
+
+                return f"\n{tbl_head}\n{tbl_subhead1}\n{tbl_subhead2}\n{tbl_subhead3}\n{tbl_subhead4}\n{tbl_subhead4a}\n{tbl_subhead4b}\n{tbl_subhead4c} \n\n{self.to_dataframe().to_string(index=False)}\n"
+
+            elif (self.samp_type == "one-sample" and self.group_names != []) or (
+                self.samp_type == "two-sample" and not self.paired
+            ):
+                tbl_subhead1 = f" Null hypothesis (Ho): mean({self.group_names[0]}) = mean({self.group_names[1]}) "
+                tbl_subhead2 = f" Equal variance assumption:"
+                tbl_subhead2a = f"  t statictics: {self.stats['t_eq_variance']:.4f}"
+                tbl_subhead2b = f"  Degrees of freedom: {self.stats['df_eq_variance']:.2f}"
+                tbl_subhead3 = f"  Alternative hypothesis (Ha):"
+                tbl_subhead3a = (
+                    f"   Prob(T < t) = {self.stats['p_value_eq_variance']['less_than']:.4f}"
+                )
+                tbl_subhead3b = (
+                    f"   Prob(|T| > |t|) = {self.stats['p_value_eq_variance']['not_equal']:.4f}"
+                )
+                tbl_subhead3c = (
+                    f"   Prob(T > t) = {self.stats['p_value_eq_variance']['greater_than']:.4f}"
+                )
+                tbl_subhead4 = f" Unequal variance assumption:"
+                tbl_subhead4a = f"  t statictics: {self.stats['t_uneq_variance']:.4f}"
+                tbl_subhead4b = f"  Degrees of freedom: {self.stats['df_uneq_variance']:.2f}"
+                tbl_subhead5 = f"  Alternative hypothesis (Ha):"
+                tbl_subhead5a = (
[…]
