+++ pyproject.toml
@@ -48,7 +48,7 @@ sphinx = "^3.0"
 sphinx-bootstrap-theme = "^0.7.1"
 nbsphinx = "^0.5.0"
 jupyterlab = "^1.2.4"
-# nb_black = "^1.0.7"
+nb_black = "^1.0.7"
 certifi = "^2019.11.28"
 pytest-cov = "^2.8.1"
 nox = "^2019.11.9"
