+++ tests/categorical/test_tabulation.py
@@ -6,7 +6,9 @@ from samplics.estimation import TaylorEstimator
 from samplics.categorical import Tabulation
 
 
-birthcat = pd.read_csv("./tests/categorical/birthcat.csv")
+birthcat = pd.read_csv(
+    "./tests/categorical/birthcat.csv", dtype={"agecat": "string", "birthcat": "category"}
+)
 
 region = birthcat["region"]
 age_cat = birthcat["agecat"]
@@ -27,27 +29,27 @@ tbl_count.tabulate(birth_cat, remove_nan=True)
 
 
 def test_oneway_count_one_var_count():
-    assert tbl_count.point_est["birthcat"][1] == 240
-    assert tbl_count.point_est["birthcat"][2] == 450
-    assert tbl_count.point_est["birthcat"][3] == 233
+    assert tbl_count.point_est["birthcat"]["1"] == 240
+    assert tbl_count.point_est["birthcat"]["2"] == 450
+    assert tbl_count.point_est["birthcat"]["3"] == 233
 
 
 def test_oneway_count_one_var_stderr():
-    assert np.isclose(tbl_count.stderror["birthcat"][1], 13.3337, atol=1e-4)
-    assert np.isclose(tbl_count.stderror["birthcat"][2], 15.1940, atol=1e-4)
-    assert np.isclose(tbl_count.stderror["birthcat"][3], 13.2050, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["birthcat"]["1"], 13.3337, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["birthcat"]["2"], 15.1940, atol=1e-4)
+    assert np.isclose(tbl_count.stderror["birthcat"]["3"], 13.2050, atol=1e-4)
 
 
 def test_oneway_count_one_var_lower_ci():
-    assert np.isclose(tbl_count.lower_ci["birthcat"][1], 213.8321, atol=1e-4)
-    assert np.isclose(tbl_count.lower_ci["birthcat"][2], 420.1812, atol=1e-4)
-    assert np.isclose(tbl_count.lower_ci["birthcat"][3], 207.0847, atol=1e-4)
+    assert np.isclose(tbl_count.lower_ci["birthcat"]["1"], 213.8321, atol=1e-4)
+    assert np.isclose(tbl_count.lower_ci["birthcat"]["2"], 420.1812, atol=1e-4)
+    assert np.isclose(tbl_count.lower_ci["birthcat"]["3"], 207.0847, atol=1e-4)
 
 
 def test_oneway_count_one_var_upper_ci():
-    assert np.isclose(tbl_count.upper_ci["birthcat"][1], 266.1679, atol=1e-4)
-    assert np.isclose(tbl_count.upper_ci["birthcat"][2], 479.8188, atol=1e-4)
-    assert np.isclose(tbl_count.upper_ci["birthcat"][3], 258.9153, atol=1e-4)
+    assert np.isclose(tbl_count.upper_ci["birthcat"]["1"], 266.1679, atol=1e-4)
+    assert np.isclose(tbl_count.upper_ci["birthcat"]["2"], 479.8188, atol=1e-4)
+    assert np.isclose(tbl_count.upper_ci["birthcat"]["3"], 258.9153, atol=1e-4)
 
 
 def test_oneway_count_one_var_deff_false():
@@ -58,36 +60,36 @@ def test_oneway_count_one_var_deff_false():
 
 
 # def test_oneway_count_one_var_deff_true():
-#     assert tbl_count.deff["birthcat"][1] == 1
[…]
