+++ .github/workflows/python-package.yml
@@ -27,8 +27,9 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8
+        python -m pip install pytest
         python -m pip install poetry
-        poetry install
+        poetry install --no-dev
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Lint with flake8
       run: |
