+++ .github/workflows/python-package.yml
@@ -27,9 +27,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install poetry
-        poetry install --no-dev
+        poetry install
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Test with pytest
       run: |
-        poetry add --dev pytest pytest-cov codecov
         pytest -v ./tests
