+++ .github/workflows/python-package.yml
@@ -27,7 +27,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install poetry
-        poetry install
+        pip install numpy pandas statsmoldels scipy
+        pip install pytest pytest-cov codecov
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Test with pytest
       run: |
