+++ .github/workflows/python-package.yml
@@ -26,8 +26,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install poetry
-        pip install numpy pandas statsmoldels scipy
+        pip install numpy pandas statsmodels scipy
         pip install pytest pytest-cov codecov
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Test with pytest
