+++ .github/workflows/tests.yml
@@ -10,12 +10,11 @@ on:
     branches: [ main ]
 
 jobs:
-  build:
-
+  samplics-test1:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.7, 3.8, 3.9]
+        python-version: [3.7, 3.8]
 
     steps:
     - uses: actions/checkout@v2
@@ -26,9 +25,9 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install numpy pandas statsmodels scipy
-        pip install pytest pytest-cov codecov
+        pip install poetry 
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Test with pytest
       run: |
-        pytest -v ./tests
+        poetry install
+        pytest -v tests/
