+++ .github/workflows/tests.yml
@@ -22,12 +22,14 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Install Python Poetry
+      uses: abatilo/actions-poetry@v2.1.0
+      with:
+        poetry-version: 1.1.0
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install poetry 
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        python -m poetry install
     - name: Test with pytest
       run: |
-        poetry install
-        pytest -v tests/
+        python -m poetry run python -m pytest -v tests
