+++ .github/workflows/tests.yml
@@ -15,7 +15,6 @@ jobs:
     strategy:
       matrix:
         python-version: [3.7, 3.8]
-
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python ${{ matrix.python-version }}
@@ -29,7 +28,34 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m poetry install
+        python -m poetry install --no-dev
+        python -m poetry add pytest
     - name: Test with pytest
       run: |
         python -m poetry run python -m pytest -v tests
+        
+  samplics-test2:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.9]
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install Python Poetry
+      uses: abatilo/actions-poetry@v2.1.0
+      with:
+        poetry-version: 1.1.0
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m poetry install --no-dev
+        python -m poetry add pytest
+        python -m poetry add codecov
+    - name: Running tests with coverage
+      run: |
+        python -m poetry run python -m pytest -v --cov-report xml --cov=src/samplics tests
+        python -m poetry run python -m codecov
