+++ .github/workflows/tests.yml
@@ -28,8 +28,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m poetry install --no-dev
-        python -m poetry add pytest
+        python -m poetry install 
     - name: Test with pytest
       run: |
         python -m poetry run python -m pytest -v tests
@@ -52,9 +51,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m poetry install --no-dev
-        python -m poetry add pytest
-        python -m poetry add codecov
+        python -m poetry install 
     - name: Running tests with coverage
       run: |
         python -m poetry run python -m pytest -v --cov-report xml --cov=src/samplics tests
