+++ .github/workflows/tests.yml
@@ -28,7 +28,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m poetry install 
+        python -m poetry add --lock pytest
+        python -m poetry install --no-dev
     - name: Test with pytest
       run: |
         python -m poetry run python -m pytest -v tests
@@ -51,7 +52,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m poetry install 
+        python -m poetry add --lock pytest pytest-cov codecov
+        python -m poetry install --no-dev
     - name: Running tests with coverage
       run: |
         python -m poetry run python -m pytest -v --cov-report xml --cov=src/samplics tests
