+++ .github/workflows/tests.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Python package
+name: Testing Samplics 
 
 on:
   push:
@@ -10,7 +10,7 @@ on:
     branches: [ main ]
 
 jobs:
-  samplics-test1:
+  Testing-previous-py-versions:
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -34,7 +34,7 @@ jobs:
       run: |
         python -m poetry run python -m pytest -v tests
         
-  samplics-test2:
+  Testing-cuurent-py-version-with-coverage:
     runs-on: ubuntu-latest
     strategy:
       matrix:
