### 2021-01-18 8071361
dropping python 3.6 support
