### 2021-01-18 8206945
dropping python 3.6 support
