+++ pyproject.toml
@@ -15,7 +15,6 @@ keywords = ["sampling", "sample", "weighting", "estimation", "survey",]
 
 classifiers = [
     "Programming Language :: Python :: Implementation :: CPython",
-    "Programming Language :: Python :: 3.6",
     "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
@@ -29,7 +28,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.6.1" #Pandas 1.0 requires at least Python 3.6.1
+python = "^3.7" #Pandas 1.0 requires at least Python 3.6.1
 numpy = "^1.15"
 pandas = "^1.0"
 scipy = "^1.1"
@@ -37,20 +36,20 @@ statsmodels = "^0.10"
 matplotlib = "^3.1"
 
 [tool.poetry.dev-dependencies]
-mypy = "^0.770"
-pytest = "^5.0.1"
-flake8 = "^3.7.9"
+mypy = "^0.790"
+pytest = "^6.2.1"
+flake8 = "^3.8.4"
 black = "^20.8b1"
 ipython = "^7.11.0"
 ipykernel = "^5.1.3"
 sphinx = "^3.0"
 sphinx-bootstrap-theme = "^0.7.1"
 nbsphinx = "^0.5.0"
-jupyterlab = "^1.2.4"
+jupyterlab = "^3.0"
 nb_black = "^1.0.7"
 certifi = "^2019.11.28"
 pytest-cov = "^2.8.1"
-nox = "^2019.11.9"
+nox = "^2020.12.31"
 isort = "^4.3.21"
 recommonmark = "^0.6"
 codecov = "^2.1.3"
