### 2021-01-18 235f393
dropping python 3.6 support
