+++ pytest.ini
@@ -1,5 +1,0 @@
-[pytest]
-
-python_files = tests.py test_*.py *_test.python_files
-
-addopts = -v --cov=./src/samplics --cov-report=html
