### 2021-01-18 cd572ef
dropping python 3.6 support
