### 2021-01-18 4f122d9
dropping python 3.6 support
