+++ docs/source/index.rst
@@ -1,7 +1,7 @@
 ==========================
 SAMPLICS: sample analytics
 ==========================
-.. image:: https://travis-ci.com/survey-methods/samplics.svg?token=WwRayqkQBt1W4ihyTzvw&branch=master
+.. image:: ![CI](https://github.com/survey-methods/samplics/tree/main/.github/workflows/Testing-Samplics/badge.svg?branch=master)
   :target: https://travis-ci.com/survey-methods/samplics
 
 .. image:: https://codecov.io/gh/survey-methods/samplics/branch/master/graph/badge.svg?token=7C0LBB5N8Y
