### 2021-01-18 373b924
dropping python 3.6 support
