### 2021-01-18 e82af80
dropping python 3.6 support
