+++ .github/workflows/tests_coverage.yml
@@ -1,0 +1,19 @@
+name: Test-Coverage 
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  name: Upload coverage to codecov
+  uses: codecov/codecov-action@v1
+    with:
+      token: ${{ secrets.CODECOV_TOKEN }}
+      codecov:
+        require_ci_to_pass: yes
+      coverage:
+        precision: 0
+        round: down
+        range: "70...100"
