+++ codecov.yml
@@ -1,7 +1,0 @@
-codecov:
-  require_ci_to_pass: yes
-
-coverage:
-  precision: 0
-  round: down
-  range: "70...100"
