### 2021-01-18 8b2e582
dropping python 3.6 support
