### 2021-01-18 46735c6
dropping python 3.6 support
