### 2021-01-18 d86165c
dropping python 3.6 support
