+++ .github/workflows/tests_coverage.yml
@@ -7,15 +7,55 @@ on:
     branches: [main]
 
 jobs:
-  test-coverage:
+
+  Testing-current-py-version-with-coverage:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.9]
     steps:
-      - uses: actions/checkout@master
-      - uses: codecov/codecov-action@v1
-        with:
-          token: ${{ secrets.CODECOV_TOKEN }} # not required for public repos
-          files: ./coverage1.xml,./coverage2.xml # optional
-          flags: unittests # optional
-          name: codecov-umbrella # optional
-          fail_ci_if_error: true # optional (default = false)
-          verbose: true # optional (default = false)
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install Python Poetry
+      uses: abatilo/actions-poetry@v2.1.0
+      with:
+        poetry-version: 1.1.0
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m poetry add --lock pytest pytest-cov codecov
+        python -m poetry install --no-dev
+    - name: Running tests with coverage
+      run: |
+        python -m poetry run python -m pytest -v --cov-report xml --cov=src/samplics tests
+        python -m poetry run python -m codecov
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v1
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        # files: ./coverage1.xml,./coverage2.xml
+        # directory: ./coverage/reports/
+        # flags: unittests
+        env_vars: OS,PYTHON
+        name: codecov-umbrella
+        fail_ci_if_error: true
+        path_to_write_report: ./coverage/codecov_report.txt
+        verbose: true
+
+
+
+  # test-coverage:
+  #   runs-on: ubuntu-latest
+  #   steps:
+  #     - uses: actions/checkout@master
+  #     - uses: codecov/codecov-action@v1
+  #       with:
+  #         token: ${{ secrets.CODECOV_TOKEN }} # not required for public repos
+  #         files: ./coverage1.xml,./coverage2.xml # optional
+  #         flags: unittests # optional
+  #         name: codecov-umbrella # optional
+  #         fail_ci_if_error: true # optional (default = false)
+  #         verbose: true # optional (default = false)
