### 2021-01-18 77fe8ca
dropping python 3.6 support
