### 2021-01-18 6b5cdf9
dropping python 3.6 support
