+++ .github/workflows/tests.yml
@@ -1,13 +1,13 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Testing-Samplics 
+name: Testing-Samplics
 
 on:
   push:
-    branches: [ main ]
+    branches: [main]
   pull_request:
-    branches: [ main ]
+    branches: [main]
 
 jobs:
   Testing-previous-py-versions:
@@ -16,45 +16,45 @@ jobs:
       matrix:
         python-version: [3.7, 3.8]
     steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install Python Poetry
-      uses: abatilo/actions-poetry@v2.1.0
-      with:
-        poetry-version: 1.1.0
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m poetry add --lock pytest
-        python -m poetry install --no-dev
-    - name: Test with pytest
-      run: |
-        python -m poetry run python -m pytest -v tests
-        
-  Testing-current-py-version-with-coverage:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.9]
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install Python Poetry
-      uses: abatilo/actions-poetry@v2.1.0
-      with:
-        poetry-version: 1.1.0
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m poetry add --lock pytest pytest-cov codecov
-        python -m poetry install --no-dev
-    - name: Running tests with coverage
-      run: |
-        python -m poetry run python -m pytest -v --cov-report xml --cov=src/samplics tests
-        python -m poetry run python -m codecov
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install Python Poetry
+        uses: abatilo/actions-poetry@v2.1.0
+        with:
+          poetry-version: 1.1.0
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m poetry add --lock pytest
+          python -m poetry install --no-dev
+      - name: Test with pytest
+        run: |
+          python -m poetry run python -m pytest -v tests
+
+  # Testing-current-py-version-with-coverage:
+  #   runs-on: ubuntu-latest
+  #   strategy:
+  #     matrix:
+  #       python-version: [3.9]
+  #   steps:
+  #   - uses: actions/checkout@v2
+  #   - name: Set up Python ${{ matrix.python-version }}
+  #     uses: actions/setup-python@v2
+  #     with:
+  #       python-version: ${{ matrix.python-version }}
+  #   - name: Install Python Poetry
+  #     uses: abatilo/actions-poetry@v2.1.0
+  #     with:
+  #       poetry-version: 1.1.0
+  #   - name: Install dependencies
+  #     run: |
+  #       python -m pip install --upgrade pip
+  #       python -m poetry add --lock pytest pytest-cov codecov
+  #       python -m poetry install --no-dev
+  #   - name: Running tests with coverage
+  #     run: |
+  #       python -m poetry run python -m pytest -v --cov-report xml --cov=src/samplics tests
+  #       python -m poetry run python -m codecov
