+++ .travis.yml
@@ -1,41 +1,0 @@
-pylanguage: python
-
-before_install:
-  - pip install poetry
-  - pip install nox
-
-jobs:
-  include:
-    - python: "3.6"
-      env: NOXSESSION="tests-3.6"
-    - python: "3.7"
-      env: NOXSESSION="tests-3.7"
-    - python: "3.8"
-      env: NOXSESSION="tests-3.8"
-    - python: "3.9"
-      env:
-        - NOXSESSION="test_coverage" #test and coverage
-        - CODECOV_TOKEN: ${{secrets.CODECOV_TOKEN}}
-    # - python: "3.8"
-    #   env: NOXSESSION="lint"
-    # - python: "3.8"
-    #   env: NOXSESSION="docs"
-script: nox --non-interactive --session "$NOXSESSION"
-
-# before-deploy:
-#   - poetry install
-#   - rm -rf dist build .egg-info
-#   - poetry build
-
-# deploy:
-#   provider: script
-#   script: poetry publish
-#   skip_cleanup: true
-#   user: __token__
-#   password:
-#     secure: vnqgRL/RhwPjjWS9830bHYueCtWI2kZwqiIUxapZ24j4fsXKXTfLeubPGGcn/xCRK8ZPQya6kYC+9Yx/qsanE6smefoWz87m5hfntKVSTf2X1V7NbGThe2GjkumcQlSdxNVGhxS7RHf0d1EPAD6ZL9uTywEwWjRfHBFKbYr/ybwCNyjZXCB5OvVaWa+BH0/RHB+uLAZzBaheSdF6NukuqgR1LTY74HQ+nywwFhZX4moahdHTHD8tv24mN40N4rdFdccca0k0976o445ZUvksnvbRG44k6RUCWFCza0Y8rOu3RQZIPbOWZB7VbDOLNI6I03vLnJDb4ajpoDTpY3k59ub757DhF56ugqvV+PV88U9KmIiasqRIWjaKirScp81LAXFj7TCOMMuhCkNI767je1MKxRp5izT7fHhhTKKVwgrnr6KDnMhGhabXWLS/uW2VVHlZqbrFFBq2IlS7x0skEpDuDFwI6+KzlrqWAP0iuqgLT1UWPtHJvRgSc+cQcKj2Ek+3mV7A5B1qS4iwHHsiePP1xZWLwzmqRTTvq6eLaprmxyrKj+zs85LmdCKaq+WPlg4QiHQiDh7YCFXlA0HNoGMvCcj/fpM/qiyZKTYMVvJ9dP7MTAu1l4f1QNJduBCA5is9g0jxiFBYzVoNey0BG+OYaGObIVcD55P45hNEYzY=
-#   on:
-#   branch: master
-
-notifications:
-  email: msdiallo@samplics.org
