+++ pyproject.toml
@@ -42,17 +42,17 @@ pytest-cov = "^2.8.1"
 codecov = "^2.1.3"
 black = "^20.8b1"
 flake8 = "^3.8.4"
-# ipython = "^7.11.0"
-# ipykernel = "^5.1.3"
 sphinx = "^3.0"
 sphinx-bootstrap-theme = "^0.7.1"
 nbsphinx = "^0.5.0"
 recommonmark = "^0.6"
+# ipython = "^7.11.0"
+# ipykernel = "^5.1.3"
 # jupyterlab = "^3.0"
 # nb_black = "^1.0.7"
 # certifi = "^2019.11.28"
 # nox = "^2020.12.31"
-isort = "^4.3.21"
+isort = "^5.7"
 
 [tool.black]
 line-length = 99
