+++ .github/workflows/tests_coverage.yml
@@ -25,8 +25,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m poetry add --lock pytest pytest-cov codecov
-          python -m poetry install --no-dev
+          python -m poetry install
       - name: Running tests with coverage
         run: |
           python -m poetry run python -m pytest -v --cov-report xml --cov=src/samplics tests
