+++ .github/workflows/tests_coverage.yml
@@ -1,4 +1,4 @@
-name: Testing-Coverage
+name: Coverage
 
 on:
   push:
