+++ .github/workflows/tests.yml
@@ -10,11 +10,11 @@ on:
     branches: [main]
 
 jobs:
-  Testing-previous-py-versions:
+  Testing-supported-py-versions:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.7, 3.8]
+        python-version: [3.7, 3.8, 3.9]
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python ${{ matrix.python-version }}
@@ -33,28 +33,3 @@ jobs:
       - name: Test with pytest
         run: |
           python -m poetry run python -m pytest -v tests
-
-  # Testing-current-py-version-with-coverage:
-  #   runs-on: ubuntu-latest
-  #   strategy:
-  #     matrix:
-  #       python-version: [3.9]
-  #   steps:
-  #   - uses: actions/checkout@v2
-  #   - name: Set up Python ${{ matrix.python-version }}
-  #     uses: actions/setup-python@v2
-  #     with:
-  #       python-version: ${{ matrix.python-version }}
-  #   - name: Install Python Poetry
-  #     uses: abatilo/actions-poetry@v2.1.0
-  #     with:
-  #       poetry-version: 1.1.0
-  #   - name: Install dependencies
-  #     run: |
-  #       python -m pip install --upgrade pip
-  #       python -m poetry add --lock pytest pytest-cov codecov
-  #       python -m poetry install --no-dev
-  #   - name: Running tests with coverage
-  #     run: |
-  #       python -m poetry run python -m pytest -v --cov-report xml --cov=src/samplics tests
-  #       python -m poetry run python -m codecov
