+++ joss/paper.md
@@ -1,16 +1,16 @@
 ---
 title: "samplics: a Python Package for selecting, weighting and analyzing data from complex sampling designs."
 tags:
-	- Python
-	- Survey sampling
-	- small area estimation
-	- Official statistics
+        - Python
+        - Survey sampling
+        - small area estimation
+        - Official statistics
 authors:
-	- name: Mamadou S. Diallo^[Corresponding author.]
+        - name: Mamadou S. Diallo^[Corresponding author.]
 	  orcid: 0000-0002-0376-3631
 	  affiliation: 1
 affiliations:
-	- name: UNICEF, The United Nations Children's Fund
+        - name: UNICEF, The United Nations Children's Fund
 	  index: 1
 date: 12 December 2020
 bibliography: paper.bib
