### 2021-01-19 #2908 @MamadouSDiallo
Thank you @kthyng I have removed the tabs.
