### 2021-01-19 #2908 @kthyng
@MamadouSDiallo you can always test your paper compilation by running the command yourself.
