### 2021-02-08 #2908 @arfon
@MamadouSDiallo - apologies it's taken so long to get this paper moving - all of our editors are extremely busy right now I'm afraid so I'll take this on myself.

Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
