### 2021-02-18 #2908 @lorenanicole
:wave: @stulacy would you be willing to review this submission for JOSS?
