+++ pyproject.toml
@@ -36,7 +36,7 @@ statsmodels = "^0.10"
 matplotlib = "^3.1"
 
 [tool.poetry.dev-dependencies]
-mypy = "^0.790"
+mypy = "^0.812"
 pytest = "^6.2.1"
 pytest-cov = "^2.8.1"
 codecov = "^2.1.3"
@@ -53,6 +53,7 @@ recommonmark = "^0.6"
 # certifi = "^2019.11.28"
 # nox = "^2020.12.31"
 isort = "^5.7"
+pylint = "^2.6.0"
 
 [tool.black]
 line-length = 99
