### 2021-02-20 233eb9e
fixed types using mypy
