+++ src/samplics/utils/types.py
@@ -10,7 +10,7 @@ import pandas as pd
 Array = Union[np.ndarray, pd.Series, list, tuple]
 Series = Union[pd.Series, list, tuple]
 
-Number = TypeVar("Number", int, float)  # Union[float, int]
-StringNumber = TypeVar("StringNumber", str, float, int)  # Union[str, float, int]
+Number = Union[float, int]
+StringNumber =  Union[str, float, int]
 
 DictStrNum = Dict[StringNumber, Number]
