+++ mypy.ini
@@ -3,7 +3,7 @@ python_version = 3.8
 
 ignore_missing_imports = True
 follow_imports = silent
-strict_optional = False
+strict_optional = True
 allow_redefinition = True
 show_error_context = False
 show_column_numbers = True
