### 2021-02-20 8b4f2a4
fixed types using mypy
