+++ src/samplics/sampling/size.py
@@ -2,7 +2,9 @@
 
 """
 
-from typing import Any, Dict, List, Union, Optional, Generic, overload
+from __future__ import annotations
+
+from typing import Any, Union, Optional
 
 import math
 
@@ -12,10 +14,10 @@ import pandas as pd
 from scipy.stats import norm as normal
 
 from samplics.utils import formats
-from samplics.utils.types import Array, Number, StringNumber
+from samplics.utils.types import Array, Number, StringNumber, DictStrNum
 
 
-class SampleSize(Generic[Number, StringNumber]):
+class SampleSize:
     """*SampleSize* implements sample size calculation methods"""
 
     def __init__(
@@ -27,28 +29,28 @@ class SampleSize(Generic[Number, StringNumber]):
         if self.method not in ("wald", "fleiss"):
             raise AssertionError("Sample size calculation method not valid.")
         self.stratification = stratification
-        # self.target: Union[Number, Dict[StringNumber, Number]] = 0
-        # self.samp_size: Union[int, Dict[StringNumber, int]] = 0
-        # self.deff_c: Union[Number, Dict[StringNumber, Number]] = 1
-        # self.deff_w: Union[Number, Dict[StringNumber, Number]] = 1
-        # self.precision: Union[Number, Dict[StringNumber, Number]] = 0
-        # self.resp_rate: Union[Number, Dict[StringNumber, Number]] = 1
+        self.target: Union[DictStrNum, Number]
+        self.samp_size: Union[DictStrNum, Number] = 0
+        self.deff_c: Union[DictStrNum, Number] = 1.0
+        self.deff_w: Union[DictStrNum, Number] = 1.0
+        self.precision: Union[DictStrNum, Number] 
+        self.resp_rate: Union[DictStrNum, Number] = 1.0
 
-    def icc(self) -> Union[Dict[StringNumber, Number], Number]:
+    def icc(self) -> Union[DictStrNum, Number]:
         pass
 
     def deff(
         self,
-        cluster_size: Union[Dict[StringNumber, Number], Number],
-        icc: Union[Dict[StringNumber, Number], Number],
-    ) -> Union[Dict[StringNumber, Number], Number]:
+        cluster_size: Union[DictStrNum, Number],
+        icc: Union[DictStrNum, Number],
+    ) -> Union[DictStrNum, Number]:
 
         if isinstance(cluster_size, (int, float)) and isinstance(icc, (int, float)):
             return max(1 + (cluster_size - 1) * icc, 0)
         elif isinstance(cluster_size, dict) and isinstance(icc, dict):
             if cluster_size.keys() != icc.keys():
                 raise AssertionError("Parameters do not have the same dictionary keys.")
-            deff_c: Dict[StringNumber, Number] = {}
+            deff_c: DictStrNum = {}
             for s in cluster_size:
                 deff_c[s] = max(1 + (cluster_size[s] - 1) * icc[s], 0)
             return deff_c
@@ -57,10 +59,10 @@ class SampleSize(Generic[Number, StringNumber]):
 
     def _calculate_wald(
         self,
-        target: Union[Dict[StringNumber, Number], Number],
-        precision: Union[Dict[StringNumber, Number], Number],
+        target: Union[DictStrNum, Number],
+        precision: Union[DictStrNum, Number],
         stratum: Optional[Array],
-    ) -> Union[int, Dict[StringNumber, int]]:
+    ) -> Union[DictStrNum, Number]:
 
         z_value = normal().ppf(1 - self.alpha / 2)
 
@@ -68,8 +70,9 @@ class SampleSize(Generic[Number, StringNumber]):
             isinstance(target, dict)
             and isinstance(precision, dict)
[…]
