### 2021-02-20 95e4c03
fixed types using mypy
