### 2021-02-20 65a2dad
fixed types using mypy and minor version number revision
