### 2021-02-20 47dbb14
fixed types using mypy and minor version number revision
