+++ src/samplics/sae/robust_unit_model.py
@@ -40,65 +40,65 @@ from samplics.sae.eb_unit_model import EbUnitModel
 class EllUnitModel:
     """*EllUnitModel* implements the basic Unit level model for complex indicators.
 
-    *EllUnitModel* takes the sample data as input and fits the basic linear mixed model. 
-    The user can pick between restricted maximum likelihood (REML), maximum likelihood (ML), 
-    and method of moments (MOM) to fit the model parameters. Also, EllUnitModel predicts the 
-    areas means and provides the point and mean squared error (MSE) estimates of the empirical 
-    Bayes linear unbiased (EBLUP). 
+    *EllUnitModel* takes the sample data as input and fits the basic linear mixed model.
+    The user can pick between restricted maximum likelihood (REML), maximum likelihood (ML),
+    and method of moments (MOM) to fit the model parameters. Also, EllUnitModel predicts the
+    areas means and provides the point and mean squared error (MSE) estimates of the empirical
+    Bayes linear unbiased (EBLUP).
 
-    *EllUnitModel* requires the user to provide the indicator function. The indicator function is 
-    expected to take the array of output sample observations as input and possibly some additional 
+    *EllUnitModel* requires the user to provide the indicator function. The indicator function is
+    expected to take the array of output sample observations as input and possibly some additional
     parameters needed to compute the indicator. The indicator function outputs an aggregated value.
-    For example, the poverty gap indicator can have the following signature 
-    pov_gap(y: array, pov_line: float) -> float. If the indicator function different outputs by 
+    For example, the poverty gap indicator can have the following signature
+    pov_gap(y: array, pov_line: float) -> float. If the indicator function different outputs by
     area then the self.area_list can be used to incorporate different logics across areas.
 
-    Also, EllUnitModel can use Boxcox to transform the output sample values in order to reduce the 
-    asymmetry in the datawhen fitting the linear mixed model. 
+    Also, EllUnitModel can use Boxcox to transform the output sample values in order to reduce the
+    asymmetry in the datawhen fitting the linear mixed model.
 
     Setting attributes
-        | method (str): the fitting method of the model parameters which can take the possible 
-        |   values restricted maximum likelihood (REML),  maximum likelihood (ML), and method of 
-        |   moments (MOM). If not specified, "MOM" is used as default.  
-        | indicator (function): a user defined function to compute the indicator. 
-        | boxcox (dict): contains the *lambda* parameter of the Boxcox and a constant for the 
-        |   log-transformation of the Boxcox.  
+        | method (str): the fitting method of the model parameters which can take the possible
+        |   values restricted maximum likelihood (REML),  maximum likelihood (ML), and method of
+        |   moments (MOM). If not specified, "MOM" is used as default.
+        | indicator (function): a user defined function to compute the indicator.
+        | boxcox (dict): contains the *lambda* parameter of the Boxcox and a constant for the
+        |   log-transformation of the Boxcox.
 
     Sample related attributes
-        | ys (array): the output sample observations. 
-        | Xs (ndarray): the auxiliary information. 
+        | ys (array): the output sample observations.
+        | Xs (ndarray): the auxiliary information.
         | scales (array): an array of scaling parameters for the unit levels errors.
         | afactors (array): sum of the inverse squared of scale.
-        | areas (array): the full vector of small areas from the sampled observations.  
+        | areas (array): the full vector of small areas from the sampled observations.
         | areas_list (array): the list of small areas from the sample data.
-        | samp_size (dict): the sample size per small areas from the sample. 
-        | ys_mean (array): sample area means of the output variable. 
+        | samp_size (dict): the sample size per small areas from the sample.
+        | ys_mean (array): sample area means of the output variable.
         | Xs_mean (ndarray): sample area means of the auxiliary variables.
 
[…]
