+++ src/samplics/sae/eblup_area_model.py
@@ -24,50 +24,50 @@ from samplics.utils.types import Array, Number
 
 
 class EblupAreaModel:
-    """*EblupAreaModel* implements the basic area level model. 
+    """*EblupAreaModel* implements the basic area level model.
 
-    *EblupAreaModel* takes the unstable survey estimates as input and model them to produce the 
-    small area estimates. The standard error of the unit level residuals are assumed known and 
-    usually correspond to the survey estimates standard error. In some cases, the survey standard 
+    *EblupAreaModel* takes the unstable survey estimates as input and model them to produce the
+    small area estimates. The standard error of the unit level residuals are assumed known and
+    usually correspond to the survey estimates standard error. In some cases, the survey standard
     errors (or variances) are modelled to reduce their variability and the modelled standard errors
     are them used for the computes the small area estimates. The user can pick between REML, ML and
-    FH to estimate the model parameters. EblupAreaModel provides both the area level point estimates and the taylor approximation of the MSEs. 
+    FH to estimate the model parameters. EblupAreaModel provides both the area level point estimates and the taylor approximation of the MSEs.
 
     Setting attributes
-        | method (str): the fitting method of the model parameters which can take the possible 
-        |   values restricted maximum likelihood (REML), maximum likelihood (ML), and 
+        | method (str): the fitting method of the model parameters which can take the possible
+        |   values restricted maximum likelihood (REML), maximum likelihood (ML), and
         |   Fay-Herriot (FH). If not specified, "REML" is used as default.
 
     Sample related attributes
-        | yhat (array): the survey output area level estimates. This is also referred to as the 
-        | direct estimates. 
-        | error_std (number): the estimated standard error of the direct estimates. 
-        | X (ndarray): the auxiliary information. 
+        | yhat (array): the survey output area level estimates. This is also referred to as the
+        | direct estimates.
+        | error_std (number): the estimated standard error of the direct estimates.
+        | X (ndarray): the auxiliary information.
         | b_const (array): an array of scaling parameters for the area random effects.
-        | area (array): the area of the survey output estimates.  
-        | samp_size (dict): the sample size per small areas from the sample. 
-        | ys_mean (array): sample area means of the output variable. 
+        | area (array): the area of the survey output estimates.
+        | samp_size (dict): the sample size per small areas from the sample.
+        | ys_mean (array): sample area means of the output variable.
         | Xs_mean (ndarray): sample area means of the auxiliary variables.
 
     Model fitting attributes
-        | fitted (boolean): indicates whether the model has been fitted or not. 
-        | fixed_effects (array): the estimated fixed effects of the regression model. 
-        | fe_std (array): the estimated standard errors of the fixed effects. 
+        | fitted (boolean): indicates whether the model has been fitted or not.
+        | fixed_effects (array): the estimated fixed effects of the regression model.
+        | fe_std (array): the estimated standard errors of the fixed effects.
         | random_effects (array): the estimated standard errors of the fixed effects.
         | re_std (number): the estimated area level standard error of the random effects.
-        | convergence (dict): a dictionnary holding the convergence status and the number of 
-        |   iterations from the model fitting algorithm. 
+        | convergence (dict): a dictionnary holding the convergence status and the number of
+        |   iterations from the model fitting algorithm.
         | goodness (dict): a dictionary holding the log-likelihood, AIC, and BIC.
 
     Prediction related attributes
[…]
