+++ tests/sampling/test_selection.py
@@ -59,10 +59,28 @@ def test_srswor_select():
 """Stratified simple random sampling"""
 continent = countries_population["continent"]
 sample_sizes = dict(
-    {"Africa": 10, "Asia": 7, "Europe": 5, "North America": 3, "South America": 3, "Oceanie": 1,}
+    {
+        "Africa": 10,
+        "Asia": 7,
+        "Europe": 5,
+        "North America": 3,
+        "South America": 3,
+        "Oceanie": 1,
+    }
 )
 
 str_srswr_design = SampleSelection(method="srs", stratification=True)
+
+
+def test_stratified_srswr_probs_same_size():
+    size = 2
+    str_srswr_probs = str_srswr_design.inclusion_probs(countries, size, continent)
+    strata = np.unique(continent)
+    initial_probs = obtained_probs = dict()
+    for s in strata:
+        initial_probs[s] = size / np.sum(continent == s)
+        obtained_probs[s] = np.unique(str_srswr_probs[continent == s])
+    assert initial_probs == obtained_probs
 
 
 def test_stratified_srswr_probs():
@@ -73,6 +91,17 @@ def test_stratified_srswr_probs():
         initial_probs[s] = sample_sizes[s] / np.sum(continent == s)
         obtained_probs[s] = np.unique(str_srswr_probs[continent == s])
     assert initial_probs == obtained_probs
+
+
+def test_stratified_srswr_select_same_size():
+    size = 2
+    str_srswr_sample, str_srswr_number_hits, _ = str_srswr_design.select(
+        countries, size, continent
+    )
+    strata = np.unique(continent)
+    obtained_sample_sizes = dict()
+    for s in strata:
+        assert size == np.sum(str_srswr_number_hits[continent == s])
 
 
 def test_stratified_srswr_select():
@@ -99,6 +128,18 @@ def test_stratified_srswor_probs():
     assert initial_probs == obtained_probs
 
 
+def test_stratified_srswor_select_same_size():
+    size = 2
+    str_srswor_sample, str_srswor_number_hits, _ = str_srswor_design.select(
+        countries, size, continent
+    )
+    strata = np.unique(continent)
+    obtained_sample_sizes = dict()
+    for s in strata:
+        assert size == np.sum(str_srswor_number_hits[continent == s])
+    assert (np.unique(str_srswor_number_hits) == (0, 1)).all()
+
+
 def test_stratified_srswor_select():
     str_srswor_sample, str_srswor_number_hits, _ = str_srswor_design.select(
         countries, sample_sizes, continent
@@ -111,6 +152,18 @@ def test_stratified_srswor_select():
     assert sample_sizes == obtained_sample_sizes
 
 
+def test_stratified_srswor_select():
+    size = 2
+    str_srswor_sample, str_srswor_number_hits, _ = str_srswor_design.select(
+        countries, size, continent
+    )
+    strata = np.unique(continent)
+    obtained_sample_sizes = dict()
+    for s in strata:
+        assert size == np.sum(str_srswor_number_hits[continent == s])
+    assert (np.unique(str_srswor_number_hits) == (0, 1)).all()
+
+
 """PPS sampling"""
 mos = countries_population["population_2019"].to_numpy()
 
@@ -120,6 +173,14 @@ pps_sys_design_wr = SampleSelection(method="pps-sys")
 def test_ppswr_sys_probs():
     pps_probs = pps_sys_design_wr.inclusion_probs(countries, sample_size, mos=mos)
[…]
