### 2021-02-20 4ab98ce
fixed types for selection.py using mypy
