+++ src/samplics/sampling/selection.py
@@ -24,7 +24,7 @@ import pandas as pd
 import math
 
 from samplics.utils import formats
-from samplics.utils.types import Array, Number, StringNumber
+from samplics.utils.types import Array, DictStrFloat, DictStrInt, DictStrNum, Number, StringNumber
 
 
 class SampleSelection:
@@ -54,10 +54,10 @@ class SampleSelection:
 
     def __init__(
         self,
-        method,
-        stratification=False,
-        with_replacement=True,
-    ):
+        method: str,
+        stratification: bool = False,
+        with_replacement: bool = True,
+    ) -> None:
         if method.lower() in (
             "srs",
             "sys",
@@ -73,18 +73,18 @@ class SampleSelection:
             raise ValueError(
                 "method must be in ('srs', 'sys', 'pps-brewer', 'pps-hv', 'pps-murphy', 'pps-rs, 'pps-sys', 'grs')"
             )
-        self.stratification = True if stratification else False
-        self.with_replacement = with_replacement
-        self.fpc = {}
+        self.stratification: bool = True if stratification else False
+        self.with_replacement: bool = with_replacement
+        self.fpc: Union[DictStrNum, Number] = {}
 
     @staticmethod
     def _to_dataframe(
-        samp_unit,
-        stratum,
-        mos,
-        sample,
-        hits,
-        probs,
+        samp_unit: np.ndarray,
+        stratum: Optional[np.ndarray],
+        mos: Optional[np.ndarray],
+        sample: np.ndarray,
+        hits: np.ndarray,
+        probs: Optional[np.ndarray],
     ) -> pd.DataFrame:
 
         df = pd.DataFrame(
@@ -108,10 +108,10 @@ class SampleSelection:
 
     def _calculate_fpc(
         self,
-        samp_unit,
-        samp_size,
-        stratum,
-    ):
+        samp_unit: np.ndarray,
+        samp_size: Union[DictStrInt, int],
+        stratum: Optional[np.ndarray],
+    ) -> None:
 
         samp_unit = formats.sample_units(samp_unit, unique=True)
         samp_size = formats.sample_size_dict(samp_size, self.stratification, stratum)
@@ -127,11 +127,11 @@ class SampleSelection:
 
     def _grs_select(
         self,
-        probs,
-        samp_unit,
-        samp_size,
-        stratum=None,
-    ):
+        probs: np.ndarray,
+        samp_unit: np.ndarray,
+        samp_size: Union[DictStrInt, int],
+        stratum: Optional[np.ndarray] = None,
+    ) -> tuple[np.ndarray, np.ndarray]:
 
         samp_size = formats.sample_size_dict(samp_size, self.stratification, stratum)
         sample = hits = np.zeros(samp_unit.size).astype(int)
@@ -168,59 +168,61 @@ class SampleSelection:
 
     @staticmethod
     def _anycertainty(
-        samp_size,
-        stratum,
-        mos,
+        samp_size: Union[DictStrInt, int],
+        stratum: Optional[np.ndarray],
+        mos: np.ndarray,
     ) -> bool:
 
-        if stratum is not None:
+        if stratum is not None and isinstance(samp_size, dict):
             probs = np.zeros(stratum.size)
             for s in np.unique(stratum):
                 stratum_units = stratum == s
                 mos_s = mos[stratum_units]
                 probs[stratum_units] = samp_size[s] * mos_s / np.sum(mos_s)
-        else:
+        elif isinstance(samp_size, (int, float)):
             probs = samp_size * mos / np.sum(mos)
 
-        certainty = (probs >= 1).any()
-
-        return certainty
+        return bool((probs >= 1).any())  # to make mypy happy
 
     # SRS methods
     def _srs_inclusion_probs(
         self,
[…]
