### 2021-02-20 5e0e293
fixed types for selection.py using mypy
