### 2021-02-21 684bfae
fixed types using mypy
