+++ tests/weighting/test_adjustment.py
@@ -60,8 +60,8 @@ def test_uk_adjustment_without_adj_class():
 def test_deff_wgt_without_domain():
     mean_wgt = np.mean(nr_wgt_without_adj_class)
     deff_wgt = 1 + np.mean(np.power(nr_wgt_without_adj_class - mean_wgt, 2) / mean_wgt ** 2)
-    assert sample_wgt_nr_without.deff_weight(nr_wgt_without_adj_class)["__none__"] == deff_wgt
-    assert sample_wgt_nr_without.deff_wgt["__none__"] == deff_wgt
+    assert sample_wgt_nr_without.deff_weight(nr_wgt_without_adj_class) == deff_wgt
+    assert sample_wgt_nr_without.deff_wgt == deff_wgt
 
 
 """Non-response adjustment WITH adjustment classes"""
