### 2021-02-21 1982728
fixed types using mypy
