### 2021-02-21 3fae671
fixed types using mypy
