+++ src/samplics/weighting/adjustment.py
@@ -8,13 +8,23 @@ sample weights.
    Calculation*, Stata Press.
 """
 
-from typing import Dict, List, Optional, Tuple, Union
+from __future__ import annotations
+from typing import Optional, Union
 
 import numpy as np
 import pandas as pd
 
 from samplics.utils import checks, formats
-from samplics.utils.types import Array, Number, StringNumber, DictStrNum
+from samplics.utils.types import (
+    Array,
+    DictStrFloat,
+    DictStrInt,
+    Number,
+    StringNumber,
+    DictStrNum,
+    DictStrInt,
+    DictStrFloat,
+)
 
 
 class SampleWeight:
@@ -48,40 +58,39 @@ class SampleWeight:
     def __init__(self) -> None:
 
         self.adjust_method: str = ""
-        self.number_units: Dict[StringNumber, int] = {}
-        self.deff_wgt: Dict[StringNumber, Number] = {}
-        self.adjust_factor: Dict[StringNumber, Number] = {}
-        self.control: Dict[StringNumber, Number] = {}
+        self.number_units: Union[DictStrInt, int] = {}
+        self.deff_wgt: Union[DictStrNum, Number] = {}
+        self.adjust_factor: Union[DictStrNum, Number] = {}
+        self.control: Union[DictStrNum, Number] = {}
 
-    def __repr__(self) -> None:
+    def __repr__(self) -> str:
         pass
 
-    def __str__(self) -> None:
+    def __str__(self) -> str:
         pass
 
-    def _number_units(self, domain: np.ndarray, samp_weight: np.ndarray) -> None:
+    def _number_units(self, domain: Optional[np.ndarray], samp_weight: np.ndarray) -> None:
         """Returns the number of units"""
 
-        if domain in (None, dict()):
+        if domain is None:
             self.number_units = len(samp_weight)
         elif domain is not None:
             keys, values = np.unique(domain, return_counts=True)
             self.number_units = dict(zip(keys, values))
 
     @staticmethod
-    def _deff_wgt(samp_weight: np.ndarray) -> float:
+    def _deff_wgt(samp_weight: np.ndarray) -> Number:
         """compute the design effect due to unequal weights -
         Page 71 of Valliant and Dever (2018)"""
 
         mean_w = np.mean(samp_weight)
         relvar_w = np.power(samp_weight - mean_w, 2) / mean_w ** 2
-        deff_w = 1 + np.mean(relvar_w)
 
-        return float(deff_w)
+        return float(1 + np.mean(relvar_w))
 
     def deff_weight(
         self, samp_weight: Array, domain: Optional[np.ndarray] = None
-    ) -> Dict[StringNumber, Number]:
+    ) -> Union[DictStrNum, Number]:
         """Computes the design effect due to unequal weights.
 
         Args:
@@ -91,41 +100,41 @@ class SampleWeight:
                 for each sample unit. Defaults to None. Defaults to None.
 
         Returns:
-            Dict[StringNumber, Number]: dictionnary pairing the domains to the design effects due
+            DictStrNum: dictionnary pairing the domains to the design effects due
                 unequal weights.
         """
 
         samp_weight = formats.numpy_array(samp_weight)
 
-        deff_w: Dict[StringNumber, Number] = {}
         if domain is None:
-            deff_w = self._deff_wgt(samp_weight)
+            self.deff_wgt = self._deff_wgt(samp_weight)
+            return self.deff_wgt
         else:
+            self.deff_wgt = {}
             for d in np.unique(domain):
-                deff_w[d] = self._deff_wgt(samp_weight[domain == d])
-        self.deff_wgt = deff_w
-
-        return deff_w
[…]
