### 2021-02-21 55bc43a
fixed types using mypy
