### 2021-02-21 cefd2e5
fixed types using mypy
