+++ src/samplics/estimation/expansion.py
@@ -11,10 +11,10 @@ and Wolter, K.M. (2007) [#w2007]_.
 .. [#w2007] Wolter, K.M. (2007), *Introduction to Variance Estimation, 2nd edn.*, 
    Springer-Verlag New York, Inc
 """
+from __future__ import annotations
+from typing import Any, Optional, Union
 
 import math
-from typing import TypeVar, Generic, Any, Dict, List, Optional, Union, Tuple
-
 import numpy as np
 import pandas as pd
 from samplics.utils.formats import fpc_as_dict, numpy_array, remove_nans, sample_size_dict
@@ -28,6 +28,7 @@ class _SurveyEstimator:
     def __init__(self, parameter: str, alpha: float = 0.05, random_seed: Optional[int] = None):
         """Initializes the instance """
 
+        self.random_seed: Optional[int]
         if random_seed is not None:
             self.random_seed = random_seed
             np.random.seed(self.random_seed)
@@ -91,8 +92,8 @@ class _SurveyEstimator:
     def _degree_of_freedom(
         self,
         samp_weight: np.ndarray,
-        stratum: np.ndarray = None,
-        psu: np.ndarray = None,
+        stratum: Optional[np.ndarray] = None,
+        psu: Optional[np.ndarray] = None,
     ) -> None:
 
         stratum = numpy_array(stratum)
@@ -119,7 +120,7 @@ class _SurveyEstimator:
             return float(np.sum(samp_weight * y) / np.sum(samp_weight))
         elif self.parameter == "total":
             return float(np.sum(samp_weight * y))
-        elif self.parameter == "ratio":
+        elif self.parameter == "ratio" and x is not None:
             return float(np.sum(samp_weight * y) / np.sum(samp_weight * x))
         else:
             raise ValueError("Parameter not valid!")
@@ -132,7 +133,7 @@ class _SurveyEstimator:
         domain: Optional[np.newaxis] = None,
         as_factor: bool = False,
         remove_nan: bool = False,
-    ) -> Union[Dict[StringNumber, Dict[StringNumber, float]], Dict[StringNumber, float], float]:
+    ) -> Union[dict[StringNumber, dict[StringNumber, float]], dict[StringNumber, float], float]:
         """Computes the parameter point estimates
 
         Args:
@@ -152,7 +153,7 @@ class _SurveyEstimator:
             raise AssertionError("Parameter x must be provided for ratio estimation.")
 
         if remove_nan:
-            if self.parameter == "ratio":
+            if self.parameter == "ratio" and x is not None:
                 excluded_units = np.isnan(y) | np.isnan(x)
             else:
                 excluded_units = np.isnan(y)
@@ -168,7 +169,7 @@ class _SurveyEstimator:
 
         if domain is None:
             if self.parameter == "proportion" or as_factor:
-                cat_dict: Dict[StringNumber, float] = {}
+                cat_dict: dict[StringNumber, float] = {}
                 for k in range(categories.size):
                     y_k = y_dummies[:, k]
                     cat_dict_k = dict({categories[k]: self._get_point_d(y_k, samp_weight)})
@@ -179,10 +180,10 @@ class _SurveyEstimator:
         else:
             domain_ids = np.unique(domain)
             if self.parameter == "proportion" or as_factor:
-                estimate1: Dict[StringNumber, Dict[StringNumber, float]] = {}
+                estimate1: dict[StringNumber, dict[StringNumber, float]] = {}
                 for d in domain_ids:
                     weight_d = samp_weight[domain == d]
-                    cat_dict_d: Dict[StringNumber, float] = {}
+                    cat_dict_d: dict[StringNumber, float] = {}
                     for k in range(categories.size):
                         y_d_k = y_dummies[domain == d, k]
[…]
