### 2021-02-21 86e7ad8
fixed types using mypy
