### 2021-02-21 2697034
fixed types using mypy
