+++ src/samplics/estimation/replication.py
@@ -2,10 +2,11 @@ import math
 from typing import TypeVar, Any, Dict, Optional, Tuple, Union
 
 import numpy as np
+from numpy.lib.arraysetops import isin
 import pandas as pd
 from samplics.estimation.expansion import _SurveyEstimator
 from samplics.utils import formats
-from samplics.utils.types import Array, Number, StringNumber
+from samplics.utils.types import Array, DictStrNum, Number, StringNumber
 from scipy.stats import t as student
 
 TypeRepEst = TypeVar("TypeRepEst", bound="ReplicateEstimator")
@@ -60,13 +61,12 @@ class ReplicateEstimator(_SurveyEstimator):
         self.conservative = False
         self.degree_of_freedom: Optional[int] = None
         self.number_reps: Optional[int] = None
-        self.rep_coefs: Optional[np.ndarray] = None
+        self.rep_coefs: Optional[Union[np.ndarray, Number]] = None
         if method == "brr" and fay_coef is not None:
             self.fay_coef = fay_coef
         elif method == "brr":
             self.fay_coef = 0
         if rep_weight_cls is not None:
-            # self.method = rep_weight_cls.rep_method
             self.number_reps = rep_weight_cls.number_reps
             self.rep_coefs = rep_weight_cls.rep_coefs
             self.degree_of_freedom = rep_weight_cls.degree_of_freedom
@@ -74,13 +74,17 @@ class ReplicateEstimator(_SurveyEstimator):
 
     def _rep_point(self, y: np.ndarray, rep_weights: np.ndarray, x: np.ndarray) -> np.ndarray:
         if self.parameter in ("proportion", "mean"):
-            return np.sum(rep_weights * y[:, None], axis=0) / np.sum(rep_weights, axis=0)
-        elif self.parameter == "total":
-            return np.sum(rep_weights * y[:, None], axis=0)
-        elif self.parameter == "ratio":
-            return np.sum(rep_weights * y[:, None], axis=0) / np.sum(
-                rep_weights * x[:, None], axis=0
+            return np.asarray(
+                np.sum(rep_weights * y[:, None], axis=0) / np.sum(rep_weights, axis=0)
             )
+        elif self.parameter == "total":
+            return np.asarray(np.sum(rep_weights * y[:, None], axis=0))
+        elif self.parameter == "ratio":
+            return np.asarray(
+                np.sum(rep_weights * y[:, None], axis=0) / np.sum(rep_weights * x[:, None], axis=0)
+            )
+        else:
+            raise AssertionError("Parameter not valid!")
 
     def _bias(
         self,
@@ -88,27 +92,29 @@ class ReplicateEstimator(_SurveyEstimator):
         samp_weight: np.ndarray,
         rep_weights: np.ndarray,
         x: np.ndarray,
-    ) -> float:
+    ) -> Number:
 
         estimate = self._get_point(y, samp_weight, x)
-        rep_estimates = self._rep_point(y, rep_weights, x)
+        if isinstance(estimate, (int, float)):
+            rep_estimates = self._rep_point(y, rep_weights, x)
+            return float(np.sum(np.mean(rep_estimates) - estimate))
+        else:
+            raise AssertionError
 
-        return float(np.sum(np.mean(rep_estimates) - estimate))
-
-    def _rep_coefs(self, rep_coefs: np.ndarray = None) -> None:
+    def _rep_coefs(self, rep_coefs: Optional[Union[np.ndarray, Number]] = None) -> None:
 
         if rep_coefs is not None:
             if isinstance(rep_coefs, np.ndarray):
                 self.rep_coefs = rep_coefs
             elif isinstance(rep_coefs, (int, float)):
-                self.rep_coefs = [rep_coefs]
-        elif self.rep_coefs is None and self.method == "bootstrap":
[…]
