### 2021-02-21 668d7dd
fixed types using mypy
