+++ src/samplics/estimation/replication.py
@@ -72,14 +72,16 @@ class ReplicateEstimator(_SurveyEstimator):
             self.degree_of_freedom = rep_weight_cls.degree_of_freedom
             self.fay_coef = rep_weight_cls.fay_coef if self.method == "brr" else None
 
-    def _rep_point(self, y: np.ndarray, rep_weights: np.ndarray, x: np.ndarray) -> np.ndarray:
+    def _rep_point(
+        self, y: np.ndarray, rep_weights: np.ndarray, x: Optional[np.ndarray]
+    ) -> np.ndarray:
         if self.parameter in ("proportion", "mean"):
             return np.asarray(
                 np.sum(rep_weights * y[:, None], axis=0) / np.sum(rep_weights, axis=0)
             )
         elif self.parameter == "total":
             return np.asarray(np.sum(rep_weights * y[:, None], axis=0))
-        elif self.parameter == "ratio":
+        elif self.parameter == "ratio" and x is not None:
             return np.asarray(
                 np.sum(rep_weights * y[:, None], axis=0) / np.sum(rep_weights * x[:, None], axis=0)
             )
@@ -91,7 +93,7 @@ class ReplicateEstimator(_SurveyEstimator):
         y: np.ndarray,
         samp_weight: np.ndarray,
         rep_weights: np.ndarray,
-        x: np.ndarray,
+        x: Optional[np.ndarray],
     ) -> Number:
 
         estimate = self._get_point(y, samp_weight, x)
@@ -124,7 +126,7 @@ class ReplicateEstimator(_SurveyEstimator):
         y: np.ndarray,
         rep_weights: np.ndarray,
         rep_coefs: np.ndarray,
-        x: np.ndarray,
+        x: Optional[np.ndarray],
         estimate: float,
         conservative: bool = False,
     ) -> Number:
@@ -168,7 +170,7 @@ class ReplicateEstimator(_SurveyEstimator):
     ) -> Any:
 
         if remove_nan:
-            if self.parameter == "ratio":
+            if self.parameter == "ratio" and x is not None:
                 excluded_units = np.isnan(y) | np.isnan(x)
             else:
                 excluded_units = np.isnan(y)
@@ -182,7 +184,7 @@ class ReplicateEstimator(_SurveyEstimator):
             categories = y_dummies.columns
             y_dummies = y_dummies.values
 
-        bias: Dict[StringNumber, Any] = {}
+        bias: Any
         if domain is None:
             if self.parameter == "proportion" and categories is not None:
                 cat_dict = dict()
@@ -240,19 +242,17 @@ class ReplicateEstimator(_SurveyEstimator):
         remove_nan: bool = False,
     ) -> Any:
 
-        categories = None
-        y_dummies = None
         if self.parameter == "proportion":
             y_dummies = pd.get_dummies(y)
             categories = y_dummies.columns
             y_dummies = y_dummies.values
 
-        variance: Dict[StringNumber, Any] = {}
         if domain is None:
             if self.parameter == "proportion":
+                estimate_k: dict
                 cat_dict = dict()
                 for k in range(categories.size):
-                    estimate_k = self._get_point(y_dummies[:, k], samp_weight, x)[1]
+                    estimate_k = self._get_point(y_dummies[:, k], samp_weight, x)
                     cat_dict_k = dict(
                         {
                             categories[k]: self._variance(
@@ -260,17 +260,19 @@ class ReplicateEstimator(_SurveyEstimator):
                                 rep_weights,
                                 rep_coefs,
                                 x,
-                                estimate_k,
+                                estimate_k[1],
                                 conservative,
                             )
                         }
                     )
                     cat_dict.update(cat_dict_k)
-                variance = cat_dict
+                return cat_dict
             else:
[…]
