### 2021-02-21 30e96d4
fixed types using mypy
