+++ src/samplics/categorical/tabulation.py
@@ -4,8 +4,8 @@ The module implements the cross-tabulation analysis.
 
 """
 
-
-from typing import Any, Dict, List, Optional, Union, Tuple
+from __future__ import annotations
+from typing import Any, Optional, Union
 
 import itertools
 
@@ -39,17 +39,17 @@ class Tabulation:
         else:
             raise ValueError("parameter must be 'count' or 'proportion'")
         self.type = "oneway"
-        self.point_est: Dict[str, Dict[StringNumber, Number]] = {}
-        self.stats: Dict[str, Dict[str, Number]] = {}
-        self.stderror: Dict[str, Dict[str, Number]] = {}
-        self.lower_ci: Dict[str, Dict[str, Number]] = {}
-        self.upper_ci: Dict[str, Dict[str, Number]] = {}
-        self.deff: Dict[str, Dict[str, Number]] = {}
+        self.point_est: dict[str, dict[StringNumber, Number]] = {}
+        self.stats: dict[str, dict[str, Number]] = {}
+        self.stderror: dict[str, dict[str, Number]] = {}
+        self.lower_ci: dict[str, dict[str, Number]] = {}
+        self.upper_ci: dict[str, dict[str, Number]] = {}
+        self.deff: dict[str, dict[str, Number]] = {}
         self.alpha: float = alpha
         self.ciprop_method: str = ciprop_method
-        self.design_info: Dict[str, Number] = {}
-        self.vars_names: List[str] = []
-        self.vars_levels: Dict[str, StringNumber] = {}
+        self.design_info: dict[str, Number] = {}
+        self.vars_names: list[str] = []
+        self.vars_levels: dict[str, list[StringNumber]] = {}
 
     def __repr__(self) -> str:
         return f"Tabulation(parameter={self.parameter}, alpha={self.alpha})"
@@ -70,16 +70,16 @@ class Tabulation:
     def _estimate(
         self,
         var_of_ones: Array,
-        var: Array,
+        var: pd.DataFrame,
         samp_weight: Array = None,
         stratum: Optional[Array] = None,
         psu: Optional[Array] = None,
         ssu: Optional[Array] = None,
-        fpc: Union[Dict, float] = 1,
+        fpc: Union[dict, float] = 1,
         deff: bool = False,
         coef_variation: bool = False,
         remove_nan: bool = False,
-    ) -> Tuple[TaylorEstimator, list, int]:
+    ) -> tuple[TaylorEstimator, list, int]:
 
         if remove_nan:
             excluded_units = var.isna().values.ravel()
@@ -90,6 +90,8 @@ class Tabulation:
         else:
             var.fillna("nan", inplace=True)
         var = var.to_numpy().ravel()
+
+        var_of_ones = numpy_array(var_of_ones)
 
         if self.parameter == "count":
             tbl_est = TaylorEstimator(parameter="total", alpha=self.alpha)
@@ -126,13 +128,13 @@ class Tabulation:
     def tabulate(
         self,
         vars: Array,
-        varnames: Optional[Union[str, List[str]]] = None,
+        varnames: Optional[Union[str, list[str]]] = None,
         samp_weight: Optional[Union[Array, Number]] = None,
         stratum: Optional[Array] = None,
         psu: Optional[Array] = None,
         ssu: Optional[Array] = None,
         # Todo: by: Optional[Array] = None,
-        fpc: Union[Dict, float] = 1,
+        fpc: Union[dict, float] = 1,
         deff: bool = False,
         coef_variation: bool = False,
         remove_nan: bool = False,
@@ -198,7 +200,6 @@ class Tabulation:
[…]
