### 2021-02-21 63a5fbb
fixed types using mypy
