### 2021-02-21 b2e9b65
fixed types using mypy
