+++ src/samplics/sae/eblup_area_model.py
@@ -14,9 +14,10 @@ level.
    Association*, **74**, 269-277.
 """
 
-import math
-from typing import Any, Dict, List, Tuple, Union
+from __future__ import annotations
+from typing import Any, Union
 
+import math
 import numpy as np
 import pandas as pd
 from samplics.utils import formats
@@ -87,16 +88,16 @@ class EblupAreaModel:
         self.fitted: bool = False
         self.fixed_effects: np.ndarray = np.array([])
         self.fe_std: np.ndarray = np.array([])
-        self.re_std: np.ndarray = np.array([])
-        self.convergence: Dict[str, Union[float, int, bool]] = {}
-        self.goodness: Dict[str, float] = {}  # loglikehood, deviance, AIC, BIC
+        self.re_std: Number
+        self.convergence: dict[str, Union[float, int, bool]] = {}
+        self.goodness: dict[str, float] = {}  # loglikehood, deviance, AIC, BIC
 
         # Predict(ino/ed) data
-        self.area_est: Dict[Any, float] = {}
-        self.area_mse: Dict[Any, float] = {}
-        self.area_mse_as1: Dict[Any, float] = {}
-        self.area_mse_as2: Dict[Any, float] = {}
-        self.area_mse_terms: Dict[str, Dict[Any, float]] = {}
+        self.area_est: dict[Any, float] = {}
+        self.area_mse: dict[Any, float] = {}
+        self.area_mse_as1: dict[Any, float] = {}
+        self.area_mse_as2: dict[Any, float] = {}
+        self.area_mse_terms: dict[str, dict[Any, float]] = {}
 
     def __str__(self) -> str:
 
@@ -122,7 +123,7 @@ class EblupAreaModel:
         sigma2_e: np.ndarray,
         sigma2_v: float,
         b_const: np.ndarray,
-    ) -> Tuple[np.ndarray, np.ndarray]:
+    ) -> tuple[np.ndarray, np.ndarray]:
 
         V = np.diag(sigma2_v * (b_const ** 2) + sigma2_e)
         V_inv = np.linalg.inv(V)
@@ -135,7 +136,7 @@ class EblupAreaModel:
 
     def _log_likelihood(
         self, y: np.ndarray, X: np.ndarray, beta: np.ndarray, V: np.ndarray
-    ) -> float:
+    ) -> Number:
 
         m = y.size
         const = m * np.log(2 * np.pi)
@@ -162,9 +163,9 @@ class EblupAreaModel:
         yhat: np.ndarray,
         X: np.ndarray,
         sigma2_e: np.ndarray,
-        sigma2_v: np.ndarray,
+        sigma2_v: Number,
         b_const: np.ndarray,
-    ) -> Tuple[float, float]:
+    ) -> tuple[Number, Number]:
 
         deriv_sigma = 0.0
         info_sigma = 0.0
@@ -199,10 +200,10 @@ class EblupAreaModel:
             P = v_inv - np.matmul(np.matmul(v_inv, x_xvinvx_x), v_inv)
             P_B = np.matmul(P, B)
             P_B_P = np.matmul(P_B, P)
-            term1 = np.trace(P_B)
+            term1 = float(np.trace(P_B))
             term2 = np.matmul(np.matmul(np.transpose(yhat), P_B_P), yhat)
             deriv_sigma = -0.5 * (term1 - term2)
-            info_sigma = 0.5 * np.trace(np.matmul(P_B_P, B))
+            info_sigma = 0.5 * float(np.trace(np.matmul(P_B_P, B)))
         elif self.method == "FH":  # Fay-Herriot approximation
             beta, beta_cov = self._fixed_coefficients(
                 area=area,
@@ -238,7 +239,7 @@ class EblupAreaModel:
         sigma2_v_start: float,
         tol: float,
         maxiter: int,
[…]
