### 2021-02-22 817389e
fixed types using mypy
