+++ src/samplics/sae/eblup_unit_model.py
@@ -25,15 +25,17 @@ see Rao, J.N.K. and Molina, I. (2015) [#rm2015]_.
    John Wiley & Sons, Hoboken, New Jersey.
 """
 
+from __future__ import annotations
+from typing import Any, Optional, Union
+
 import warnings
-from typing import Any, Dict, List, Optional, Union
 
 import numpy as np
 import pandas as pd
 import statsmodels.api as sm
 from samplics.sae.sae_core_functions import area_stats
 from samplics.utils import formats, basic_functions
-from samplics.utils.types import Array, Number
+from samplics.utils.types import Array, DictStrNum, Number, StringNumber
 
 
 class EblupUnitModel:
@@ -101,35 +103,35 @@ class EblupUnitModel:
             raise AssertionError("Value provided for method is not valid!")
 
         # Sample data
-        self.scales: np.ndarray = np.array([])
-        self.afactors: Dict[Any, float] = {}
-        self.ys: np.ndarray = np.array([])
-        self.Xs: np.ndarray = np.array([])
-        self.areas: np.ndarray = np.array([])
-        self.areas_list: np.ndarray = np.array([])
-        self.samp_size: Dict[Any, int] = {}
-        self.ys_mean: np.ndarray = np.array([])
-        self.Xs_mean: np.ndarray = np.array([])
+        self.scales: np.ndarray 
+        self.afactors: DictStrNum
+        self.ys: np.ndarray 
+        self.Xs: np.ndarray
+        self.areas: np.ndarray 
+        self.areas_list: np.ndarray 
+        self.samp_size: DictStrNum
+        self.ys_mean: np.ndarray
+        self.Xs_mean: np.ndarray
 
         # Fitting stats
         self.fitted: bool = False
-        self.fixed_effects: np.ndarray = np.array([])
-        self.fe_std: np.ndarray = np.array([])
-        self.random_effects: np.ndarray = np.array([])
-        self.re_std: float = 0
-        self.error_std: float = 0
-        self.convergence: Dict[str, Union[float, int, bool]] = {}
-        self.goodness: Dict[str, float] = {}  # loglikehood, deviance, AIC, BIC
-        self.gamma: Dict[Any, float] = {}
+        self.fixed_effects: np.ndarray 
+        self.fe_std: np.ndarray 
+        self.random_effects: np.ndarray 
+        self.re_std: Number = 0
+        self.error_std: Number = 0
+        self.convergence: dict[str, Union[float, int, bool]] = {}
+        self.goodness: dict[str, Number] = {}  # loglikehood, deviance, AIC, BIC
+        self.gamma: DictStrNum
 
         # Predict(ion/ed) data
-        self.areap: np.ndarray = np.array([])
-        self.Xp_mean: np.ndarray = np.array([])
+        self.areap: np.ndarray
+        self.Xp_mean: np.ndarray
         self.number_reps: int = 0
-        self.samp_rate: Dict[Any, float] = None
-        self.area_est: Dict[Any, float] = None
-        self.area_mse: Dict[Any, float] = None
-        self.area_mse_boot: Optional[Dict[Any, float]] = None
+        self.samp_rate: DictStrNum
+        self.area_est: DictStrNum
+        self.area_mse: DictStrNum
+        self.area_mse_boot: Optional[DictStrNum] = None
 
     def _beta(
         self,
@@ -154,9 +156,8 @@ class EblupUnitModel:
             beta1 = beta1 + np.matmul(np.transpose(Xw_d), resid_d_w)
             beta2 = beta2 + np.sum(resid_d_w * y_d[:, None] * w_d[:, None], axis=0)
 
-        beta = np.matmul(np.linalg.inv(beta1), beta2)
-
[…]
