### 2021-02-22 5865990
fixed types using mypy
