+++ src/samplics/sae/robust_unit_model.py
@@ -23,7 +23,8 @@ see Rao, J.N.K. and Molina, I. (2015) [#rm2015]_.
    John Wiley & Sons, Hoboken, New Jersey.
 """
 
-from typing import Any, Dict, Optional, Tuple, Union, Callable
+from __future__ import annotations
+from typing import Any, Optional, Union, Callable
 
 import numpy as np
 import pandas as pd
@@ -31,7 +32,7 @@ import pandas as pd
 import statsmodels.api as sm
 
 from samplics.utils import formats, basic_functions
-from samplics.utils.types import Array, Number
+from samplics.utils.types import Array, DictStrNum, Number
 
 from samplics.sae.sae_core_functions import area_stats
 from samplics.sae.eb_unit_model import EbUnitModel
@@ -105,7 +106,7 @@ class EllUnitModel:
         self,
         method: str = "MOM",
         boxcox: Optional[float] = None,
-        constant: Number = 0,
+        constant: Optional[Number] = None,
         indicator: Optional[Any] = None,
     ):
 
@@ -114,36 +115,36 @@ class EllUnitModel:
         if self.method not in ("REML", "ML", "MOM"):
             raise AssertionError("Value provided for method is not valid!")
         self.indicator = indicator
-        self.boxcox = {"lambda": boxcox, "constant": constant}
+        self.boxcox: dict[str, Optional[Number]] = {"lambda": boxcox, "constant": constant}
 
         # Sample data
-        self.scales: np.ndarray = np.array([])
-        self.afactors: Dict[Any, float] = {}
-        self.ys: np.ndarray = np.array([])
-        self.Xs: np.ndarray = np.array([])
-        self.areas: np.ndarray = np.array([])
-        self.areas_list: np.ndarray = np.array([])
-        self.samp_size: Dict[Any, int] = {}
-        self.ys_mean: np.ndarray = np.array([])
-        self.Xs_mean: np.ndarray = np.array([])
+        self.scales: np.ndarray
+        self.afactors: DictStrNum
+        self.ys: np.ndarray
+        self.Xs: np.ndarray
+        self.areas: np.ndarray
+        self.areas_list: np.ndarray
+        self.samp_size: DictStrNum
+        self.ys_mean: np.ndarray
+        self.Xs_mean: np.ndarray
 
         # Fitted data
         self.fitted: bool = False
-        self.fixed_effects: np.ndarray = np.array([])
-        self.fe_std: np.ndarray = np.array([])
-        self.random_effects: np.ndarray = np.array([])
-        self.re_std: float = 0
-        self.error_std: float = 0
-        self.convergence: Dict[str, Union[float, int, bool]] = {}
-        self.goodness: Dict[str, float] = {}  # loglikehood, deviance, AIC, BIC
-        self.gamma: Dict[Any, float] = {}
+        self.fixed_effects: np.ndarray
+        self.fe_std: np.ndarray
+        self.random_effects: np.ndarray
+        self.re_std: float
+        self.error_std: float
+        self.convergence: dict[str, Union[float, int, bool]] = {}
+        self.goodness: dict[str, Number] = {}  # loglikehood, deviance, AIC, BIC
+        self.gamma: DictStrNum
 
         # Predict(ion/ed) data
-        self.areap: np.ndarray = np.array([])
-        self.Xp_mean: np.ndarray = np.array([])
-        self.number_reps: int = 0
-        self.area_est: Dict[Any, float] = {}
-        self.area_mse: Dict[Any, float] = {}
+        self.areap: np.ndarray
+        self.Xp_mean: np.ndarray
+        self.number_reps: int
+        self.area_est: DictStrNum
+        self.area_mse: DictStrNum
 
     def fit(
         self,
[…]

[…]
