### 2021-02-22 0d6ce97
fixed types using mypy
