### 2021-02-22 3f807ad
fixed types using mypy
