### 2021-02-22 2750065
fixed types using mypy
