### 2021-02-22 13c9681
fixed types using mypy
