+++ src/samplics/__init__.py
@@ -21,4 +21,4 @@ __all__ = [
     "transform",
 ]
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
