### 2021-02-22 dca11d5
fixed types using mypy
