### 2021-02-23 #2908 @rchew
Hi Lorena,

I'd be happy to review the submission -- let me know how I can help.

Best,

Rob

On Mon, Feb 22, 2021 at 11:10 AM Lorena Mesa <notifications@github.com>
wrote:

> @rchew <https://github.com/rchew> would you be willing to review this
> submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
