### 2021-02-24 #2908 @lorenanicole
> Hi Lorena, I'd be happy to review the submission -- let me know how I can help. Best, Rob
> […](#)
> On Mon, Feb 22, 2021 at 11:10 AM Lorena Mesa ***@***.***> wrote: @rchew <https://github.com/rchew> would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub <[#2908 (comment)](https://github.com/openjournals/joss-reviews/issues/2908#issuecomment-783485760)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AAYCRWXXA2ZGR5EQRNRGFL3TAJ6X7ANCNFSM4U3G2WKQ> .

Excellent! We are seeking for a second reviewer and when we have that locked down you'll be notified on next step(s).
