### 2021-02-24 #2908 @lorenanicole
👋 @acolum would you be willing to review this submission for JOSS?

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html .
