### 2021-02-25 #2908 @acolum
Hi Lorena,

Thank you for contacting me, but I have too many other commitments at the
moment to review this submission.

Best,
Alyssa

On Wed, Feb 24, 2021 at 10:56 AM Lorena Mesa <notifications@github.com>
wrote:

> 👋 @acolum <https://github.com/acolum> would you be willing to review
> this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html .
>
