### 2021-02-27 f4511bd
adding features to OneSampleSize
