### 2021-02-27 3c64f93
adding features to OneSampleSize
