### 2021-02-27 b2534bc
adding features to OneSampleSize
