### 2021-02-27 4c9fa9f
adding features to OneSampleSize
