### 2021-02-27 f5be076
adding features to OneSampleSize
