+++ src/samplics/sampling/__init__.py
@@ -1,4 +1,4 @@
 from samplics.sampling.selection import SampleSelection
-from samplics.sampling.size import OneSampleSize
+from samplics.sampling.size import SampleSize
 
 __all__ = ["SampleSelection", "SampleSize"]
