+++ src/samplics/sampling/__init__.py
@@ -1,4 +1,4 @@
 from samplics.sampling.selection import SampleSelection
-from samplics.sampling.size import SampleSize
+from samplics.sampling.size import SampleSize, allocate
 
-__all__ = ["SampleSelection", "SampleSize"]
+__all__ = ["allocate", "SampleSelection", "SampleSize"]
