### 2021-02-27 750122b
adding features to OneSampleSize
