### 2021-02-27 4fcf522
adding features to OneSampleSize
