### 2021-02-27 56d7e7c
adding allocate function
