### 2021-02-28 e177e9a
adding allocate function
