+++ tests/sampling/test_size_calculation.py
@@ -8,74 +8,128 @@ pop_size = {"Dakar": 500, "Kaolack": 300, "Ziguinchor": 200}
 
 
 @pytest.mark.xfail(reason="stratum is required")
-def test_deff_equal_stratum_error():
+def testation_equal_stratum_error():
     allocate(method="equal")
 
 
-def test_deff_equal_alloc():
-    equal_alloc = allocate(method="equal", stratum=region, constant=15)
-    assert equal_alloc["Dakar"] == 15
-    assert equal_alloc["Kaolack"] == 15
-    assert equal_alloc["Ziguinchor"] == 15
+def testation_equal():
+    equal = allocate(method="equal", stratum=region, pop_size=pop_size, constant=15)
+    assert equal["Dakar"] == 15
+    assert equal["Kaolack"] == 15
+    assert equal["Ziguinchor"] == 15
 
 
-def test_deff_equal_alloc_error():
+def testation_equal_error():
     with pytest.raises(ValueError):
-        allocate(method="equal", stratum=region, constant=[23])
+        allocate(method="equal", stratum=region, pop_size=pop_size, constant=[23])
 
 
-def test_deff_proportional_alloc():
-    equal_alloc = allocate(method="proportional", stratum=region, samp_size=100, pop_size=pop_size)
-    assert equal_alloc["Dakar"] == 50
-    assert equal_alloc["Kaolack"] == 30
-    assert equal_alloc["Ziguinchor"] == 20
+def testation_proportional():
+    equal = allocate(method="proportional", stratum=region, samp_size=100, pop_size=pop_size)
+    assert equal["Dakar"] == 50
+    assert equal["Kaolack"] == 30
+    assert equal["Ziguinchor"] == 20
 
 
-def test_deff_proportional_alloc_error():
+def testation_proportional_error():
     with pytest.raises(ValueError):
         allocate(method="equal", stratum=region, samp_size=100, pop_size=5)
 
 
-def test_deff_rate_number_alloc():
-    equal_alloc = allocate(method="fixed_rate", stratum=region, pop_size=pop_size, rate=0.05)
-    assert equal_alloc["Dakar"] == 25
-    assert equal_alloc["Kaolack"] == 15
-    assert equal_alloc["Ziguinchor"] == 10
+def testation_fixed_rate_number():
+    equal = allocate(method="fixed_rate", stratum=region, pop_size=pop_size, rate=0.05)
+    assert equal["Dakar"] == 25
+    assert equal["Kaolack"] == 15
+    assert equal["Ziguinchor"] == 10
 
 
-def test_deff_rate_dict_alloc():
-    rate = {"Dakar": 0.05, "Kaolack": 0.10, "Ziguinchor": 0.20}
-    equal_alloc = allocate(method="fixed_rate", stratum=region, pop_size=pop_size, rate=rate)
-    assert equal_alloc["Dakar"] == 25
-    assert equal_alloc["Kaolack"] == 30
-    assert equal_alloc["Ziguinchor"] == 40
-
-
-def test_deff_rate_alloc_error():
+def testation_fixed_rate_error():
     with pytest.raises(ValueError):
         allocate(method="fixed_rate", stratum=region, pop_size=[5])
 
 
-def test_deff_proportional_rate_alloc():
+def testation_variable_rate():
+    rate = {"Dakar": 0.05, "Kaolack": 0.10, "Ziguinchor": 0.20}
+    equal = allocate(method="variable_rate", stratum=region, pop_size=pop_size, rate=rate)
+    assert equal["Dakar"] == 25
+    assert equal["Kaolack"] == 30
+    assert equal["Ziguinchor"] == 40
+
+
[…]
