### 2021-02-28 4fa5340
adding allocate function
