### 2021-02-28 6739a85
adding allocate function
