### 2021-02-28 c23c220
adding allocate function
