### 2021-02-28 0f9b123
adding allocate function
