### 2021-02-28 45e541f
adding allocate function
