### 2021-02-28 679ac53
adding calculate_clusters function
