### 2021-03-06 3be0bdc
adding sample size calculation for one sample
