+++ tests/sampling/test_sample_size.py
@@ -250,7 +250,7 @@ def test_size_nat_wald_size():
     assert size_nat_wald.samp_size == 62
     assert size_nat_wald.deff_c == 1.0
     assert size_nat_wald.target == 0.80
-    assert size_nat_wald.precision == 0.1
+    assert size_nat_wald.half_ci == 0.1
 
 
 def test_size_nat_wald_size_with_deff():
@@ -258,20 +258,20 @@ def test_size_nat_wald_size_with_deff():
     assert size_nat_wald.samp_size == 93
     assert size_nat_wald.deff_c == 1.5
     assert size_nat_wald.target == 0.80
-    assert size_nat_wald.precision == 0.1
+    assert size_nat_wald.half_ci == 0.1
 
 
 def test_size_nat_wald_df():
     size_nat_wald.calculate(0.80, 0.10)
     size_df = size_nat_wald.to_dataframe()
-    assert (size_df.columns == ["_target", "_precision", "_samp_size"]).all()
+    assert (size_df.columns == ["_target", "_half_ci", "_samp_size"]).all()
 
 
 ## Wald's method - stratified
 size_str_wald = SampleSize(parameter="Proportion", method="Wald", stratification=True)
 
 target = {"stratum1": 0.95, "stratum2": 0.70, "stratum3": 0.30}
-precision = {"stratum1": 0.30, "stratum2": 0.10, "stratum3": 0.15}
+half_ci = {"stratum1": 0.30, "stratum2": 0.10, "stratum3": 0.15}
 deff = {"stratum1": 1, "stratum2": 1.5, "stratum3": 2.5}
 
 
@@ -289,7 +289,7 @@ def test_size_str_wald_size1():
 
 
 def test_size_str_wald_size2():
-    size_str_wald.calculate(0.8, precision)
+    size_str_wald.calculate(0.8, half_ci)
     assert size_str_wald.samp_size["stratum1"] == 7
     assert size_str_wald.samp_size["stratum2"] == 62
     assert size_str_wald.samp_size["stratum3"] == 28
@@ -303,7 +303,7 @@ def test_size_str_wald_size3():
 
 
 def test_size_str_wald_size4():
-    size_str_wald.calculate(target, precision, deff)
+    size_str_wald.calculate(target, half_ci, deff)
     assert size_str_wald.samp_size["stratum1"] == 3
     assert size_str_wald.samp_size["stratum2"] == 122
     assert size_str_wald.samp_size["stratum3"] == 90
@@ -322,7 +322,7 @@ def test_size_str_wald_df():
     size_str_wald.calculate(0.80, 0.10, number_strata=5)
     size_df = size_str_wald.to_dataframe()
     assert size_df.shape[0] == 5
-    assert (size_df.columns == ["_stratum", "_target", "_precision", "_samp_size"]).all()
+    assert (size_df.columns == ["_stratum", "_target", "_half_ci", "_samp_size"]).all()
 
 
 ## Fleiss' method
@@ -341,7 +341,7 @@ def test_size_nat_fleiss_size1a():
     assert size_nat_fleiss.samp_size == 88
     assert size_nat_fleiss.deff_c == 1.0
     assert size_nat_fleiss.target == 0.80
-    assert size_nat_fleiss.precision == 0.1
+    assert size_nat_fleiss.half_ci == 0.1
 
 
 def test_size_nat_fleiss_size1b():
@@ -349,7 +349,7 @@ def test_size_nat_fleiss_size1b():
     assert size_nat_fleiss.samp_size == 88
     assert size_nat_fleiss.deff_c == 1.0
[…]
