+++ tests/utils/test_formats.py
@@ -86,3 +86,55 @@ def test_dict_to_dataframe():
     df = dict_to_dataframe(["domain", "size"], samp_size)
     assert isinstance(df, pd.DataFrame) == True
     assert (df.columns == ["domain", "size"]).all()
+
+
+dict1 = {"one": 1, "two": 2, "three": 3}
+dict2 = {"one": 11, "two": 22, "three": 33}
+dict3 = {"one": 11, "two": 22}
+dict4 = {"one": 11, "three": 22}
+
+
+def test_convert_numbers_to_dicts1():
+    dicts = convert_numbers_to_dicts(3, 100, 1000, 10000)
+    assert dicts[0]["_stratum_1"] == 100
+    assert dicts[0]["_stratum_2"] == 100
+    assert dicts[0]["_stratum_3"] == 100
+    assert dicts[1]["_stratum_1"] == 1000
+    assert dicts[1]["_stratum_2"] == 1000
+    assert dicts[1]["_stratum_3"] == 1000
+    assert dicts[2]["_stratum_1"] == 10000
+    assert dicts[2]["_stratum_2"] == 10000
+    assert dicts[2]["_stratum_3"] == 10000
+
+
+def test_convert_numbers_to_dicts2():
+    dicts = convert_numbers_to_dicts(3, 100, dict1, dict2)
+    assert dicts[0]["one"] == 100
+    assert dicts[0]["two"] == 100
+    assert dicts[0]["three"] == 100
+    assert dicts[1]["one"] == 1
+    assert dicts[1]["two"] == 2
+    assert dicts[1]["three"] == 3
+    assert dicts[2]["one"] == 11
+    assert dicts[2]["two"] == 22
+    assert dicts[2]["three"] == 33
+
+
+def test_convert_numbers_to_dicts3():
+    dicts = convert_numbers_to_dicts(3, dict1, dict2)
+    assert dicts[0]["one"] == 1
+    assert dicts[0]["two"] == 2
+    assert dicts[0]["three"] == 3
+    assert dicts[1]["one"] == 11
+    assert dicts[1]["two"] == 22
+    assert dicts[1]["three"] == 33
+
+
+@pytest.mark.xfail(reason="dictionnaries have different keys")
+def test_convert_numbers_to_dicts4():
+    convert_numbers_to_dicts(3, dict1, dict3)
+
+
+@pytest.mark.xfail(reason="dictionnaries have different keys")
+def test_convert_numbers_to_dicts5():
+    convert_numbers_to_dicts(3, dict3, dict4)
