+++ src/samplics/sampling/size.py
@@ -11,7 +11,7 @@ import math
 import numpy as np
 import pandas as pd
 
-from scipy.stats import norm as normal
+from scipy.stats import norm as normal, t as student
 
 from samplics.utils.formats import numpy_array, dict_to_dataframe
 from samplics.utils.types import Array, Number, StringNumber, DictStrNum
@@ -33,7 +33,7 @@ class SampleSize:
         self.samp_size: Union[DictStrNum, Number] = 0
         self.deff_c: Union[DictStrNum, Number] = 1.0
         self.deff_w: Union[DictStrNum, Number] = 1.0
-        self.precision: Union[DictStrNum, Number]
+        self.half_ci: Union[DictStrNum, Number]
         self.resp_rate: Union[DictStrNum, Number] = 1.0
         self.pop_size: Optional[Union[DictStrNum, Number]] = None
 
@@ -61,7 +61,7 @@ class SampleSize:
     def _calculate_wald(
         self,
         target: Union[DictStrNum, Number],
-        precision: Union[DictStrNum, Number],
+        half_ci: Union[DictStrNum, Number],
         pop_size: Optional[Union[DictStrNum, Number]],
         stratum: Optional[Array],
     ) -> Union[DictStrNum, Number]:
@@ -70,7 +70,7 @@ class SampleSize:
 
         if (
             isinstance(target, dict)
-            and isinstance(precision, dict)
+            and isinstance(half_ci, dict)
             and isinstance(self.deff_c, dict)
             and stratum is not None
         ):
@@ -83,16 +83,16 @@ class SampleSize:
                         * pop_size[s]
                         * z_value ** 2
                         * sigma_s
-                        / ((pop_size[s] - 1) * precision[s] ** 2 + z_value * sigma_s)
+                        / ((pop_size[s] - 1) * half_ci[s] ** 2 + z_value * sigma_s)
                     )
                 else:
                     samp_size[s] = math.ceil(
-                        self.deff_c[s] * z_value ** 2 * sigma_s / precision[s] ** 2
+                        self.deff_c[s] * z_value ** 2 * sigma_s / half_ci[s] ** 2
                     )
             return samp_size
         elif (
             isinstance(target, (int, float))
-            and isinstance(precision, (int, float))
+            and isinstance(half_ci, (int, float))
             and isinstance(self.deff_c, (int, float))
         ):
             sigma = target * (1 - target)
@@ -102,17 +102,17 @@ class SampleSize:
                     * pop_size
                     * z_value ** 2
                     * sigma
-                    / ((pop_size - 1) * precision ** 2 + z_value * sigma)
+                    / ((pop_size - 1) * half_ci ** 2 + z_value * sigma)
                 )
             else:
-                return math.ceil(self.deff_c * z_value ** 2 * sigma / precision ** 2)
+                return math.ceil(self.deff_c * z_value ** 2 * sigma / half_ci ** 2)
         else:
-            raise TypeError("target and precision must be numbers or dictionnaires!")
+            raise TypeError("target and half_ci must be numbers or dictionnaires!")
 
     def _calculate_fleiss(
         self,
         target: Union[DictStrNum, Number],
-        precision: Union[DictStrNum, Number],
+        half_ci: Union[DictStrNum, Number],
         # pop_size: Optional[Union[DictStrNum, Number]],
         stratum: Optional[Array],
     ) -> Union[DictStrNum, Number]:
@@ -135,18 +135,18 @@ class SampleSize:
         if (
             self.stratification
             and isinstance(target, dict)
-            and isinstance(precision, dict)
+            and isinstance(half_ci, dict)
             and isinstance(self.deff_c, dict)
             and stratum is not None
         ):
[…]
