### 2021-03-07 2b018d8
adding sample size calculation for one sample
