### 2021-03-07 52077d1
adding sample size calculation for one sample
