+++ src/samplics/sampling/__init__.py
@@ -1,4 +1,4 @@
 from samplics.sampling.selection import SampleSelection
-from samplics.sampling.size import SampleSize, allocate
+from samplics.sampling.size import allocate, SampleSize, OneMeanSampleSize, TwoMeanSampleSize
 
-__all__ = ["allocate", "SampleSelection", "SampleSize"]
+__all__ = ["allocate", "SampleSelection", "SampleSize", "OneMeanSample", "TwoMeanSampleSize"]
