### 2021-03-07 b817604
adding sample size calculation for one sample
