### 2021-03-07 8566731
adding sample size calculation for one sample
