+++ src/samplics/sampling/size.py
@@ -552,6 +552,42 @@ class OneMeanSampleSize:
         self.alpha = alpha
         self.beta = beta
 
+        prob_alpha = (
+            normal.ppf(1 - self.alpha / 2)
+            if self.test_type == "two-side"
+            else normal.ppf(1 - self.alpha)
+        )
+        prob_beta = normal.ppf(1 - self.beta)
+        if self.stratification:
+            self.samp_size = {}
+            self.power = {}
+            for key in mean1:
+                self.samp_size[key] = math.ceil(
+                    pow(
+                        (prob_alpha + prob_beta) * std_dev[key] / (mean1[key] - mean0[key]),
+                        2,
+                    )
+                )
+                adj_fct = (mean0[key] - mean1[key]) / (
+                    std_dev[key] / math.sqrt(self.samp_size[key])
+                )
+                self.power[key] = (
+                    1 - normal.cdf(prob_alpha + adj_fct) + normal.cdf(-prob_alpha + adj_fct)
+                )
+        else:
+            self.samp_size = math.ceil(
+                pow(
+                    (prob_alpha + prob_beta)
+                    * std_dev["_stratum_1"]
+                    / (mean1["_stratum_1"] - mean0["_stratum_1"]),
+                    2,
+                )
+            )
+            adj_fct = (mean0["_stratum_1"] - mean1["_stratum_1"]) / (
+                std_dev["_stratum_1"] / math.sqrt(self.samp_size)
+            )
+            self.power = 1 - normal.cdf(prob_alpha + adj_fct) + normal.cdf(-prob_alpha + adj_fct)
+
         if self.estimated_mean:
             pass
             # prob_alpha and prob_beta should be solved iteratively, see pages 66 and 67 of Ryan's book
@@ -559,44 +595,6 @@ class OneMeanSampleSize:
             #     self.samp_size[key] = pow(
             #         (prob_alpha + prob_beta) * std_dev[key] / (mean1[key] - mean0[key]), 2
             #     )
-        else:
-            prob_alpha = (
-                normal.ppf(1 - self.alpha / 2)
-                if self.test_type == "two-side"
-                else normal.ppf(1 - self.alpha)
-            )
-            prob_beta = normal.ppf(1 - self.beta)
-            if self.stratification:
-                self.samp_size = {}
-                self.power = {}
-                for key in mean1:
-                    self.samp_size[key] = math.ceil(
-                        pow(
-                            (prob_alpha + prob_beta) * std_dev[key] / (mean1[key] - mean0[key]),
-                            2,
-                        )
-                    )
-                    adj_fct = (mean0[key] - mean1[key]) / (
-                        std_dev[key] / math.sqrt(self.samp_size[key])
-                    )
-                    self.power[key] = (
-                        1 - normal.cdf(prob_alpha + adj_fct) + normal.cdf(-prob_alpha + adj_fct)
-                    )
-            else:
-                self.samp_size = math.ceil(
-                    pow(
-                        (prob_alpha + prob_beta)
-                        * std_dev["_stratum_1"]
-                        / (mean1["_stratum_1"] - mean0["_stratum_1"]),
-                        2,
-                    )
-                )
-                adj_fct = (mean0["_stratum_1"] - mean1["_stratum_1"]) / (
-                    std_dev["_stratum_1"] / math.sqrt(self.samp_size)
-                )
-                self.power = (
-                    1 - normal.cdf(prob_alpha + adj_fct) + normal.cdf(-prob_alpha + adj_fct)
-                )
 
 
 class TwoMeanSampleSize:
