+++ src/samplics/weighting/__init__.py
@@ -1,4 +1,5 @@
 from samplics.weighting.adjustment import SampleWeight
 from samplics.weighting.replicates import ReplicateWeight
 
+
 __all__ = ["SampleWeight", "ReplicateWeight"]
