+++ src/samplics/__init__.py
@@ -1,14 +1,16 @@
 from samplics.estimation import ReplicateEstimator, TaylorEstimator
-from samplics.sampling import SampleSelection, SampleSize
-from samplics.weighting import ReplicateWeight, SampleWeight
 from samplics.sae import EblupAreaModel, EblupUnitModel, EbUnitModel, EllUnitModel
-
-from samplics.utils.formats import array_to_dict
+from samplics.sampling import OneMeanSampleSize, SampleSelection, SampleSize, allocate
 from samplics.utils.basic_functions import transform
+from samplics.utils.formats import array_to_dict
+from samplics.weighting import ReplicateWeight, SampleWeight
+
 
 __all__ = [
+    "allocate",
     "SampleSelection",
     "SampleSize",
+    "OneMeanSampleSize",
     "SampleWeight",
     "ReplicateWeight",
     "ReplicateEstimator",
