+++ tests/simulated_population.py
@@ -1,5 +1,6 @@
 import numpy as np
 
+
 population_size = 35000000
 
 admin1_nb = 10
