+++ tests/sampling/test_selection.py
@@ -3,6 +3,7 @@ import pandas as pd
 
 from samplics.sampling import SampleSelection
 
+
 countries_population = pd.read_csv("./tests/sampling/countries_population_2019.csv")
 
 countries = countries_population["country"]
