+++ tests/sae/simulate_lmm_population.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pandas as pd
-
 import statsmodels.api as sm
 
 
