+++ src/samplics/sae/__init__.py
@@ -1,6 +1,6 @@
+from samplics.sae.eb_unit_model import EbUnitModel
 from samplics.sae.eblup_area_model import EblupAreaModel
 from samplics.sae.eblup_unit_model import EblupUnitModel
-from samplics.sae.eb_unit_model import EbUnitModel
 from samplics.sae.robust_unit_model import EllUnitModel
 
 
