+++ tests/utils/test_formats.py
@@ -1,7 +1,6 @@
-import pytest
-
 import numpy as np
 import pandas as pd
+import pytest
 
 from samplics.utils.formats import *
 
