+++ tests/weighting/data_weights.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pandas as pd
 
+
 # from scipy.stats import gamma
 
 np.random.seed(181336)
