### 2021-04-10 #2908 @rchew
@lorenanicole this is my first time as a reviewer for JOSS -- we're still waiting on getting a 2nd reviewer in order to move this into the `[REVIEW]` stage, right?  Just wanted to confirm I wasn't holding the process up.
