+++ src/samplics/__init__.py
@@ -2,6 +2,7 @@ from samplics.datasets.datasets import (
     Birth,
     CountyCrop,
     CountyCropMeans,
+    ExpenditureMilk,
     Nhanes2,
     Nhanes2brr,
     Nhanes2jk,
@@ -28,6 +29,7 @@ __all__ = [
     "EblupUnitModel",
     "EbUnitModel",
     "EllUnitModel",
+    "ExpenditureMilk",
     "Nhanes2",
     "Nhanes2brr",
     "Nmihs",
