+++ src/samplics/datasets/__init__.py
@@ -1,0 +1,28 @@
+from samplics.datasets.datasets import (
+    Birth,
+    CountyCrop,
+    CountyCropMeans,
+    ExpenditureMilk,
+    Nhanes2,
+    Nmihs,
+    Nhanes2brr,
+    Nhanes2jk,
+    PSUFrame,
+    PSUSample,
+    SSUSample,
+)
+
+
+__all__ = [
+    "Birth",
+    "CountyCrop",
+    "CountyCropMeans",
+    "ExpenditureMilk",
+    "Nhanes2",
+    "Nhanes2brr",
+    "Nhanes2jk",
+    "Nmihs",
+    "PSUFrame",
+    "PSUSample",
+    "SSUSample",
+]
