+++ pyproject.toml
@@ -28,16 +28,16 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.7" #Pandas 1.0 requires at least Python 3.6.1
-numpy = "^1.15"
-pandas = "^1.0"
+python = "^3.7.1" #Pandas 1.0 requires at least Python 3.6.1
+numpy = "^1.16"
+pandas = "^1.2"
 scipy = "^1.1"
-statsmodels = "^0.10"
-matplotlib = "^3.1"
+statsmodels = "^0.12"
+matplotlib = "^3.4"
 
 [tool.poetry.dev-dependencies]
 mypy = "^0.812"
-pytest = "^6.2.1"
+pytest = "^6.2"
 pytest-cov = "^2.8.1"
 codecov = "^2.1.3"
 black = "^20.8b1"
