+++ src/samplics/__init__.py
@@ -47,4 +47,4 @@ __all__ = [
     "transform",
 ]
 
-__version__ = "0.3.2"
+__version__ = "0.3.4"
