+++ src/samplics/__init__.py
@@ -1,4 +1,5 @@
 from samplics.datasets.datasets import (
+    Auto,
     Birth,
     CountyCrop,
     CountyCropMeans,
@@ -22,6 +23,7 @@ from samplics.weighting import ReplicateWeight, SampleWeight
 __all__ = [
     "allocate",
     "array_to_dict",
+    "Auto",
     "Birth",
     "CountyCrop",
     "CountyCropMeans",
@@ -47,4 +49,4 @@ __all__ = [
     "transform",
 ]
 
-__version__ = "0.3.4"
+__version__ = "0.3.5"
