+++ docs/source/tutorial/index.rst
@@ -6,15 +6,15 @@ Samplics is a Python package designed to be a comprehensive tool for selecting, 
 
 This tutorial is intended for users with basic knowledge of Python. It is assumed that the user has a basic understanding of Python syntax. Also, the tutorial is not intended to teach survey sampling methods. To learn survey sampling methods, we refer the user to the `UNStats Handbook 2005 <https://unstats.un.org/unsd/hhsurveys/>`_, `Designing Household Survey Samples: Practical Guidelines <https://unstats.un.org/unsd/demographic/sources/surveys/Series_F98en.pdf>`_, and the reference material mentioned throughout this tutorial. 
 
-The module *datasets* allows the user to access the datasets used in this tutorial. Note that the datasets are only used to illustrate the the syntax and APIs of *samplics*. The datasets should not be used for any statistical analysis. **Many of the datasets in this tutorial are subsets of real-life samples but DO NOT represent these real-life samples**. The datasets were subseted from existing samples to reduce the size of the files.  Each datasets is represented by a class with the members: *name*, *description*, *nrows*, *ncols*, *design*, *source*, and, *data*. The list of datasets are the following:
-* Birth: This dataset represent a city data of categories of birth by age group. The dataset was obtained through the public stata API.
-* CountyCrop and CountryCropMeans: These datasets were used by Battese, G.E., Harter, R.M. and Fuller, W.A. [#battese1988]_ for their pioner paper on small area estimation.
-* ExpenditureMilk: 
-* Nhanes2, Nhanes2brr, Nhanes2jk: these datasets were obtained from the NHANES[#mcdowell1981]_. As mentioned above, the datasets are only subsets of the full sample and do not represent the NHANES II study. This data is only useful for illustrating the syntax of *samplics*. These datasets should not be used to conduct any analysis of NHANES nor use the numbers for any statistical analysis. The original data was obtained through the public stata API.
-* Nmihs: The dataset is a subset of the National Maternal and Infant Health Survey (NMIHS) sample[#gonzalez1992]_. The dataset should not be used to conduct any analysis of NMIHS nor use the numbers for any statistical analysis. The original data was obtained through the public stata API.
-* PSUFrame, PSUSample, SSUSample: these are simulated datasets to illustrate the selectionof primary and secondary sampling units.
+The module *datasets* allows the user to load the datasets used in this tutorial. Note that the datasets are only used to illustrate the syntax and APIs of *samplics*. The datasets should not be used for any statistical analysis. **Many of the datasets in this tutorial are subsets of actual samples but DO NOT represent these samples**. The datasets were subseted from existing samples to reduce the size of the files.  Each dataset is represented by a class with the members: *name*, *description*, *nrows*, *ncols*, *design*, *source*, and, *data*. The list of datasets are the following:
+- Birth: This dataset represent a city data of categories of birth by age group. The dataset was obtained through the public stata API.
+- CountyCrop and CountryCropMeans: These datasets were used by Battese, G.E., Harter, R.M. and Fuller, W.A. [#battese1988]_ for their pioner paper on small area estimation.
+- ExpenditureMilk: The Milk Expenditure data contains 43 observations on the average expenditure on fresh milk for the year 1989. This dataset was originally used by Arora and Lahiri (1997) and later by You and Chapman (2006).
+- Nhanes2, Nhanes2brr, Nhanes2jk: these datasets were obtained from the NHANES [#mcdowell1981]_. As mentioned above, the datasets are only subsets of the full sample and do not represent the NHANES II study. This data is only useful for illustrating the syntax of *samplics*. These datasets should not be used to conduct any analysis of NHANES nor use the numbers for any statistical analysis. The original data was obtained through the public stata API.
[…]
