### 2021-06-17 samplics-org/samplics#29 @soodoku
fix use of some deprecated things
```
sample_size = SampleSize(parameter = "proportion")
sample_size.calculate(target=0.80, precision=0.10)
```

```
FutureWarning: pandas.util.testing is deprecated. Use the functions in the public API at pandas.testing instead.
  import pandas.util.testing as tm
```
