### 2021-06-17 samplics-org/samplics#30 @soodoku
testing
dear @MamadouSDiallo,

from joss.
it will be cool to have some validation around correctness.

one idea = confirm some outputs w/ lumley's survey sampling package?
