### 2021-06-17 samplics-org/samplics#31 @soodoku
comm. guidelines
dear @MamadouSDiallo,

couldn't find the following:

check other JOSS software to see what you can add to the readme.
