### 2021-06-17 #3376 @soodoku
For stats packages, I am going to sign off on 'functionality' without checking for correctness as that will take a sig. investment. I have proposed a way to match outputs to other implementations like Lumley's R version. But I am assuming the bar for acceptance is lower.
