### 2021-06-17 samplics-org/samplics#32 @soodoku
minor issues w/ writing
Dear @MamadouSDiallo,

to sign off on " Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?", i recommend you fix all the minor grammatical issues. one idea is to use grammarly. just take a close look.
