### 2021-10-05 samplics-org/samplics#34 @rchew
Documentation outdated in README and ReadTheDocs
Followed instructions to `pip install samplics` (samplics==0.3.10) and then tried some of the example code snippets from the project README page.  It seems like there are newly supported keywords arguments that are not supported on the version on pip.

For example, the following throws `TypeError: calculate() got an unexpected keyword argument 'precision'`
```
import samplics
from samplics.sampling import SampleSize

sample_size = SampleSize(parameter = "proportion")
sample_size.calculate(target=0.80, precision=0.10)
```

And the following references a local file ("psu_frame.csv") which users do not appear to have access to for the pip version, though it appears the code on GH supports importing the reference dataset using something like `from samplics.dataset import psu_frame`.
```
import samplics
from samplics.sampling import SampleSelection

psu_frame = pd.read_csv("psu_frame.csv")
psu_sample_size = {"East":3, "West": 2, "North": 2, "South": 3}
pps_design = SampleSelection(
   method="pps-sys",
   stratification=True,
   with_replacement=False
   )

frame["psu_prob"] = pps_design.inclusion_probs(
   psu_frame["cluster"],
   psu_sample_size,
   psu_frame["region"],
   psu_frame["number_households_census"]
   )
```

As another general note, I would also add code to the examples to import all libraries needed to run each code snippet.  For example, in the snippet above, I would add `import pandas as pd` to the top if you're going to use Panda's `read_csv` function.
