### 2021-10-05 #3376 @rchew
@MamadouSDiallo Apologies once again for how long this has taken to review.  Started adding some issues to your repo and wanted to tag in this thread as well.

Started with following the examples on the README using the version on pip, but it appears that the pip version is out of date with the documentation and the code on GH.  Will clone the repo locally for the time being to address the other components of the library.
https://github.com/samplics-org/samplics/issues/34
