### 2021-10-05 #3376 @rchew
@MamadouSDiallo Finished an initial review, and overall, this is looking in pretty good shape!  Still assessing Functionality and Documentation, but wanted to bring your attention to @soodoku 's  comments if you hadn't yet seen them:

- https://github.com/samplics-org/samplics/issues/29
- https://github.com/samplics-org/samplics/issues/30
- https://github.com/samplics-org/samplics/issues/31
- https://github.com/samplics-org/samplics/issues/32
