### 2021-10-29 samplics-org/samplics#30 @MamadouSDiallo
Dear @soodoku

samplics has an extensive test suit. The vast majority of the tests were created by using results of the same run from Stata. When Stata did not have a feature, I used the R survey package. Earlier versions of the tutorial had some comparisons with Stata but I remove them.

I will produce a limited set of comparisons with Stata for this review.

Best regards
