### 2021-10-29 samplics-org/samplics#32 @MamadouSDiallo
Dear @soodoku I updated the paper for your review. Best
