### 2021-10-30 samplics-org/samplics#31 @MamadouSDiallo
Dear @soodoku
I added the contributing guide
Best
