### 2021-10-30 #3376 @MamadouSDiallo
Dear @arfon, @rchew, @soodoku

I believe I have addressed all the issues but the testing. Please review and let me know.

The samplics testing suite is extensive (about half of the codebase). The vast majority of those tests were created by checking the results against Stata and sometimes the R survey package when Stata did not have the feature. In fact, earlier versions of the tutorial had comparisons with Stata but I removed those for readability. However, for the purpose of JOSS, I will create a limited set of tests to compare samplics output to Stata (and R survey if feature not available on Stata).

Best regards
