### 2021-10-31 #3376 @arfon
Thanks @soodoku!

@rchew - it looks like the authors have addressed the [issues you referenced here](https://github.com/openjournals/joss-reviews/issues/3376#issuecomment-934573758). Could you come and take a look at the submission once more and see if the few remaining checkboxes can be marked as complete now?
