### 2021-12-06 #3376 @arfon
@MamadouSDiallo – is there any way you can tweak the layout of your equations to stop the text falling into the margins...

![Uploading Screenshot 2021-12-06 at 13.19.08.png…]()
