### 2021-12-07 #3376 @MamadouSDiallo
Somehow the editor was witching _ to *
I hope it's fixed now. Let me know.
