### 2022-01-10 #3376 @MamadouSDiallo
Dear @whedon

I found a typo in the paper I don't know if it's still possible to fix.
The typo is "let us us us" in _Sample Estimation_ section.

I fixed it in my github copy.

Best regards
