### 2022-01-10 #3376 @arfon
@MamadouSDiallo – I just fixed this in the JOSS copy too. It may take 24 hours for the new version to show up on the JOSS website due to caching.
